refactor(web): split route pages by business domain
