+++ apps/web/src/App.tsx
@@ -1,85 +1,8 @@
-import { FormEvent, type ReactNode, useCallback, useEffect, useId, useRef, useState } from "react";
-import { createPortal } from "react-dom";
-import { Activity, BarChart3, ChevronRight, ClipboardCheck, Database, Eye, EyeOff, FileClock, FileUp, LogOut, Network, PauseCircle, PlayCircle, Plus, RefreshCw, Search, ShieldCheck, Target, UsersRound, X } from "lucide-react";
-import type { ChinaMap } from "./china-map";
-import { PAGE_ORDER, PAGE_ROUTES, readPageId, type PageId } from "./page-routes";
-
-type Capabilities = { viewPerformance:boolean; viewGoals:boolean; viewAudits:boolean; viewOrganization:boolean; viewApprovals:boolean; editPerformance:boolean; exportPerformance:boolean; exportGoals:boolean; manageAccounts:boolean; manageOrganization:boolean };
-type User = { id: string; personId:string; username: string; displayName: string; mustChangePassword: boolean; roles: string[]; capabilities:Capabilities };
-type AuthState = { status: "loading" } | { status: "guest" } | { status: "authenticated"; user: User };
-type OrderLifecycle = "draft" | "active" | "paused" | "zero" | "historical_review_required";
-type Order = { id: string; orderNo: string; customerName: string; customerUnit: string; salespersonName: string; serviceType: string | null; sourceReceivedOn: string; originalAmount: string; currentRevenue: string; countedAmount: string; lifecycleState: OrderLifecycle; postedAt: string; departmentName:string; groupName:string; leaderName:string|null; supervisorName:string|null };
-type OrderFilters = { search:string;month:string;status:string;salesperson:string;department:string;group:string;region:string;customerUnit:string };
-const emptyOrderFilters:OrderFilters={search:"",month:"",status:"",salesperson:"",department:"",group:"",region:"",customerUnit:""};
-const orderFilterUrlKeys:Record<keyof OrderFilters,string>={search:"orderSearch",month:"orderMonth",status:"orderStatus",salesperson:"orderSalesperson",department:"orderDepartment",group:"orderGroup",region:"orderRegion",customerUnit:"orderCustomerUnit"};
-function readOrderUrlState(){const params=new URLSearchParams(window.location.search);const filters={...emptyOrderFilters};for(const key of Object.keys(orderFilterUrlKeys) as Array<keyof OrderFilters>)filters[key]=params.get(orderFilterUrlKeys[key])??"";return{filters,cursor:params.get("orderCursor")};}
-function writeOrderUrlState(filters:OrderFilters,cursor:string|null,mode:"push"|"replace"="push"){const params=new URLSearchParams(window.location.search);params.set("page","orders");for(const key of Object.keys(orderFilterUrlKeys) as Array<keyof OrderFilters>){if(filters[key])params.set(orderFilterUrlKeys[key],filters[key]);else params.delete(orderFilterUrlKeys[key]);}if(cursor)params.set("orderCursor",cursor);else params.delete("orderCursor");window.history[mode==="push"?"pushState":"replaceState"]({},"",`${window.location.pathname}?${params.toString()}${window.location.hash}`);}
-type PerformanceEvent = { id:string; sequence:number; eventType:string; deltaAmount:string; resultingCurrentRevenue:string; resultingCountedAmount:string; resultingLifecycleState:OrderLifecycle|null; accountingMonth:string; occurredOn:string; occurredAt:string; reason:string|null; actorName:string|null; salespersonName:string; departmentName:string|null; groupName:string|null; leaderName:string|null; supervisorName:string|null; businessRegionCode:string|null; businessRegionSourceText:string|null; customerUnit:string|null };
-type AccountingPeriod={periodMonth:string;status:"open"|"closed";version:number;needsReclose:boolean;verifiedAt:string|null;verifiedBy:string|null;closedAt:string|null;closedBy:string|null};
-type AccountingCorrection={id:string;periodMonth:string;orderId:string;orderNo:string;eventType:string;occurredOn:string;reason:string;businessRegionCode:string|null;businessRegionSourceText:string|null;customerUnit:string|null;analysisDimensionEvidence:string|null;status:"pending"|"approved"|"rejected"|"consumed"|"revoked";requestedBy:string;reviewedBy:string|null;reviewNote:string|null;expiresAt:string|null};
-type HistoricalReview={id:string;orderId:string;orderNo:string;lifecycleState:"active"|"paused"|"zero";currentRevenue:string;conclusion:string;evidence:string;reason:string;status:"pending"|"approved"|"rejected";requestedBy:string;reviewedBy:string|null;reviewNote:string|null};
-type PersonalAchievementEvent = {id:string;orderId:string;orderNo:string;customerName:string;eventType:string;deltaAmount:string;accountingMonth:string;occurredOn:string;sequence:number;reason:string|null;resultingCountedAmount:string;resultingLifecycleState:"active"|"paused"|"zero"|null;departmentName:string|null;groupName:string|null};
-type PersonalAchievement = { goalId:string|null;periodMonth:string;targetAmount:string|null;actualAmount:string;gapAmount:string|null;achievementRate:string|null;timeProgressRate:string|null;progressVariance:string|null;calculationReason:"PERIOD_IN_FUTURE"|"TARGET_NOT_ACTIVE"|"TARGET_AMOUNT_NOT_POSITIVE"|"TARGET_SCOPE_AMBIGUOUS"|null;eventCount:number };
-type PersonalAchievementDrilldown = PersonalAchievement&{events:PersonalAchievementEvent[]};
-type GroupAchievement = PersonalAchievement&{groupId:string;groupName:string;memberCount:number};
-type GroupAchievementOrder = {orderId:string;orderNo:string;customerName:string;actualAmount:string;eventCount:number;events:PersonalAchievementEvent[]};
-type GroupAchievementMember = {personId:string|null;personKey:string;name:string;actualAmount:string;eventCount:number;orders:GroupAchievementOrder[]};
-type GroupAchievementDrilldown = GroupAchievement&{members:GroupAchievementMember[]};
-type OrganizationGroupAchievement = PersonalAchievement&{groupId:string|null;groupKey:string;groupName:string;memberCount:number;members:GroupAchievementMember[]};
-type DepartmentAchievement = PersonalAchievement&{departmentId:string|null;departmentKey:string;departmentName:string;groupCount:number};
-type DepartmentAchievementDrilldown = DepartmentAchievement&{groups:OrganizationGroupAchievement[]};
-type SalesAchievement = PersonalAchievement&{departmentCount:number};
-type SalesAchievementDrilldown = SalesAchievement&{departments:DepartmentAchievementDrilldown[]};
-type DashboardData = { month: string; metrics: { total: string; eventCount: number; negativeTotal: string; pendingApprovals: number }; monthly: Array<{ month: string; total: string }>; groups: Array<{ id:string;name:string;total:string }>; recent: Array<{ orderNo: string; salespersonName: string; eventType: string; month: string; amount: string; groupName: string }>; personalAchievement:PersonalAchievement|null;groupAchievements:GroupAchievement[];departmentAchievements:DepartmentAchievement[];salesAchievement:SalesAchievement|null };
-type AnalysisAmount = { eventCount:number;totalAmount:string };
-type AnalysisProvince = {regionCode:string;regionName:string;eventCount:number;totalAmount:string};
-type AnalysisCustomer = {customerUnit:string;eventCount:number;totalAmount:string};
-type PerformanceAnalysis = { month:string;ledger:AnalysisAmount;mapped:AnalysisAmount;pending:AnalysisAmount;reconciled:boolean;provinces:AnalysisProvince[];foreignTrade:{regionCode:"EXT-TRADE";regionName:string;eventCount:number;totalAmount:string};customers:Array<AnalysisProvince&{customerUnit:string}> };
-type AnalysisCustomersDrilldown = {level:"customers";regionCode:string;regionName:string;month:string;eventCount:number;totalAmount:string;customerCount:number;nextCursor:string|null;pageSize:number;customers:AnalysisCustomer[]};
-type AnalysisMonthsDrilldown = {level:"months";regionCode:string;regionName:string;customerUnit:string;year:string;eventCount:number;totalAmount:string;months:Array<{month:string;eventCount:number;totalAmount:string}>};
-type AnalysisDrilldownEvent = {id:string;eventType:string;deltaAmount:string;resultingCurrentRevenue:string;resultingCountedAmount:string;resultingLifecycleState:"active"|"paused"|"zero"|null;accountingMonth:string;occurredOn:string;reason:string|null;salespersonName:string;departmentName:string|null;groupName:string|null;sequence:number;businessRegionCode:string;businessRegionSourceText:string;customerUnit:string};
-type AnalysisDrilldownOrder = {orderId:string;orderNo:string;customerName:string;eventCount:number;totalAmount:string;events:AnalysisDrilldownEvent[]};
-type AnalysisEventsDrilldown = {level:"events";regionCode:string;regionName:string;customerUnit:string;month:string;eventCount:number;totalAmount:string;nextCursor:string|null;pageSize:number;orders:AnalysisDrilldownOrder[]};
-type GoalLevel="sales_manager"|"department"|"group"|"personal";
-type Goal = { id:string;periodMonth:string;level:GoalLevel;ownerUsername:string|null;ownerName:string;ownerPersonId:string;orgUnitId:string|null;orgUnitName:string|null;parentGoalId:string|null;versionId:string;versionNo:string;amount:string;effectiveAmount:string|null;status:string;signatureText:string|null;signedAt:string|null;changeReason:string;allocatedAmount:string;allocationDifference:string;allocationType:"unallocated"|"overallocated"|"balanced";allocationRatio:string|null };
-type GoalOption={personId:string;name:string;orgUnitId:string|null;orgUnitName:string|null};
-type ParentGoalOption={id:string;ownerName:string;orgUnitId:string|null;orgUnitName:string|null;amount:string};
-type GoalChangeRequest={id:string;goalId:string;level:GoalLevel;ownerName:string;requestedAmount:string|null;reason:string;status:string;outcomeComment:string|null;createdVersionId:string|null;createdAt:string;currentAmount:string|null;newAmount:string|null;amountDifference:string|null;requestedByName:string;handledByName:string|null;canHandle:boolean;canWithdraw:boolean};
-type GoalLinkageDecision={id:string;parentGoalId:string;triggeringChildVersionId:string;decision:"keep_parent"|"adjust_parent"|null;status:string;reason:string|null;generatedChangeRequestId:string|null;childGoalId:string;childOwnerName:string;childAmount:string;decidedAt:string|null;parentLevel:GoalLevel;parentAmount:string;canDecide:boolean};
-type GoalWorkflows={changeRequests:GoalChangeRequest[];linkageDecisions:GoalLinkageDecision[]};
-type GoalHistory={versions:Array<{id:string;versionNo:string;amount:string;previousAmount:string|null;amountDifference:string|null;status:string;createdByName:string;signedByName:string|null;signatureText:string|null;signedAt:string|null;changeReason:string;createdAt:string}>;approvals:Array<{id:string;goalVersionId:string;stage:string;decision:string;decidedByName:string;comment:string|null;decidedAt:string}>;changeRequests:Array<{id:string;requestedAmount:string|null;reason:string;status:string;outcomeComment:string|null;createdAt:string;requestedByName:string;handledByName:string|null}>;audit:Array<{id:string;action:string;entityType:string;entityId:string;actorName:string|null;beforeData:unknown;afterData:unknown;createdAt:string}>};
-const GOAL_CONFIRMATION_STATEMENT="本人已核对并确认承担本目标版本。";
-type FormalReport = { goalId:string; periodMonth:string; level:Goal["level"]; ownerName:string; targetAmount:string; actualAmount:string; gapAmount:string; achievementRate:string|null };
-type AdminUser = { id:string; username:string; displayName:string; isActive:boolean; mustChangePassword:boolean; roles:string[] };
-type PersonOption = { id:string; displayName:string; username:string|null };
-type RolePermission = { code:string; name:string; businessScope:"none"|"self"|"group"|"department"|"all"; businessOperations:readonly string[]; targetResponsibilities:string; exportPermission:string; forbidden:readonly string[] };
-type OrgUnit = { id:string; name:string; unitType:"department"|"group"; parentId:string|null; parentName:string|null; isActive:boolean };
-type Assignment = { id:string; username:string; displayName:string; departmentName:string|null; groupName:string|null; effectiveFrom:string; effectiveTo:string|null };
-type Responsibility = { id:string; personId:string; username:string|null; displayName:string; unitId:string; unitName:string; unitType:"department"|"group"; responsibilityType:"leader"|"supervisor"; effectiveFrom:string; effectiveTo:string|null };
-type AuditRow = { id:string; actorPersonId:string|null; actorUsername:string|null; actorDisplayName:string|null; action:string; entityType:string; entityId:string|null; beforeData:unknown; afterData:unknown; createdAt:string };
-type AuditFilters = { person:string; action:string; entityType:string; entityId:string; from:string; to:string };
-const emptyAuditFilters:AuditFilters={person:"",action:"",entityType:"",entityId:"",from:"",to:""};
-const roleNames: Record<string, string> = { system_admin: "系统管理员", sales_assistant: "销售助理", sales_assistant_leader: "销售助理组长", sales_manager: "销售经理", sales_supervisor: "业务主管", sales_leader: "业务员组长", salesperson: "业务员", hr: "人事部", general_manager: "总经理" };
-function businessDateToday():string{const parts=new Intl.DateTimeFormat("zh-CN",{timeZone:"Asia/Shanghai",year:"numeric",month:"2-digit",day:"2-digit"}).formatToParts(new Date());const value=(type:Intl.DateTimeFormatPartTypes)=>parts.find((part)=>part.type===type)?.value??"";return `${value("year")}-${value("month")}-${value("day")}`;}
-
-function readCsrfToken(): string | null {
-  const cookie = document.cookie.split("; ").find((item) => item.startsWith("sampleflow_csrf="));
-  return cookie ? decodeURIComponent(cookie.slice("sampleflow_csrf=".length)) : null;
-}
-
-async function apiFetch(input: RequestInfo | URL, init: RequestInit = {}): Promise<Response> {
-  const method = (init.method ?? "GET").toUpperCase();
-  const headers = new Headers(init.headers);
-  if (!["GET", "HEAD", "OPTIONS"].includes(method)) {
-    const csrfToken = readCsrfToken();
-    if (csrfToken) headers.set("x-csrf-token", csrfToken);
-  }
-  return fetch(input, { ...init, headers });
-}
-
-async function readResponseJson<T>(response:Response,fallback:string):Promise<T>{
-  try{return await response.json() as T;}catch{throw new Error(fallback);}
-}
+import { type FormEvent, useCallback, useEffect, useRef, useState } from "react";
+import { Activity, ChevronRight, Database, Eye, EyeOff, ShieldCheck } from "lucide-react";
+import { apiFetch, readResponseJson } from "./app-api";
+import type { AuthState, User } from "./app-types";
+import { Dashboard } from "./dashboard";
 
 function passwordStrength(password: string): "弱" | "中" | "强" {
   const categoryCount = [/[a-z]/, /[A-Z]/, /[0-9]/, /[^A-Za-z0-9]/].filter((pattern) => pattern.test(password)).length;
@@ -165,612 +88,3 @@ function Login({ onLogin }: { onLogin: (user: User) => void }) {
     <section className="login-panel"><div className="login-card"><div className="login-heading"><p>销售到样业绩管理</p><h2>登录系统</h2><span>{import.meta.env.DEV ? "开发环境已预填销售助理演示账号" : "请使用管理员分配的账号登录"}</span></div><form noValidate onSubmit={submit}><label htmlFor="username">账号</label><input id="username" name="username" value={username} onChange={(e) => setUsername(e.target.value)} autoComplete="username" /><PasswordInput id="password" name="password" label="密码" value={password} onChange={setPassword} autoComplete="current-password"/><button type="submit" disabled={submitting}>{submitting ? "正在登录…" : "进入 SampleFlow"}<ChevronRight size={18} /></button>{message ? <p className="form-error" role="alert">{message}</p> : null}</form><div className={`readiness readiness-${readiness}`} role="status" aria-live="polite"><span />{{checking:"正在检查 API 与数据库",ready:"前端、API 与数据库已连接",unavailable:"API 或数据库暂不可用"}[readiness]}</div></div></section>
   </main>;
 }
-
-function Dashboard({ user, onLogout }: { user: User; onLogout: () => void }) {
-  const pageIcons = {overview:BarChart3,goals:Target,orders:ClipboardCheck,analysis:Activity,organization:Network,approvals:FileClock,audits:Search,accounts:UsersRound};
-  const canOpen=(page:PageId)=>user.capabilities[PAGE_ROUTES[page].capability];
-  const pages=PAGE_ORDER.filter(canOpen).map((id)=>({id,Icon:pageIcons[id],label:PAGE_ROUTES[id].label}));
-  const defaultPage:PageId=user.capabilities.manageAccounts?"accounts":pages[0]?.id??"overview";
-  const [active,setActive]=useState<PageId>(()=>readPageId(window.location.search)??defaultPage);
-  const navigate=useCallback((page:PageId,mode:"push"|"replace"="push")=>{if(mode==="push"&&readPageId(window.location.search)===page)return;const params=new URLSearchParams(window.location.search);params.set("page",page);window.history[`${mode}State`]({},"",`${window.location.pathname}?${params.toString()}${window.location.hash}`);setActive(page);},[]);
-  useEffect(()=>{if(!readPageId(window.location.search))navigate(defaultPage,"replace");const restore=()=>{const page=readPageId(window.location.search);if(page)setActive(page);else navigate(defaultPage,"replace");};window.addEventListener("popstate",restore);return()=>window.removeEventListener("popstate",restore);},[defaultPage,navigate]);
-  useEffect(()=>{document.title=`${PAGE_ROUTES[active].label} — SampleFlow`;},[active]);
-  async function logout() { await apiFetch("/api/auth/logout", { method: "POST" }); onLogout(); }
-  const content = !canOpen(active)
-    ? <main className="dashboard"><header><div><h1>无法访问{PAGE_ROUTES[active].label}</h1><p>该页面不在当前账号的角色权限范围内</p></div></header><div className="permission-note"><ShieldCheck size={18}/>403 · 当前账号没有{PAGE_ROUTES[active].label}权限。</div></main>
-    : active === "orders"
-    ? <OrdersPage user={user} />
-    : active === "goals"
-      ? <GoalsPage user={user} pendingOnly={false} />
-      : active === "approvals"
-        ? <GoalsPage user={user} pendingOnly />
-      : active === "organization"
-        ? <OrganizationPage user={user}/>
-      : active === "analysis"
-        ? <AnalysisPage/>
-      : active === "audits"
-        ? <AuditPage/>
-      : active === "accounts"
-        ? <AccountsPage user={user}/>
-      : <Overview canEdit={user.capabilities.editPerformance} canExport={user.capabilities.exportPerformance} onEnterOrders={() => navigate("orders")} />;
-  return <div className="app-shell"><aside className="sidebar"><div className="sidebar-brand"><img src="/brand-logo.png" alt="瑞源生物 Pronetbio"/></div><nav>{pages.map(({Icon,label,id}) => <button className={id === active ? "active" : ""} key={id} onClick={() => navigate(id)} aria-label={label} aria-current={id===active?"page":undefined}><Icon size={18}/><span>{label}</span></button>)}</nav><div className="sidebar-user"><div className="avatar">{user.displayName.slice(0,1)}</div><div><strong>{user.displayName}</strong><span>{user.roles.map((r) => roleNames[r] ?? r).join("、")}</span></div><button onClick={logout} aria-label="退出登录"><LogOut size={17}/></button></div></aside>{content}</div>;
-}
-
-function auditData(value:unknown):string{return value===null||value===undefined?"—":JSON.stringify(value);}
-function auditDateTime(value:string):string{return new Intl.DateTimeFormat("zh-CN",{dateStyle:"medium",timeStyle:"medium",timeZone:"Asia/Shanghai"}).format(new Date(value));}
-function auditTimeParameter(value:string):string{return value?`${value}${value.length===16?":00":""}+08:00`:"";}
-
-function AuditPage(){
-  const[draft,setDraft]=useState<AuditFilters>(emptyAuditFilters);const[filters,setFilters]=useState<AuditFilters>(emptyAuditFilters);const[audits,setAudits]=useState<AuditRow[]>([]);const[cursor,setCursor]=useState<string|null>(null);const[nextCursor,setNextCursor]=useState<string|null>(null);const[loading,setLoading]=useState(false);const[error,setError]=useState("");const[revision,setRevision]=useState(0);
-  const requestKey=JSON.stringify([filters,cursor,revision]);
-  useEffect(()=>{const controller=new AbortController();const params=new URLSearchParams();for(const key of ["person","action","entityType","entityId"] as const)if(filters[key])params.set(key,filters[key]);if(filters.from)params.set("from",auditTimeParameter(filters.from));if(filters.to)params.set("to",auditTimeParameter(filters.to));if(cursor)params.set("cursor",cursor);setLoading(true);setError("");setNextCursor(null);if(!cursor)setAudits([]);fetch(`/api/audits${params.size?`?${params.toString()}`:""}`,{signal:controller.signal}).then(async(response)=>{const data=await readResponseJson<{audits?:AuditRow[];nextCursor?:string|null;message?:string}>(response,"审计响应无效，请重试。");if(!response.ok)throw new Error(data.message??"审计查询失败");setAudits((current)=>cursor?[...current,...(data.audits??[])]:data.audits??[]);setNextCursor(data.nextCursor??null);}).catch((reason)=>{if(reason instanceof DOMException&&reason.name==="AbortError")return;setError(reason instanceof Error?reason.message:"审计查询失败");}).finally(()=>{if(!controller.signal.aborted)setLoading(false);});return()=>controller.abort();},[requestKey]);
-  function update(key:keyof AuditFilters,value:string){setDraft((current)=>({...current,[key]:value}));}
-  function query(event:FormEvent){event.preventDefault();setFilters(Object.fromEntries(Object.entries(draft).map(([key,value])=>[key,value.trim()])) as AuditFilters);setCursor(null);setRevision((value)=>value+1);}
-  function clear(){setDraft(emptyAuditFilters);setFilters(emptyAuditFilters);setCursor(null);setRevision((value)=>value+1);}
-  return <main className="dashboard"><header><div><h1>审计查询</h1><p>按账号、动作、实体和时间追溯不可变记录；数据范围沿用当前角色权限</p></div></header><div className="permission-note"><ShieldCheck size={18}/>此页面只提供查询，不提供修改或删除入口；敏感凭据字段不会返回。</div><section className="orders-card" aria-labelledby="audit-table-title"><div className="orders-toolbar"><div><h2 id="audit-table-title">审计记录</h2><span role="status">{loading?"正在查询…":`已加载 ${audits.length} 条记录`}</span></div></div><form noValidate className="order-filters" onSubmit={query}><div className="order-filter-grid"><label className="field"><span>人员</span><input value={draft.person} onChange={(event)=>update("person",event.target.value)} placeholder="姓名、账号或人员标识"/></label><label className="field"><span>动作</span><input value={draft.action} onChange={(event)=>update("action",event.target.value)} placeholder="如 performance.order_posted"/></label><label className="field"><span>实体类型</span><input value={draft.entityType} onChange={(event)=>update("entityType",event.target.value)} placeholder="如 performance_order"/></label><label className="field"><span>实体标识</span><input value={draft.entityId} onChange={(event)=>update("entityId",event.target.value)} placeholder="稳定标识"/></label><label className="field"><span>开始时间</span><input type="datetime-local" value={draft.from} onChange={(event)=>update("from",event.target.value)}/></label><label className="field"><span>结束时间</span><input type="datetime-local" value={draft.to} onChange={(event)=>update("to",event.target.value)}/></label></div><div className="order-filter-actions"><button type="button" onClick={clear}>清除条件</button><button type="submit">查询审计</button></div></form>{error?<div className="query-feedback"><p className="page-message" role="alert">{error}</p><button type="button" onClick={()=>setRevision((value)=>value+1)}>重试查询</button></div>:null}<div className="orders-table-wrap" tabIndex={0} role="region" aria-label="审计记录，可横向滚动"><table className="audit-table"><thead><tr><th>时间</th><th>人员</th><th>动作</th><th>实体</th><th>变更前</th><th>变更后</th></tr></thead><tbody>{!loading&&!error&&audits.length===0?<tr><td className="empty-cell" colSpan={6}>没有符合当前权限和条件的审计记录。</td></tr>:audits.map((row)=><tr key={row.id}><td><time dateTime={row.createdAt}>{auditDateTime(row.createdAt)}</time></td><td>{row.actorDisplayName??"系统"}<small className="table-secondary">{row.actorUsername??row.actorPersonId??"—"}</small></td><td>{row.action}</td><td>{row.entityType}<small className="table-secondary">{row.entityId??"—"}</small></td><td className="audit-data"><code>{auditData(row.beforeData)}</code></td><td className="audit-data"><code>{auditData(row.afterData)}</code></td></tr>)}</tbody></table></div><nav className="table-pagination" aria-label="审计分页"><span>按不可变审计编号倒序加载</span>{nextCursor?<button type="button" disabled={loading} onClick={()=>setCursor(nextCursor)}>{loading?"正在加载…":"加载更多"}</button>:null}</nav></section></main>;
-}
-
-function AccountsPage({user}:{user:User}){
-  const initialSearch=useRef(new URLSearchParams(window.location.search).get("accountSearch")??"").current;
-  const [users,setUsers]=useState<AdminUser[]>([]);const [roles,setRoles]=useState<Array<{code:string;name:string}>>([]);const [permissionMatrix,setPermissionMatrix]=useState<RolePermission[]>([]);const [message,setMessage]=useState("");const [loadError,setLoadError]=useState("");const [loading,setLoading]=useState(false);const [searchInput,setSearchInput]=useState(initialSearch);const [search,setSearch]=useState(initialSearch);const [isComposing,setIsComposing]=useState(false);const [cursors,setCursors]=useState<Array<string|null>>([null]);const [pageIndex,setPageIndex]=useState(0);const [nextCursor,setNextCursor]=useState<string|null>(null);const [revision,setRevision]=useState(0);const [showCreate,setShowCreate]=useState(false);const [roleTarget,setRoleTarget]=useState<AdminUser|null>(null);const [temporaryCredential,setTemporaryCredential]=useState<{username:string;password:string;expiresAt:string}|null>(null);const searchRef=useRef<HTMLInputElement>(null);const isAdmin=user.capabilities.manageAccounts;const cursor=cursors[pageIndex]??null;
-  const commitSearch=useCallback((value:string,historyMode:"push"|"replace"="push",updateInput=true)=>{const normalized=value.trim();const params=new URLSearchParams(window.location.search);if(normalized)params.set("accountSearch",normalized);else params.delete("accountSearch");window.history[historyMode==="push"?"pushState":"replaceState"]({},"",`${window.location.pathname}${params.size?`?${params.toString()}`:""}${window.location.hash}`);if(updateInput)setSearchInput(normalized);setSearch(normalized);setCursors([null]);setPageIndex(0);setRevision((current)=>current+1);},[]);
-  useEffect(()=>{const restore=()=>{const restored=new URLSearchParams(window.location.search).get("accountSearch")??"";setSearchInput(restored);setSearch(restored);setCursors([null]);setPageIndex(0);setRevision((current)=>current+1);};window.addEventListener("popstate",restore);return()=>window.removeEventListener("popstate",restore);},[]);
-  useEffect(()=>{if(isComposing||searchInput.trim()===search)return;const timer=window.setTimeout(()=>commitSearch(searchInput,"push",false),300);return()=>window.clearTimeout(timer);},[searchInput,isComposing,search,commitSearch]);
-  const requestKey=JSON.stringify([search,cursor]);const lastRequestKey=useRef<string|null>(null);
-  useEffect(()=>{if(!isAdmin)return;if(lastRequestKey.current!==requestKey){lastRequestKey.current=requestKey;setUsers([]);}const controller=new AbortController();const params=new URLSearchParams();if(search)params.set("search",search);if(cursor)params.set("cursor",cursor);setLoading(true);setLoadError("");setNextCursor(null);fetch(`/api/admin/users${params.size?`?${params.toString()}`:""}`,{signal:controller.signal}).then(async(response)=>{const data=await readResponseJson<{users?:AdminUser[];roles?:Array<{code:string;name:string}>;permissionMatrix?:RolePermission[];nextCursor?:string|null;message?:string}>(response,"账号响应无效，请重试。");if(!response.ok)throw new Error(data.message??"账号加载失败");setUsers(data.users??[]);setRoles(data.roles??[]);setPermissionMatrix(data.permissionMatrix??[]);setNextCursor(data.nextCursor??null);}).catch((error)=>{if(error instanceof DOMException&&error.name==="AbortError")return;setLoadError(error instanceof Error?error.message:"账号加载失败");}).finally(()=>{if(!controller.signal.aborted)setLoading(false);});return()=>controller.abort();},[isAdmin,search,cursor,requestKey,revision]);
-  function refresh(){setRevision((value)=>value+1);}
-  function restart(){setCursors([null]);setPageIndex(0);setRevision((value)=>value+1);}
-  function submitSearch(event:FormEvent){event.preventDefault();if(isComposing)return;setMessage("");commitSearch(searchInput);}
-  function clearSearch(){setSearchInput("");commitSearch("","push",false);window.requestAnimationFrame(()=>searchRef.current?.focus());}
-  function nextPage(){if(!nextCursor)return;setCursors((current)=>[...current.slice(0,pageIndex+1),nextCursor]);setPageIndex((value)=>value+1);}
-  async function toggle(item:AdminUser){const response=await apiFetch(`/api/admin/users/${item.id}/status`,{method:"PATCH",headers:{"content-type":"application/json"},body:JSON.stringify({isActive:!item.isActive})});const data=await readResponseJson<{message?:string}>(response,"状态修改响应无效，请重试。");if(!response.ok){setMessage(data.message??"状态修改失败");return;}refresh();}
-  async function resetPassword(item:AdminUser){const response=await apiFetch(`/api/admin/users/${item.id}/reset-password`,{method:"POST",headers:{"content-type":"application/json"},body:"{}"});const data=await readResponseJson<{message?:string;temporaryPassword?:string;temporaryPasswordExpiresAt?:string}>(response,"密码重置响应无效，请重试。");if(!response.ok||!data.temporaryPassword||!data.temporaryPasswordExpiresAt){setMessage(data.message??"密码重置失败");return;}setTemporaryCredential({username:item.username,password:data.temporaryPassword,expiresAt:data.temporaryPasswordExpiresAt});refresh();}
-  return <main className="dashboard"><header><div><h1>账号管理</h1><p>系统管理权限与业务权限分离；业务角色必须显式分配</p></div>{isAdmin?<button className="primary-action" onClick={()=>setShowCreate(true)}><Plus size={16}/>创建账号</button>:null}</header>{!isAdmin?<div className="permission-note"><ShieldCheck size={18}/>仅独立系统管理员可以维护账号和角色。</div>:null}{message?<p className="page-message" role="status">{message}</p>:null}{isAdmin?<><section className="orders-card" aria-labelledby="accounts-table-title"><div className="orders-toolbar"><div><h2 id="accounts-table-title">系统账号</h2><span role="status">{loading?"正在查询…":`第 ${pageIndex+1} 页 · 本页 ${users.length} 个账号`}</span></div><form noValidate className="account-search" role="search" onSubmit={submitSearch}><div className="account-search-field"><Search size={16}/><input ref={searchRef} aria-label="搜索账号" type="search" value={searchInput} onCompositionStart={()=>setIsComposing(true)} onCompositionEnd={()=>setIsComposing(false)} onChange={(event)=>setSearchInput(event.target.value)} placeholder="账号或姓名"/>{searchInput?<button type="button" className="account-search-clear" aria-label="清除账号搜索" onClick={clearSearch}><X size={15}/></button>:null}</div><button type="submit">搜索账号</button></form></div>{loadError?<div className="query-feedback"><p className="page-message" role="alert">{loadError}</p><button type="button" onClick={refresh}>重试查询</button></div>:null}<div className="orders-table-wrap"><table><thead><tr><th>账号</th><th>姓名</th><th>角色</th><th>状态</th><th>操作</th></tr></thead><tbody>{!loading&&!loadError&&users.length===0?<tr><td colSpan={5} className="empty-cell">{search?"没有符合条件的账号。":"暂无账号数据。"}</td></tr>:users.map((item)=><tr key={item.id}><td>{item.username}</td><td>{item.displayName}</td><td>{item.roles.map((role)=>roleNames[role]??role).join("、")}</td><td><span className={`status ${item.isActive?"status-active":"status-paused"}`}>{item.isActive?"启用":"停用"}</span></td><td><div className="table-actions"><button className="table-action" onClick={()=>setRoleTarget(item)}>修改角色</button><button className="table-action" onClick={()=>resetPassword(item)}>重置密码</button><button className="table-action" onClick={()=>toggle(item)}>{item.isActive?"停用":"启用"}</button></div></td></tr>)}</tbody></table></div><nav className="table-pagination" aria-label="账号分页"><span>固定按账号创建顺序分页</span><button type="button" disabled={loading||pageIndex===0} onClick={()=>setPageIndex((value)=>value-1)}>上一页</button><button type="button" disabled={loading||!nextCursor} onClick={nextPage}>下一页</button></nav></section><section className="orders-card permission-matrix-card" aria-labelledby="permission-matrix-title"><div className="orders-toolbar"><div><h2 id="permission-matrix-title">角色权限说明</h2><span>多角色账号取各角色权限并集；系统管理员角色本身不增加业务权限</span></div></div><div className="orders-table-wrap" tabIndex={0} role="region" aria-label="角色权限矩阵，可横向滚动"><table className="permission-matrix-table"><thead><tr><th>角色</th><th>数据范围</th><th>业务操作</th><th>目标职责</th><th>导出权限</th><th>明确禁止</th></tr></thead><tbody>{permissionMatrix.length===0?<tr><td colSpan={6} className="empty-cell">暂无角色权限定义</td></tr>:permissionMatrix.map((item)=><tr key={item.code}><td><strong>{item.name}</strong></td><td><span className={`scope-badge scope-${item.businessScope}`}>{scopeName(item.businessScope)}</span></td><td>{item.businessOperations.join("；")}</td><td>{item.targetResponsibilities}</td><td>{item.exportPermission}</td><td>{item.forbidden.join("；")}</td></tr>)}</tbody></table></div></section></>:null}{showCreate?<CreateAccount roles={roles} onClose={()=>setShowCreate(false)} onUncertain={(username)=>{setShowCreate(false);setMessage("创建结果不确定，已重新查询该账号；如账号已存在，请重置密码生成新的临时密码。");commitSearch(username);}} onSaved={async()=>{setShowCreate(false);restart();}}/>:null}{roleTarget?<EditAccountRoles item={roleTarget} roles={roles} onClose={()=>setRoleTarget(null)} onSaved={async()=>{const changedSelf=roleTarget.id===user.id;setRoleTarget(null);setMessage(`${roleTarget.username} 的角色已更新。`);if(changedSelf)window.location.reload();else refresh();}}/>:null}{temporaryCredential?<Modal title="临时密码已生成" note="该密码只显示一次，24 小时后失效" onClose={()=>setTemporaryCredential(null)}><div className="temporary-password-result"><p>账号：{temporaryCredential.username}</p><strong>{temporaryCredential.password}</strong><span>失效时间：{new Date(temporaryCredential.expiresAt).toLocaleString("zh-CN")}</span><div className="modal-actions"><button type="button" onClick={()=>setTemporaryCredential(null)}>我已安全保存</button></div></div></Modal>:null}</main>;
-}
-
-function scopeName(scope:RolePermission["businessScope"]):string{return{none:"无业务范围",self:"仅本人",group:"本人及所负责小组",department:"本人及所负责部门",all:"全公司 / 销售组织"}[scope];}
-
-function EditAccountRoles({item,roles,onClose,onSaved}:{item:AdminUser;roles:Array<{code:string;name:string}>;onClose:()=>void;onSaved:()=>Promise<void>}){const[selected,setSelected]=useState<string[]>(item.roles);const[error,setError]=useState("");const[saving,setSaving]=useState(false);function toggle(role:string){setSelected((current)=>current.includes(role)?current.filter((item)=>item!==role):[...current,role]);}async function submit(event:FormEvent){event.preventDefault();if(saving)return;if(selected.length===0){setError("至少保留一个固定角色。");return;}setSaving(true);setError("");try{const response=await apiFetch(`/api/admin/users/${item.id}/roles`,{method:"PATCH",headers:{"content-type":"application/json"},body:JSON.stringify({roles:selected})});const data=await readResponseJson<{message?:string}>(response,"角色修改响应无效，请重试。");if(!response.ok){setError(data.message??"角色修改失败");return;}await onSaved();}catch(failure){setError(failure instanceof Error?failure.message:"网络异常，角色修改失败，请重试。");}finally{setSaving(false);}}return <Modal title={`修改 ${item.username} 的角色`} note="只能选择固定角色；系统管理员角色本身不增加业务查看或导出权限" onClose={onClose}><form noValidate className="business-form" onSubmit={submit}><fieldset className="role-options"><legend>固定角色</legend>{roles.map((role)=><label key={role.code}><input type="checkbox" checked={selected.includes(role.code)} onChange={()=>toggle(role.code)}/><span>{role.name}</span></label>)}</fieldset>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving||selected.length===0} aria-busy={saving}>{saving?"正在保存…":"保存角色"}</button></div></form></Modal>}
-
-function CreateAccount({roles,onClose,onUncertain,onSaved}:{roles:Array<{code:string;name:string}>;onClose:()=>void;onUncertain:(username:string)=>void;onSaved:()=>Promise<void>}){
-  const[username,setUsername]=useState("");const[displayName,setDisplayName]=useState("");const[role,setRole]=useState(roles[0]?.code??"salesperson");const[personId,setPersonId]=useState("");const[people,setPeople]=useState<PersonOption[]>([]);const[error,setError]=useState("");const[created,setCreated]=useState<{temporaryPassword:string;temporaryPasswordExpiresAt:string}|null>(null);const[saving,setSaving]=useState(false);
-  useEffect(()=>{fetch("/api/admin/people").then(async(response)=>{const data=await response.json() as {people?:PersonOption[];message?:string};if(!response.ok)throw new Error(data.message??"人员列表加载失败");setPeople((data.people??[]).filter((person)=>!person.username));}).catch((reason)=>setError(reason instanceof Error?reason.message:"人员列表加载失败"));},[]);
-  async function submit(event:FormEvent){event.preventDefault();if(saving)return;setSaving(true);setError("");try{const response=await apiFetch("/api/admin/users",{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({username,displayName,roles:[role],personId:personId?Number(personId):null})});if(!response.ok){const data=await response.json().catch(()=>({})) as {message?:string};setError(data.message??"创建账号失败，请重试。");return;}const data=await readResponseJson<{temporaryPassword?:string;temporaryPasswordExpiresAt?:string}>(response,"创建账号响应无效，请重试。");if(!data.temporaryPassword||!data.temporaryPasswordExpiresAt){onUncertain(username);return;}setCreated({temporaryPassword:data.temporaryPassword,temporaryPasswordExpiresAt:data.temporaryPasswordExpiresAt});}catch{onUncertain(username);}finally{setSaving(false);}}
-  return <Modal title="创建系统账号" note="可绑定已有人员身份；临时密码只显示一次，24 小时内必须完成首次改密" onClose={onClose} preventClose={saving}>{created?<div className="temporary-password-result"><p>请立即安全保存临时密码，关闭后无法再次查看。</p><strong>{created.temporaryPassword}</strong><span>失效时间：{new Date(created.temporaryPasswordExpiresAt).toLocaleString("zh-CN")}</span><div className="modal-actions"><button type="button" onClick={async()=>{await onSaved();}}>我已安全保存</button></div></div>:<form noValidate className="business-form" onSubmit={submit}><div className="form-grid"><Field label="登录账号" value={username} onChange={setUsername}/><Field label="账号显示姓名" value={displayName} onChange={setDisplayName}/><label className="field"><span>绑定已有人员（可选）</span><select value={personId} onChange={(event)=>setPersonId(event.target.value)}><option value="">新建人员身份</option>{people.map((person)=><option key={person.id} value={person.id}>{person.displayName}</option>)}</select></label><label className="field"><span>角色</span><select value={role} onChange={(event)=>setRole(event.target.value)}>{roles.map((item)=><option key={item.code} value={item.code}>{item.name}</option>)}</select></label></div>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose} disabled={saving}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在创建…":"创建账号"}</button></div></form>}</Modal>;
-}
-
-function OrganizationPage({user}:{user:User}){
-  const[units,setUnits]=useState<OrgUnit[]>([]);const[assignments,setAssignments]=useState<Assignment[]>([]);const[responsibilities,setResponsibilities]=useState<Responsibility[]>([]);
-  const[message,setMessage]=useState("");const[dialog,setDialog]=useState<"unit"|"assignment"|"transfer"|null>(null);const[closing,setClosing]=useState<Assignment|null>(null);const[replacing,setReplacing]=useState<Responsibility|null>(null);
-  const loadSequence=useRef(0);
-  const isAdmin=user.capabilities.manageOrganization;
-  const load=useCallback(async()=>{const sequence=++loadSequence.current;try{const response=await fetch("/api/organization");const data=await readResponseJson<{units?:OrgUnit[];assignments?:Assignment[];responsibilities?:Responsibility[];message?:string}>(response,"组织架构响应无效，请重试。");if(!response.ok)throw new Error(data.message??"组织架构加载失败");if(sequence!==loadSequence.current)return;setUnits(data.units??[]);setAssignments(data.assignments??[]);setResponsibilities(data.responsibilities??[]);setMessage("");}catch(error){if(sequence===loadSequence.current)setMessage(error instanceof Error?error.message:"组织架构加载失败");throw error;}},[]);
-  useEffect(()=>{load().catch(()=>undefined);},[load]);
-  return <main className="dashboard"><header><div><h1>组织架构</h1><p>组织与人员任职按生效日期管理；业绩入账时固化组织快照</p></div>{isAdmin?<div className="header-actions"><button className="secondary-action" onClick={()=>setDialog("transfer")}>办理组织异动</button><button className="secondary-action" onClick={()=>setDialog("assignment")}>新增任职</button><button className="primary-action" onClick={()=>setDialog("unit")}><Plus size={16}/>新增组织</button></div>:null}</header>{message?<p className="page-message" role="alert">{message}</p>:null}<section className="org-grid"><article className="orders-card"><div className="orders-toolbar"><div><h2>部门与小组</h2><span>{units.length} 个组织单元</span></div></div><div className="compact-list">{units.map((unit)=><div key={unit.id}><span>{unit.unitType==="department"?"部门":"小组"}</span><strong>{unit.name}</strong><small>{unit.parentName??"顶层"}</small></div>)}</div></article><article className="orders-card"><div className="orders-toolbar"><div><h2>人员任职</h2><span>{assignments.length} 条有效期记录</span></div></div><div className="compact-list">{assignments.length?assignments.map((item)=><div key={item.id}><span>{item.effectiveFrom}{item.effectiveTo?` 至 ${item.effectiveTo}`:" 起"}</span><strong>{item.displayName}</strong><div className="compact-list-detail"><small>{[item.departmentName,item.groupName].filter(Boolean).join(" / ")}</small>{isAdmin&&!item.effectiveTo?<button type="button" onClick={()=>setClosing(item)}>关闭任职</button>:null}</div></div>):<p className="empty-cell">暂无人员任职记录</p>}</div></article>{isAdmin?<article className="orders-card responsibility-card"><div className="orders-toolbar"><div><h2>负责人职责</h2><span>{responsibilities.length} 条有效期记录</span></div></div><div className="compact-list responsibility-list">{responsibilities.length?responsibilities.map((item)=><div key={item.id}><span>{item.effectiveFrom}{item.effectiveTo?` 至 ${item.effectiveTo}`:" 起"}</span><strong>{item.displayName}</strong><div className="compact-list-detail"><small>{item.unitName} · {item.responsibilityType==="leader"?"小组负责人":"部门主管"}</small>{!item.effectiveTo?<button type="button" onClick={()=>setReplacing(item)}>更换负责人</button>:null}</div></div>):<p className="empty-cell">暂无负责人职责记录</p>}</div></article>:null}</section>{dialog==="unit"?<CreateOrgUnit units={units} onClose={()=>setDialog(null)} onSaved={async()=>{await load();setDialog(null);}}/>:null}{dialog==="assignment"||dialog==="transfer"?<CreateAssignment units={units} closePrevious={dialog==="transfer"} onClose={()=>setDialog(null)} onSaved={async()=>{await load();setDialog(null);}}/>:null}{closing?<CloseAssignment assignment={closing} onClose={()=>setClosing(null)} onSaved={async(effectiveTo)=>{setAssignments((current)=>current.map((item)=>item.id===closing.id?{...item,effectiveTo}:item));setClosing(null);try{await load();}catch{setMessage("任职已关闭，但组织列表刷新失败，请刷新页面重试。");}}}/>:null}{replacing?<ReplaceResponsibility responsibility={replacing} onClose={()=>setReplacing(null)} onSaved={async(predecessorEffectiveTo)=>{setResponsibilities((current)=>current.map((item)=>item.id===replacing.id?{...item,effectiveTo:predecessorEffectiveTo}:item));setReplacing(null);try{await load();}catch{setMessage("负责人已更换，但组织列表刷新失败，请刷新页面重试。");}}}/>:null}</main>
-}
-
-function CreateOrgUnit({units,onClose,onSaved}:{units:OrgUnit[];onClose:()=>void;onSaved:()=>Promise<void>}){
-  const[name,setName]=useState("");const[unitType,setUnitType]=useState<"department"|"group">("department");const[parentId,setParentId]=useState("");const[error,setError]=useState("");const[saving,setSaving]=useState(false);
-  async function submit(event:FormEvent){event.preventDefault();if(saving)return;setSaving(true);setError("");try{const response=await apiFetch("/api/admin/organization/units",{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({name,unitType,parentId:parentId?Number(parentId):null})});const data=await readResponseJson<{message?:string}>(response,"新增组织响应无效，请重试。");if(!response.ok){setError(data.message??"新增组织失败");return;}await onSaved();}catch(reason){setError(reason instanceof Error?reason.message:"网络异常，新增组织失败，请重试。");}finally{setSaving(false);}}
-  return <Modal title="新增组织单元" note="小组必须归属于一个部门" onClose={onClose} preventClose={saving}><form noValidate className="business-form" onSubmit={submit}><Field label="名称" value={name} onChange={setName}/><label className="field"><span>类型</span><select value={unitType} onChange={(e)=>setUnitType(e.target.value as "department"|"group")}><option value="department">部门</option><option value="group">小组</option></select></label>{unitType==="group"?<label className="field"><span>所属部门</span><select required value={parentId} onChange={(e)=>setParentId(e.target.value)}><option value="">请选择</option>{units.filter((u)=>u.unitType==="department").map((u)=><option key={u.id} value={u.id}>{u.name}</option>)}</select></label>:null}{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose} disabled={saving}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在保存…":"保存组织"}</button></div></form></Modal>;
-}
-
-function CreateAssignment({units,closePrevious=false,onClose,onSaved}:{units:OrgUnit[];closePrevious?:boolean;onClose:()=>void;onSaved:()=>Promise<void>}){
-  const[people,setPeople]=useState<PersonOption[]>([]);const[personId,setPersonId]=useState("");const[departmentId,setDepartmentId]=useState("");const[groupId,setGroupId]=useState("");const[leaderPersonId,setLeaderPersonId]=useState("");const[supervisorPersonId,setSupervisorPersonId]=useState("");const[effectiveFrom,setEffectiveFrom]=useState(businessDateToday);const[error,setError]=useState("");const[saving,setSaving]=useState(false);
-  useEffect(()=>{fetch("/api/admin/people").then(async(response)=>{const data=await response.json() as {people?:PersonOption[];message?:string};if(!response.ok)throw new Error(data.message??"人员列表加载失败");setPeople(data.people??[]);}).catch((reason)=>setError(reason instanceof Error?reason.message:"人员列表加载失败"));},[]);
-  async function submit(event:FormEvent){event.preventDefault();if(saving)return;setSaving(true);setError("");try{const response=await apiFetch("/api/admin/organization/assignments",{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({personId:Number(personId),departmentId:Number(departmentId),groupId:Number(groupId),leaderPersonId:Number(leaderPersonId),supervisorPersonId:Number(supervisorPersonId),effectiveFrom,effectiveTo:null,closePrevious})});const data=await readResponseJson<{message?:string}>(response,"任职保存响应无效，请重试。");if(!response.ok){setError(data.message??"任职保存失败");return;}await onSaved();}catch(reason){setError(reason instanceof Error?reason.message:"网络异常，任职保存失败，请重试。");}finally{setSaving(false);}}
-  const personSelect=(label:string,value:string,onChange:(value:string)=>void)=><label className="field"><span>{label}</span><select required value={value} onChange={(event)=>onChange(event.target.value)}><option value="">请选择</option>{people.map((item)=><option key={item.id} value={item.id}>{item.displayName}{item.username?`（${item.username}）`:"（无登录账号）"}</option>)}</select></label>;
-  return <Modal title={closePrevious?"办理组织异动":"新增人员任职"} note={closePrevious?"系统将于新任职生效日前一天结束原任职；历史业绩归属不会重写":"成员归属与负责人职责均按生效日期保存；新任职不改写历史事件"} onClose={onClose} preventClose={saving}><form noValidate className="business-form" onSubmit={submit}><div className="form-grid">{personSelect("任职人员",personId,setPersonId)}<Field label="生效日期" value={effectiveFrom} type="date" onChange={setEffectiveFrom}/><label className="field"><span>部门</span><select required value={departmentId} onChange={(event)=>{setDepartmentId(event.target.value);setGroupId("");}}><option value="">请选择</option>{units.filter((unit)=>unit.unitType==="department").map((unit)=><option key={unit.id} value={unit.id}>{unit.name}</option>)}</select></label><label className="field"><span>小组</span><select required value={groupId} onChange={(event)=>setGroupId(event.target.value)}><option value="">请选择</option>{units.filter((unit)=>unit.unitType==="group"&&unit.parentId===departmentId).map((unit)=><option key={unit.id} value={unit.id}>{unit.name}</option>)}</select></label>{personSelect("小组负责人",leaderPersonId,setLeaderPersonId)}{personSelect("部门主管",supervisorPersonId,setSupervisorPersonId)}</div>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose} disabled={saving}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在保存…":closePrevious?"确认异动":"保存任职"}</button></div></form></Modal>;
-}
-
-function CloseAssignment({assignment,onClose,onSaved}:{assignment:Assignment;onClose:()=>void;onSaved:(effectiveTo:string)=>Promise<void>}){
-  const[effectiveOn,setEffectiveOn]=useState(businessDateToday);const[error,setError]=useState("");const[saving,setSaving]=useState(false);
-  async function submit(event:FormEvent){event.preventDefault();setSaving(true);setError("");try{const response=await apiFetch(`/api/admin/organization/memberships/${assignment.id}/close`,{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({effectiveOn})});const data=await readResponseJson<{effectiveTo?:string;message?:string}>(response,"任职关闭响应无效，请重试。");if(!response.ok){setError(data.message??"任职关闭失败");return;}if(!data.effectiveTo)throw new Error("任职关闭响应无效，请重试。");await onSaved(data.effectiveTo);}catch(reason){setError(reason instanceof Error?reason.message:"网络异常，任职关闭失败");}finally{setSaving(false);}}
-  return <Modal title="关闭人员任职" note={`${assignment.displayName} · ${[assignment.departmentName,assignment.groupName].filter(Boolean).join(" / ")}；生效日起不再具有该成员归属，历史事件不会改写`} onClose={onClose}><form noValidate className="business-form" onSubmit={submit}><Field label="离任生效日期" value={effectiveOn} type="date" onChange={setEffectiveOn}/>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在关闭…":"确认关闭"}</button></div></form></Modal>;
-}
-
-function ReplaceResponsibility({responsibility,onClose,onSaved}:{responsibility:Responsibility;onClose:()=>void;onSaved:(predecessorEffectiveTo:string)=>Promise<void>}){
-  const[people,setPeople]=useState<PersonOption[]>([]);const[successorPersonId,setSuccessorPersonId]=useState("");const[effectiveOn,setEffectiveOn]=useState(businessDateToday);const[error,setError]=useState("");const[saving,setSaving]=useState(false);
-  useEffect(()=>{fetch("/api/admin/people").then(async(response)=>{const data=await readResponseJson<{people?:PersonOption[];message?:string}>(response,"人员列表响应无效，请重试。");if(!response.ok)throw new Error(data.message??"人员列表加载失败");setPeople(data.people??[]);}).catch((reason)=>setError(reason instanceof Error?reason.message:"人员列表加载失败"));},[]);
-  async function submit(event:FormEvent){event.preventDefault();setSaving(true);setError("");try{const response=await apiFetch(`/api/admin/organization/responsibilities/${responsibility.id}/replace`,{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({successorPersonId,effectiveOn})});const data=await readResponseJson<{predecessorEffectiveTo?:string;message?:string}>(response,"负责人继任响应无效，请重试。");if(!response.ok){setError(data.message??"负责人继任失败");return;}if(!data.predecessorEffectiveTo)throw new Error("负责人继任响应无效，请重试。");await onSaved(data.predecessorEffectiveTo);}catch(reason){setError(reason instanceof Error?reason.message:"网络异常，负责人继任失败");}finally{setSaving(false);}}
-  return <Modal title="更换负责人" note={`${responsibility.unitName} · 现任 ${responsibility.displayName}；旧职责于继任生效日前一天结束，不允许负责人空档`} onClose={onClose}><form noValidate className="business-form" onSubmit={submit}><label className="field"><span>继任负责人</span><select required value={successorPersonId} onChange={(event)=>setSuccessorPersonId(event.target.value)}><option value="">请选择</option>{people.filter((person)=>person.id!==responsibility.personId).map((person)=><option key={person.id} value={person.id}>{person.displayName}{person.username?`（${person.username}）`:"（无登录账号）"}</option>)}</select></label><Field label="继任生效日期" value={effectiveOn} type="date" onChange={setEffectiveOn}/>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在更换…":"确认继任"}</button></div></form></Modal>;
-}
-
-type GoalDialog={kind:"sign"|"approve"|"reject"|"request";goal:Goal};
-type WorkflowDialog={kind:"accept"|"reject";request:GoalChangeRequest}|{kind:"linkage";linkage:GoalLinkageDecision};
-
-function allocationText(goal:Goal){if(goal.allocationType==="balanced")return "已完全分配";const label=goal.allocationType==="overallocated"?"超分配":"未分配";return `${label} ${formatMoney(Math.abs(Number(goal.allocationDifference)))}${goal.allocationRatio?`（${goal.allocationRatio}%）`:""}`;}
-function differenceText(value:string|null){if(value===null)return "—";const number=Number(value);return `${number>0?"+":""}${formatMoney(number)}`;}
-function auditDataText(value:unknown){return value===null||value===undefined?"—":JSON.stringify(value);}
-async function fetchGoalHistory(goalId:string,signal?:AbortSignal){const response=await fetch(`/api/goals/${goalId}/history`,signal?{signal}:{});const result=await readResponseJson<GoalHistory&{message?:string}>(response,"目标历史响应无效，请重试。");if(!response.ok)throw new Error(result.message??"目标历史加载失败");return result;}
-
-function GoalsPage({ user, pendingOnly }: { user: User; pendingOnly: boolean }) {
-  const [goals,setGoals]=useState<Goal[]>([]);const[workflows,setWorkflows]=useState<GoalWorkflows>({changeRequests:[],linkageDecisions:[]});
-  const[showCreate,setShowCreate]=useState(false);const[message,setMessage]=useState("");const[loading,setLoading]=useState(true);
-  const[formalGoal,setFormalGoal]=useState<Goal|null>(null);const[historyGoal,setHistoryGoal]=useState<Goal|null>(null);
-  const[goalDialog,setGoalDialog]=useState<GoalDialog|null>(null);const[workflowDialog,setWorkflowDialog]=useState<WorkflowDialog|null>(null);const[busyWorkflowId,setBusyWorkflowId]=useState<string|null>(null);
-  const loadRevision=useRef(0);
-  const canCreate=user.roles.some((role)=>["sales_manager","sales_supervisor","sales_leader"].includes(role));
-  const load=useCallback(async()=>{const revision=++loadRevision.current;setLoading(true);setGoals([]);setWorkflows({changeRequests:[],linkageDecisions:[]});try{const [goalResponse,workflowResponse]=await Promise.all([fetch(pendingOnly?"/api/goals?pendingOnly=true":"/api/goals"),pendingOnly?fetch("/api/goal-workflows"):Promise.resolve(null)]);const goalData=await readResponseJson<{goals?:Goal[];message?:string}>(goalResponse,"目标响应无效，请重试。");if(!goalResponse.ok)throw new Error(goalData.message??"目标加载失败");let workflowData:GoalWorkflows={changeRequests:[],linkageDecisions:[]};if(workflowResponse){const parsed=await readResponseJson<GoalWorkflows&{message?:string}>(workflowResponse,"目标待办响应无效，请重试。");if(!workflowResponse.ok)throw new Error(parsed.message??"目标待办加载失败");workflowData={changeRequests:parsed.changeRequests??[],linkageDecisions:parsed.linkageDecisions??[]};}if(revision!==loadRevision.current)return;setGoals(goalData.goals??[]);setWorkflows(workflowData);}catch(error){if(revision!==loadRevision.current)return;setGoals([]);setWorkflows({changeRequests:[],linkageDecisions:[]});throw error;}finally{if(revision===loadRevision.current)setLoading(false);}},[pendingOnly]);
-  useEffect(()=>{load().catch((error)=>setMessage(error instanceof Error?error.message:"目标加载失败"));},[load]);
-  const refreshed=async(note:string)=>{setGoalDialog(null);setWorkflowDialog(null);try{await load();setMessage(note);}catch{setMessage(`${note} 列表刷新失败，请手动刷新页面。`);}};
-  async function withdraw(item:GoalChangeRequest){if(busyWorkflowId)return;setBusyWorkflowId(item.id);setMessage("");try{const response=await apiFetch(`/api/goal-change-requests/${item.id}/withdraw`,{method:"POST",headers:{"content-type":"application/json"},body:"{}"});const data=await readResponseJson<{message?:string}>(response,"撤回响应无效，请重试。");if(!response.ok){if(response.status===403||response.status===409){await refreshed("状态已变化，已重新读取权威状态。");return;}setMessage(data.message??"撤回失败");return;}await refreshed("修改申请已撤回，可以基于当前生效目标重新申请。");}catch{await refreshed("操作结果不确定，已重新读取权威状态。");}finally{setBusyWorkflowId(null);}}
-  return <main className="dashboard goals-page"><header><div><h1>{pendingOnly?"审批中心":"目标管理"}</h1><p>{pendingOnly?"处理目标审批、修改申请和逐级联动选择":"按组织责任逐级下达，实名确认和审批均绑定具体版本"}</p></div><div className="header-actions">{user.capabilities.exportGoals?<button className="secondary-action" onClick={()=>{window.location.href="/api/exports/goals.csv"}}>导出目标</button>:null}{canCreate&&!pendingOnly?<button className="primary-action" onClick={()=>setShowCreate(true)}><Plus size={16}/>下达目标</button>:null}</div></header>{message?<p className="page-message" role="status">{message}</p>:null}<section className="orders-card" aria-labelledby="goal-table-title"><div className="orders-toolbar"><div><h2 id="goal-table-title">{pendingOnly?"待确认与待审批目标":"目标责任台账"}</h2><span role="status">{loading?"正在读取…":`${goals.length} 条记录`}</span></div></div><div className="orders-table-wrap"><table><thead><tr><th>月份</th><th>层级 / 范围</th><th>责任人</th><th>目标</th><th>直接下级</th><th>分配差额</th><th>状态</th><th>操作</th></tr></thead><tbody>{!loading&&goals.length===0?<tr><td colSpan={8} className="empty-cell">{pendingOnly?"当前没有待确认或待审批目标。":"暂无目标；有下达权限的负责人可创建第一条目标。"}</td></tr>:goals.map((goal)=><tr key={goal.id}><td>{goal.periodMonth}</td><td><strong>{goalLevelName(goal.level)}</strong>{goal.orgUnitName?<small className="table-secondary">{goal.orgUnitName}</small>:null}</td><td>{goal.ownerName}</td><td>{formatMoney(goal.amount)}{goal.effectiveAmount!==null&&goal.effectiveAmount!==goal.amount?<small className="table-secondary">当前生效 {formatMoney(goal.effectiveAmount)}</small>:null}</td><td>{formatMoney(goal.allocatedAmount)}</td><td className={goal.allocationType==="overallocated"?"warning":""}>{allocationText(goal)}</td><td><span className={`status goal-${goal.status}`}>{goalStatusName(goal.status)}</span></td><td><div className="row-actions">{goal.status==="active"?<button onClick={()=>setFormalGoal(goal)}>查看正式报表</button>:null}<button onClick={()=>setHistoryGoal(goal)}>版本与记录</button>{goal.ownerPersonId===user.personId&&goal.status==="pending_signature"?<button onClick={()=>setGoalDialog({kind:"sign",goal})}>确认目标</button>:null}{goal.ownerPersonId===user.personId&&goal.level!=="sales_manager"&&goal.status==="active"?<button onClick={()=>setGoalDialog({kind:"request",goal})}>申请修改</button>:null}{goal.status==="pending_gm"&&user.roles.includes("general_manager")?<><button onClick={()=>setGoalDialog({kind:"approve",goal})}>批准</button><button className="danger-action" onClick={()=>setGoalDialog({kind:"reject",goal})}>拒绝</button></>:null}{goal.status==="pending_hr"&&user.roles.includes("hr")?<><button onClick={()=>setGoalDialog({kind:"approve",goal})}>批准</button><button className="danger-action" onClick={()=>setGoalDialog({kind:"reject",goal})}>拒绝</button></>:null}</div></td></tr>)}</tbody></table></div></section>{pendingOnly?<><WorkflowTable title="目标修改申请" empty="当前没有相关修改申请。" rows={workflows.changeRequests.map((item)=><tr key={item.id}><td>{goalLevelName(item.level)} · {item.ownerName}</td><td>{item.requestedByName}</td><td>{item.currentAmount?formatMoney(item.currentAmount):"—"}</td><td>{item.requestedAmount?formatMoney(item.requestedAmount):"由直属上级填写"}</td><td>{item.newAmount?formatMoney(item.newAmount):"尚未生成"}</td><td>{differenceText(item.amountDifference)}</td><td className="wrap-cell">{item.reason}{item.outcomeComment?<small className="table-secondary">处理意见：{item.outcomeComment}</small>:null}</td><td><span className={`status workflow-${item.status}`}>{workflowStatusName(item.status)}</span></td><td><div className="row-actions">{item.canHandle?<><button onClick={()=>setWorkflowDialog({kind:"accept",request:item})}>接受并填金额</button><button className="danger-action" onClick={()=>setWorkflowDialog({kind:"reject",request:item})}>拒绝</button></>:null}{item.canWithdraw?<button disabled={busyWorkflowId!==null} aria-busy={busyWorkflowId===item.id} onClick={()=>withdraw(item)}>{busyWorkflowId===item.id?"正在撤回…":"撤回"}</button>:null}</div></td></tr>)} headers={["目标","申请人","当前生效","建议金额","候选版本","版本差异","原因 / 意见","状态","操作"]}/><WorkflowTable title="目标联动选择" empty="当前没有待处理的目标联动。" rows={workflows.linkageDecisions.map((item)=><tr key={item.id}><td>{goalLevelName(item.parentLevel)}</td><td>{item.childOwnerName}</td><td>{formatMoney(item.childAmount)}</td><td>{formatMoney(item.parentAmount)}</td><td><span className={`status workflow-${item.status}`}>{workflowStatusName(item.status)}</span></td><td><div className="row-actions">{item.canDecide?<button onClick={()=>setWorkflowDialog({kind:"linkage",linkage:item})}>选择是否调整</button>:null}</div></td></tr>)} headers={["本级目标","变化的下级责任人","下级新目标","本级生效目标","状态","操作"]}/></>:null}{showCreate?<CreateGoal user={user} onClose={()=>setShowCreate(false)} onSaved={async()=>{setShowCreate(false);try{await load();setMessage("目标已下达，等待责任人确认。");}catch{setMessage("目标已下达，但列表刷新失败，请手动刷新页面。");}}}/>:null}{goalDialog?<GoalActionDialog dialog={goalDialog} onClose={()=>setGoalDialog(null)} onSaved={()=>refreshed(goalDialog.kind==="request"?"修改申请已提交。":"目标状态已更新。")} onStale={()=>refreshed("状态已变化，已重新读取权威状态。")} onUncertain={()=>refreshed("操作结果不确定，已重新读取权威状态。")}/>:null}{workflowDialog?<WorkflowActionDialog dialog={workflowDialog} onClose={()=>setWorkflowDialog(null)} onSaved={()=>refreshed("目标待办已处理。")} onStale={()=>refreshed("状态已变化，已重新读取权威状态。")} onUncertain={()=>refreshed("操作结果不确定，已重新读取权威状态。")}/>:null}{historyGoal?<GoalHistoryDialog goal={historyGoal} onClose={()=>setHistoryGoal(null)}/>:null}{formalGoal?<FormalReportDialog goal={formalGoal} onClose={()=>setFormalGoal(null)}/>:null}</main>;
-}
-
-function WorkflowTable({title,empty,headers,rows}:{title:string;empty:string;headers:string[];rows:ReactNode[]}){return <section className="orders-card workflow-card"><div className="orders-toolbar"><div><h2>{title}</h2><span>{rows.length} 条记录</span></div></div><div className="orders-table-wrap"><table><thead><tr>{headers.map((header)=><th key={header}>{header}</th>)}</tr></thead><tbody>{rows.length?rows:<tr><td colSpan={headers.length} className="empty-cell">{empty}</td></tr>}</tbody></table></div></section>}
-
-function GoalActionDialog({dialog,onClose,onSaved,onStale,onUncertain}:{dialog:GoalDialog;onClose:()=>void;onSaved:()=>Promise<void>;onStale:()=>Promise<void>;onUncertain:()=>Promise<void>}){
-  const[reason,setReason]=useState("");const[requestedAmount,setRequestedAmount]=useState("");const[comment,setComment]=useState("");const[error,setError]=useState("");const[saving,setSaving]=useState(false);
-  const[history,setHistory]=useState<GoalHistory|null>(null);const[historyError,setHistoryError]=useState("");const[historyRevision,setHistoryRevision]=useState(0);
-  const needsHistory=dialog.kind==="approve"||dialog.kind==="reject";
-  const title=dialog.kind==="sign"?"确认目标版本":dialog.kind==="request"?"申请修改目标":dialog.kind==="approve"?"批准目标":"拒绝目标";
-  useEffect(()=>{if(!needsHistory)return;const controller=new AbortController();setHistory(null);setHistoryError("");fetchGoalHistory(dialog.goal.id,controller.signal).then(setHistory).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setHistoryError(failure instanceof Error?failure.message:"目标历史加载失败");});return()=>controller.abort();},[dialog.goal.id,historyRevision,needsHistory]);
-  const currentVersion=history?.versions.find((version)=>version.id===dialog.goal.versionId)??null;
-  const previousVersion=currentVersion?history?.versions.find((version)=>Number(version.versionNo)===Number(currentVersion.versionNo)-1)??null:null;
-  const currentApprovals=currentVersion?history?.approvals.filter((approval)=>approval.goalVersionId===currentVersion.id)??[]:[];
-  const latestRequest=history?.changeRequests[0]??null;
-  async function submit(event:FormEvent){event.preventDefault();if(saving||needsHistory&&!currentVersion)return;setSaving(true);setError("");let url=`/api/goals/${dialog.goal.id}/decision`;let body:Record<string,unknown>={expectedVersionId:dialog.goal.versionId,decision:dialog.kind==="reject"?"rejected":"approved",comment};if(dialog.kind==="sign"){url=`/api/goal-versions/${dialog.goal.versionId}/confirm`;body={};}else if(dialog.kind==="request"){url=`/api/goals/${dialog.goal.id}/change-requests`;body={reason,...(requestedAmount?{requestedAmount:Number(requestedAmount)}:{})};}try{const response=await apiFetch(url,{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify(body)});const data=await readResponseJson<{message?:string}>(response,"服务响应无效，请重试。");if(!response.ok){if(response.status===403||response.status===409){await onStale();return;}setError(data.message??"操作失败");return;}await onSaved();}catch{await onUncertain();}finally{setSaving(false);}}
-  return <Modal title={title} note={`${dialog.goal.periodMonth} · ${goalLevelName(dialog.goal.level)} · 版本 ${dialog.goal.versionNo}`} onClose={onClose}><form noValidate className="business-form" onSubmit={submit}>{needsHistory?historyError?<section className="approval-evidence"><p className="form-error" role="alert">{historyError}</p><button type="button" onClick={()=>setHistoryRevision((value)=>value+1)}>重新读取审批依据</button></section>:!history?<p role="status">正在读取审批依据…</p>:!currentVersion?<section className="approval-evidence"><p className="form-error" role="alert">当前待办版本已变化，请返回审批中心重新处理。</p><button type="button" onClick={onStale}>重新读取审批中心</button></section>:<><section className="approval-evidence" aria-labelledby="approval-evidence-title"><h3 id="approval-evidence-title">审批依据</h3><div className="approval-facts"><p>原目标 <strong>{currentVersion.previousAmount===null?"—":formatMoney(currentVersion.previousAmount)}</strong></p><p>当前目标 <strong>{formatMoney(currentVersion.amount)}</strong></p><p>差异 <strong>{differenceText(currentVersion.amountDifference)}</strong></p><p>状态 <strong>{previousVersion?goalStatusName(previousVersion.status):"初始版本"} → {goalStatusName(currentVersion.status)}</strong></p></div><div className="approval-confirmation"><strong>目标确认</strong><p>确认人 {currentVersion.signedByName??"尚未确认"}{currentVersion.signedAt?` · ${formatOperationTime(currentVersion.signedAt)}`:""}</p><p>{currentVersion.signatureText??"未留下确认声明"}</p></div><div className="approval-opinions"><h4>原因与既有意见</h4><p>版本原因：{currentVersion.changeReason}</p>{latestRequest?<p>修改申请：{latestRequest.reason}{latestRequest.outcomeComment?` · 处理意见 ${latestRequest.outcomeComment}`:""}</p>:null}{currentApprovals.map((approval)=><p key={approval.id}>{approval.stage==="general_manager"?"总经理":"人事部"} {approval.decidedByName}：{approval.comment||"未填写意见"}</p>)}</div></section><label className="field"><span>审批意见</span><textarea className="resize-none" rows={4} value={comment} onChange={(event)=>setComment(event.target.value)} placeholder={dialog.kind==="reject"?"拒绝时必须说明原因":"填写本次审批意见"}/></label></>:dialog.kind==="sign"?<section className="form-note" aria-labelledby="goal-confirmation-statement"><strong id="goal-confirmation-statement">固定确认声明</strong><p>{GOAL_CONFIRMATION_STATEMENT}</p><p>责任人：{dialog.goal.ownerName} · 目标金额：{formatMoney(dialog.goal.amount)}</p></section>:<div className="form-grid"><Field label="修改原因" value={reason} onChange={setReason}/><label className="field"><span>建议金额（可选）</span><input value={requestedAmount} type="number" min="0" step="0.01" onChange={(event)=>setRequestedAmount(event.target.value)}/></label></div>}{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving||needsHistory&&!currentVersion} aria-busy={saving}>{saving?"正在提交…":dialog.kind==="sign"?"确认目标":dialog.kind==="request"?"提交修改申请":dialog.kind==="approve"?"确认批准":"确认拒绝"}</button></div></form></Modal>;
-}
-
-function WorkflowActionDialog({dialog,onClose,onSaved,onStale,onUncertain}:{dialog:WorkflowDialog;onClose:()=>void;onSaved:()=>Promise<void>;onStale:()=>Promise<void>;onUncertain:()=>Promise<void>}){const isLinkage=dialog.kind==="linkage";const[decision,setDecision]=useState<"keep_parent"|"adjust_parent">("keep_parent");const[newAmount,setNewAmount]=useState(!isLinkage&&dialog.kind==="accept"?(dialog.request.requestedAmount??""):"");const[reason,setReason]=useState("");const[error,setError]=useState("");const[saving,setSaving]=useState(false);async function submit(event:FormEvent){event.preventDefault();if(saving)return;setSaving(true);setError("");let url:string;let body:Record<string,unknown>;if(isLinkage){url=`/api/goal-linkage-decisions/${dialog.linkage.id}/decide`;body={decision,reason,...(decision==="adjust_parent"&&dialog.linkage.parentLevel==="sales_manager"?{newAmount:Number(newAmount)}:{})};}else{url=`/api/goal-change-requests/${dialog.request.id}/${dialog.kind}`;body=dialog.kind==="accept"?{newAmount:Number(newAmount),comment:reason}:{comment:reason};}try{const response=await apiFetch(url,{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify(body)});const data=await readResponseJson<{message?:string}>(response,"服务响应无效，请重试。");if(!response.ok){if(response.status===403||response.status===409){await onStale();return;}setError(data.message??"待办处理失败");return;}await onSaved();}catch{await onUncertain();}finally{setSaving(false);}}const title=isLinkage?"处理目标联动":dialog.kind==="accept"?"接受修改申请":"拒绝修改申请";return <Modal title={title} note={isLinkage?"不调整会保留本级确认；调整会逐级发起申请，绝不自动改金额":"直属目标下达人填写新金额，责任人需要重新确认"} onClose={onClose}><form noValidate className="business-form" onSubmit={submit}>{isLinkage?<><label className="field"><span>处理方式</span><select value={decision} onChange={(event)=>setDecision(event.target.value as "keep_parent"|"adjust_parent")}><option value="keep_parent">不调整本级目标，保留原确认</option><option value="adjust_parent">调整本级目标，发起上一级流程</option></select></label>{decision==="adjust_parent"&&dialog.linkage.parentLevel==="sales_manager"?<Field label="销售经理总目标新金额" value={newAmount} type="number" onChange={setNewAmount}/>:null}</>:dialog.kind==="accept"?<Field label="新目标金额" value={newAmount} type="number" onChange={setNewAmount}/>:null}<label className="field"><span>{isLinkage?"联动原因":"处理意见"}</span><textarea className="resize-none" rows={4} value={reason} onChange={(event)=>setReason(event.target.value)}/></label>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在提交…":isLinkage?"确认联动选择":dialog.kind==="accept"?"接受并创建新版本":"确认拒绝"}</button></div></form></Modal>}
-
-function GoalHistoryDialog({goal,onClose}:{goal:Goal;onClose:()=>void}){const[data,setData]=useState<GoalHistory|null>(null);const[error,setError]=useState("");useEffect(()=>{const controller=new AbortController();fetchGoalHistory(goal.id,controller.signal).then(setData).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setError(failure instanceof Error?failure.message:"目标历史加载失败");});return()=>controller.abort();},[goal.id]);return <Modal title="目标版本与操作记录" note={`${goal.periodMonth} · ${goal.ownerName} · 每次确认和审批均绑定具体版本`} onClose={onClose}>{error?<p className="form-error history-content" role="alert">{error}</p>:!data?<p className="history-content">正在读取版本与审计记录…</p>:<div className="history-content"><section><h3>版本与差异</h3><ol className="history-list">{data.versions.map((version)=><li key={version.id}><div><strong>版本 {version.versionNo} · {formatMoney(version.amount)}</strong><span className={`status goal-${version.status}`}>{goalStatusName(version.status)}</span></div><p>{version.previousAmount===null?"初始版本":`基准 ${formatMoney(version.previousAmount)} · 差异 ${differenceText(version.amountDifference)}`} · {version.changeReason} · 录入 {version.createdByName} · 确认者 {version.signedByName??"尚未确认"}</p></li>)}</ol></section><section><h3>修改申请</h3>{data.changeRequests.length?<ol className="history-list">{data.changeRequests.map((request)=><li key={request.id}><div><strong>{request.requestedByName} · {workflowStatusName(request.status)}</strong><time>{formatOperationTime(request.createdAt)}</time></div><p>{request.reason} · 建议金额 {request.requestedAmount?formatMoney(request.requestedAmount):"未填写"} · 处理人 {request.handledByName??"尚未处理"}{request.outcomeComment?` · 意见 ${request.outcomeComment}`:""}</p></li>)}</ol>:<p>暂无修改申请。</p>}</section><section><h3>审批</h3>{data.approvals.length?<ol className="history-list">{data.approvals.map((approval)=><li key={approval.id}><div><strong>{approval.stage==="general_manager"?"总经理":"人事部"} · {approval.decision==="approved"?"批准":"拒绝"}</strong><time>{formatOperationTime(approval.decidedAt)}</time></div><p>{approval.decidedByName} · {approval.comment}</p></li>)}</ol>:<p>暂无审批记录。</p>}</section><section><h3>审计轨迹</h3><ol className="history-list">{data.audit.map((item)=><li key={item.id}><div><strong>{goalAuditName(item.action)}</strong><time>{formatOperationTime(item.createdAt)}</time></div><p>{item.actorName??"系统"} · 变更前 {auditDataText(item.beforeData)} · 变更后 {auditDataText(item.afterData)}</p></li>)}</ol></section></div>}</Modal>}
-
-function FormalReportDialog({goal,onClose}:{goal:Goal;onClose:()=>void}){
-  const[report,setReport]=useState<FormalReport|null>(null);const[error,setError]=useState("");
-  useEffect(()=>{fetch(`/api/performance/formal-reports/${goal.id}`).then(async(response)=>{const data=await response.json() as FormalReport&{message?:string};if(!response.ok)throw new Error(data.message??"正式报表加载失败");setReport(data);}).catch((reason)=>setError(reason instanceof Error?reason.message:"正式报表加载失败"));},[goal.id]);
-  return <Modal title="正式业绩报表" note="目标已生效，达成率按该层级目标独立计算" onClose={onClose}>{error?<p className="form-error formal-report-content">{error}</p>:report?<div className="formal-report-content"><div><span>目标月份</span><strong>{report.periodMonth}</strong></div><div><span>责任层级</span><strong>{goalLevelName(report.level)}</strong></div><div><span>生效目标</span><strong>{formatMoney(report.targetAmount)}</strong></div><div><span>实际业绩</span><strong>{formatMoney(report.actualAmount)}</strong></div><div><span>目标差距</span><strong>{formatMoney(report.gapAmount)}</strong></div><div><span>达成率</span><strong>{report.achievementRate===null?"不计算":`${report.achievementRate}%`}</strong></div><div className="modal-actions"><button type="button" onClick={()=>{window.location.href=`/api/exports/formal-reports/${goal.id}.csv`;}}>导出正式报表</button></div></div>:<p className="formal-report-content">正在计算正式报表…</p>}</Modal>;
-}
-
-function CreateGoal({user,onClose,onSaved}:{user:User;onClose:()=>void;onSaved:()=>Promise<void>}){const available:GoalLevel[]=user.roles.includes("sales_manager")?["sales_manager","department"]:user.roles.includes("sales_supervisor")?["group"]:["personal"];const[periodMonth,setPeriodMonth]=useState(businessDateToday().slice(0,7));const[level,setLevel]=useState<GoalLevel>(available[0]!);const[parentGoalId,setParentGoalId]=useState("");const[parentGoals,setParentGoals]=useState<ParentGoalOption[]>([]);const[owners,setOwners]=useState<GoalOption[]>([]);const[ownerKey,setOwnerKey]=useState("");const[amount,setAmount]=useState("");const[changeReason,setChangeReason]=useState("目标下达");const[error,setError]=useState("");const[loading,setLoading]=useState(true);const[saving,setSaving]=useState(false);useEffect(()=>{const controller=new AbortController();setLoading(true);setError("");const params=new URLSearchParams({periodMonth,level});if(parentGoalId)params.set("parentGoalId",parentGoalId);fetch(`/api/goals/options?${params}`,{signal:controller.signal}).then(async(response)=>{const data=await readResponseJson<{parentGoals?:ParentGoalOption[];owners?:GoalOption[];message?:string}>(response,"目标选项响应无效，请重试。");if(!response.ok)throw new Error(data.message??"目标选项加载失败");const nextParents=data.parentGoals??[];const nextOwners=data.owners??[];setParentGoals(nextParents);setOwners(nextOwners);if(level!=="sales_manager"&&!parentGoalId&&nextParents[0]){setParentGoalId(nextParents[0].id);return;}const keys=nextOwners.map((owner)=>`${owner.personId}:${owner.orgUnitId??""}`);setOwnerKey((current)=>keys.includes(current)?current:(keys[0]??""));}).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setError(failure instanceof Error?failure.message:"目标选项加载失败");}).finally(()=>{if(!controller.signal.aborted)setLoading(false);});return()=>controller.abort();},[periodMonth,level,parentGoalId]);const selectedOwner=owners.find((owner)=>`${owner.personId}:${owner.orgUnitId??""}`===ownerKey);async function submit(event:FormEvent){event.preventDefault();if(saving)return;if(!selectedOwner||!amount||!changeReason.trim()){setError("请选择合法责任人，并填写目标金额和下达原因。");return;}setSaving(true);setError("");try{const response=await apiFetch("/api/goals",{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({periodMonth,level,ownerPersonId:Number(selectedOwner.personId),orgUnitId:selectedOwner.orgUnitId?Number(selectedOwner.orgUnitId):null,parentGoalId:parentGoalId?Number(parentGoalId):null,amount:Number(amount),changeReason})});const data=await readResponseJson<{message?:string}>(response,"服务响应无效，请重试。");if(!response.ok){setError(data.message??"目标下达失败");return;}await onSaved();}catch(failure){setError(failure instanceof Error?failure.message:"网络异常，目标下达失败，请重试。");}finally{setSaving(false);}}return <Modal title="下达目标" note="责任人从当前月有效组织关系中选择；服务端再次校验直属关系" onClose={onClose}><form noValidate className="business-form" onSubmit={submit}><div className="form-grid"><Field label="目标月份" value={periodMonth} type="month" onChange={(value)=>{setPeriodMonth(value);setParentGoalId("");setOwnerKey("");}}/><label className="field"><span>目标层级</span><select value={level} onChange={(event)=>{setLevel(event.target.value as GoalLevel);setParentGoalId("");setOwnerKey("");}}>{available.map((item)=><option key={item} value={item}>{goalLevelName(item)}</option>)}</select></label>{level!=="sales_manager"?<label className="field"><span>直属上级目标</span><select value={parentGoalId} onChange={(event)=>{setParentGoalId(event.target.value);setOwnerKey("");}}><option value="">请选择已生效的直属上级目标</option>{parentGoals.map((parent)=><option key={parent.id} value={parent.id}>{parent.ownerName} · {parent.orgUnitName??goalLevelName(level)} · {formatMoney(parent.amount)}</option>)}</select></label>:null}<label className="field"><span>目标责任人</span><select value={ownerKey} disabled={loading||owners.length===0} onChange={(event)=>setOwnerKey(event.target.value)}><option value="">{loading?"正在加载合法责任人…":"请选择责任人"}</option>{owners.map((owner)=><option key={`${owner.personId}:${owner.orgUnitId??""}`} value={`${owner.personId}:${owner.orgUnitId??""}`}>{owner.name}{owner.orgUnitName?` · ${owner.orgUnitName}`:""}</option>)}</select></label><Field label="目标金额" value={amount} type="number" onChange={setAmount}/><Field label="下达原因" value={changeReason} onChange={setChangeReason}/></div>{level!=="sales_manager"&&parentGoals.length===0&&!loading?<p className="form-note">当前月份没有可用的已生效直属上级目标，请先完成上一级审批。</p>:null}{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving||loading} aria-busy={saving}>{saving?"正在提交…":"提交待确认"}</button></div></form></Modal>}
-
-function Overview({ canEdit, canExport, onEnterOrders }: { canEdit: boolean; canExport: boolean; onEnterOrders: () => void }) {
-  const [data, setData] = useState<DashboardData | null>(null);
-  const [showPersonalDetails,setShowPersonalDetails]=useState(false);
-  const [personalDetailKind,setPersonalDetailKind]=useState<"actual"|"gap">("actual");
-  const [personalDetails,setPersonalDetails]=useState<PersonalAchievementDrilldown|null>(null);
-  const [personalDetailsError,setPersonalDetailsError]=useState("");
-  const [selectedGroup,setSelectedGroup]=useState<GroupAchievement|null>(null);
-  const [groupDetailKind,setGroupDetailKind]=useState<"actual"|"gap">("actual");
-  const [groupDetails,setGroupDetails]=useState<GroupAchievementDrilldown|null>(null);
-  const [groupDetailsError,setGroupDetailsError]=useState("");
-  const groupDetailsRequest=useRef(0);
-  const [organizationDialog,setOrganizationDialog]=useState<{kind:"sales"|"department";metric:"actual"|"gap";achievement:SalesAchievement|DepartmentAchievement}|null>(null);
-  const [organizationDetails,setOrganizationDetails]=useState<SalesAchievementDrilldown|DepartmentAchievementDrilldown|null>(null);
-  const [organizationDetailsError,setOrganizationDetailsError]=useState("");
-  const organizationDetailsRequest=useRef(0);
-  useEffect(() => { fetch("/api/performance/dashboard").then(async (response) => { if (!response.ok) throw new Error("总览加载失败"); setData(await response.json() as DashboardData); }).catch(() => setData(null)); }, []);
-  const maxGroup = Math.max(1, ...(data?.groups.map((group) => Number(group.total)) ?? [1]));
-  const personal=data?.personalAchievement;
-  const reason=personal?achievementReasonText(personal.calculationReason,"个人"):null;
-  async function openPersonalDetails(kind:"actual"|"gap"){if(!personal)return;setPersonalDetailKind(kind);setShowPersonalDetails(true);setPersonalDetails(null);setPersonalDetailsError("");try{const response=await fetch(`/api/performance/personal-achievement/events?month=${personal.periodMonth}`);const result=await readResponseJson<PersonalAchievementDrilldown&{message?:string}>(response,"个人业绩构成响应无效");if(!response.ok)throw new Error(result.message??"个人业绩构成加载失败");const{events,...achievement}=result;if(kind==="gap"&&result.gapAmount===null)setPersonalDetailKind("actual");setPersonalDetails(result);setData((current)=>current?{...current,personalAchievement:achievement}:current);}catch(error){setPersonalDetailsError(error instanceof Error?error.message:"个人业绩构成加载失败");}}
-  async function openGroupDetails(group:GroupAchievement,kind:"actual"|"gap"){const requestId=++groupDetailsRequest.current;setSelectedGroup(group);setGroupDetailKind(kind);setGroupDetails(null);setGroupDetailsError("");try{const response=await fetch(`/api/performance/group-achievement/events?month=${group.periodMonth}&groupId=${group.groupId}`);const result=await readResponseJson<GroupAchievementDrilldown&{message?:string}>(response,"小组业绩构成响应无效");if(requestId!==groupDetailsRequest.current)return;if(!response.ok)throw new Error(result.message??"小组业绩构成加载失败");const{members,...achievement}=result;if(kind==="gap"&&result.gapAmount===null)setGroupDetailKind("actual");setGroupDetails(result);setData((current)=>current?{...current,groupAchievements:current.groupAchievements.map((item)=>item.groupId===result.groupId?achievement:item)}:current);}catch(error){if(requestId===groupDetailsRequest.current)setGroupDetailsError(error instanceof Error?error.message:"小组业绩构成加载失败");}}
-  async function openOrganizationDetails(kind:"sales"|"department",achievement:SalesAchievement|DepartmentAchievement,metric:"actual"|"gap"){
-    const requestId=++organizationDetailsRequest.current;
-    setOrganizationDialog({kind,achievement,metric});setOrganizationDetails(null);setOrganizationDetailsError("");
-    try{
-      const url=kind==="sales"
-        ? `/api/performance/sales-achievement/events?month=${achievement.periodMonth}`
-        : (achievement as DepartmentAchievement).departmentId===null
-          ? `/api/performance/sales-achievement/events?month=${achievement.periodMonth}`
-          : `/api/performance/department-achievement/events?month=${achievement.periodMonth}&departmentId=${(achievement as DepartmentAchievement).departmentId}`;
-      const response=await fetch(url);
-      const result=await readResponseJson<(SalesAchievementDrilldown|DepartmentAchievementDrilldown)&{message?:string}>(response,"组织业绩构成响应无效");
-      if(requestId!==organizationDetailsRequest.current)return;
-      if(!response.ok)throw new Error(result.message??"组织业绩构成加载失败");
-      const details=kind==="department"&&"departments" in result
-        ? result.departments.find((item)=>item.departmentKey===(achievement as DepartmentAchievement).departmentKey)??null
-        : result;
-      if(!details)throw new Error("部门业绩构成不存在");
-      setOrganizationDetails(details);
-      if(metric==="gap"&&details.gapAmount===null)setOrganizationDialog({kind,achievement,metric:"actual"});
-    }catch(error){if(requestId===organizationDetailsRequest.current)setOrganizationDetailsError(error instanceof Error?error.message:"组织业绩构成加载失败");}
-  }
-  const sales=data?.salesAchievement;
-  const departments=data?.departmentAchievements??[];
-  const organizationDepartments=organizationDetails?("departments" in organizationDetails?organizationDetails.departments:[organizationDetails]):[];
-  const dashboardDescription=sales?"销售组织目标与不可变业绩事件":departments.length?"个人、部门目标与不可变业绩事件":personal?"个人目标与不可变业绩事件":"原始账本，不代表正式绩效结果";
-  return <main className="dashboard"><header><div><h1>业绩账本总览</h1><p>{data ? `${data.month.replace("-", " 年 ")} 月 · ${dashboardDescription}` : "正在加载真实业绩账本…"}</p></div><div className="header-actions">{canExport?<button className="secondary-action" onClick={()=>{window.location.href="/api/exports/performance.csv"}}>导出全部订单</button>:null}{canEdit?<button className="primary-action" onClick={onEnterOrders}>录入订单业绩</button>:null}</div></header>
-      {personal?<><section className="metric-band personal-metrics" aria-label="个人目标达成"><Metric label="生效目标" value={personal.targetAmount===null?"—":formatMoney(personal.targetAmount)} note={reason??"当前月个人目标"}/><Metric label="个人业绩" value={formatMoney(personal.actualAmount)} note={`${personal.eventCount} 条本人事件`} onClick={()=>openPersonalDetails("actual")} actionLabel="查看个人业绩构成"/><Metric label="差距" value={personal.gapAmount===null?"不计算":formatMoney(personal.gapAmount)} note={reason??"生效目标减个人业绩"} onClick={personal.gapAmount===null?undefined:()=>openPersonalDetails("gap")} actionLabel={personal.gapAmount===null?undefined:"查看个人差距构成"}/><Metric label="达成率" value={personal.achievementRate===null?"不计算":`${personal.achievementRate}%`} note={reason??"个人业绩 ÷ 生效目标"}/><Metric label="自然日进度" value={personal.timeProgressRate===null?"不计算":`${personal.timeProgressRate}%`} note={reason??"按上海自然日计算"}/><Metric label="进度偏差" value={personal.progressVariance===null?"不计算":`${personal.progressVariance}%`} note={reason??"达成率减自然日进度"} warning={personal.progressVariance!==null&&Number(personal.progressVariance)<0}/></section>{showPersonalDetails?<Modal title={personalDetailKind==="gap"?"个人差距构成":"个人业绩构成"} note={personalDetails?(personalDetailKind==="gap"?`目标 ${formatMoney(personalDetails.targetAmount!)} − 事件净额 ${formatMoney(personalDetails.actualAmount)} = 差距 ${formatMoney(personalDetails.gapAmount!)}`:`${personalDetails.events.length} 条事件 · 净额 ${formatMoney(personalDetails.actualAmount)}`):`正在读取 ${personal.periodMonth} 个人业绩事件`} onClose={()=>setShowPersonalDetails(false)}>{personalDetailsError?<p className="form-error personal-drilldown" role="alert">{personalDetailsError}</p>:personalDetails?<div className="personal-drilldown"><div className="orders-table-wrap"><table><thead><tr><th>订单</th><th>客户</th><th>不可变事件</th><th>业务日 / 记账月</th><th>金额方向</th><th>状态 / 计入</th><th>组织归属</th><th>订单当前分析资料</th><th>原因</th></tr></thead><tbody>{personalDetails.events.map((event)=><tr key={event.id}><td>{event.orderNo}</td><td>{event.customerName}</td><td>{eventTypeName(event.eventType)} · 第 {event.sequence} 条</td><td>{event.occurredOn} / {event.accountingMonth}</td><td className={event.deltaAmount.startsWith("-")?"negative":""}>{event.deltaAmount.startsWith("-")?"负向":"正向"} {formatMoney(event.deltaAmount.startsWith("-")?event.deltaAmount.slice(1):event.deltaAmount)}</td><td>{event.resultingLifecycleState?<><Status state={event.resultingLifecycleState}/><small className="table-secondary">计入至 {formatMoney(event.resultingCountedAmount)}</small></>:"原始状态未推断"}</td><td>{[event.departmentName,event.groupName].filter(Boolean).join(" / ")||"—"}</td><td>分析维度待补齐</td><td>{event.reason??"—"}</td></tr>)}</tbody></table></div>{personalDetails.events.length===0?<p className="empty-cell">当前月份没有个人业绩事件。</p>:null}</div>:<p className="personal-drilldown">正在读取个人业绩构成…</p>}</Modal>:null}</>:null}
-      {sales?<OrganizationAchievementCard scope="销售组织" title="销售组织目标达成" countNote={`${sales.departmentCount} 个部门 · 当前仅一个销售组织`} achievement={sales} onOpen={(metric)=>openOrganizationDetails("sales",sales,metric)}/>:null}
-      {departments.map((department)=><OrganizationAchievementCard key={department.departmentKey} scope="部门" title={`部门目标达成 · ${department.departmentName}`} countNote={`${department.groupCount} 个小组 · 按事件发生时组织归属`} achievement={department} onOpen={(metric)=>openOrganizationDetails("department",department,metric)}/>)}
-      {organizationDialog?<Modal title={organizationDialog.kind==="sales"?`销售组织 · ${organizationDialog.metric==="gap"?"差距构成":"业绩构成"}`:`${(organizationDialog.achievement as DepartmentAchievement).departmentName} · ${organizationDialog.metric==="gap"?"部门差距构成":"部门业绩构成"}`} note={organizationDetails?(organizationDialog.metric==="gap"?`目标 ${formatMoney(organizationDetails.targetAmount!)} − 事件净额 ${formatMoney(organizationDetails.actualAmount)} = 差距 ${formatMoney(organizationDetails.gapAmount!)}`:`${organizationDetails.eventCount} 条事件 · 净额 ${formatMoney(organizationDetails.actualAmount)}`):"正在读取部门、小组、人员、订单和事件"} onClose={()=>{organizationDetailsRequest.current+=1;setOrganizationDialog(null);}}>{organizationDetailsError?<p className="form-error group-drilldown" role="alert">{organizationDetailsError}</p>:organizationDetails?<OrganizationHierarchy departments={organizationDepartments}/>:<p className="group-drilldown">正在读取组织业绩构成…</p>}</Modal>:null}
-      {data?.groupAchievements.map((group)=>{const groupReason=achievementReasonText(group.calculationReason,"小组");return <article className="team-achievement orders-card" aria-label={`${group.groupName}目标达成`} key={group.groupId}><PanelTitle title={`小组目标达成 · ${group.groupName}`} note={`${group.memberCount} 位事件责任人 · 按事件发生时组织归属`}/><section className="metric-band personal-metrics"><Metric label="小组生效目标" value={group.targetAmount===null?"—":formatMoney(group.targetAmount)} note={groupReason??"当前月小组目标"}/><Metric label="小组业绩" value={formatMoney(group.actualAmount)} note={`${group.eventCount} 条小组事件`} onClick={()=>openGroupDetails(group,"actual")} actionLabel={`查看${group.groupName}业绩构成`}/><Metric label="小组差距" value={group.gapAmount===null?"不计算":formatMoney(group.gapAmount)} note={groupReason??"生效目标减小组业绩"} onClick={group.gapAmount===null?undefined:()=>openGroupDetails(group,"gap")} actionLabel={group.gapAmount===null?undefined:`查看${group.groupName}差距构成`}/><Metric label="小组达成率" value={group.achievementRate===null?"不计算":`${group.achievementRate}%`} note={groupReason??"小组业绩 ÷ 生效目标"}/><Metric label="自然日进度" value={group.timeProgressRate===null?"不计算":`${group.timeProgressRate}%`} note={groupReason??"按上海自然日计算"}/><Metric label="进度偏差" value={group.progressVariance===null?"不计算":`${group.progressVariance}%`} note={groupReason??"达成率减自然日进度"} warning={group.progressVariance!==null&&Number(group.progressVariance)<0}/></section></article>})}
-      {selectedGroup?<Modal
-        title={`${selectedGroup.groupName} · ${groupDetailKind==="gap"?"小组差距构成":"小组业绩构成"}`}
-        note={groupDetails?(groupDetailKind==="gap"?`目标 ${formatMoney(groupDetails.targetAmount!)} − 事件净额 ${formatMoney(groupDetails.actualAmount)} = 差距 ${formatMoney(groupDetails.gapAmount!)}`:`${groupDetails.memberCount} 位事件责任人 · ${groupDetails.eventCount} 条事件 · 净额 ${formatMoney(groupDetails.actualAmount)}`):`正在读取 ${selectedGroup.groupName} 小组业绩事件`}
-        onClose={()=>{groupDetailsRequest.current+=1;setSelectedGroup(null);}}
-      >{groupDetailsError?<p className="form-error group-drilldown" role="alert">{groupDetailsError}</p>:groupDetails?<div className="group-drilldown"><AchievementMembers members={groupDetails.members}/></div>:<p className="group-drilldown">正在读取小组业绩构成…</p>}</Modal>:null}
-      <section className="metric-band"><Metric label="账本净额" value={data ? formatMoney(data.metrics.total) : "—"} note={`${data?.metrics.eventCount ?? 0} 条授权范围事件`}/><Metric label="正式报表" value="从生效目标进入" note="未生效不计算达成率"/><Metric label="待处理审批" value={String(data?.metrics.pendingApprovals ?? 0)} note="目标确认与变更" warning/><Metric label="负向调整" value={data ? formatMoney(data.metrics.negativeTotal) : "—"} note="暂停与金额变更" negative/></section>
-      <section className="dashboard-grid"><article className="trend-panel"><PanelTitle title="月度业绩趋势" note={`${data?.month.slice(0,4)??"当前"} 年授权范围业绩净额`}/><TrendChart year={data?.month.slice(0,4)??""} monthly={data?.monthly??[]}/></article><article className="ranking-panel"><PanelTitle title="小组业绩" note="按事件发生时组织归属"/>{data?.groups.map((group,i) => <div className="rank-row" key={group.id}><span>{i+1}</span><div><strong>{group.name}</strong><span><i style={{width:`${Math.max(0, Number(group.total)) / maxGroup * 100}%`}}/></span></div><b>{(Number(group.total)/10000).toFixed(1)}万</b></div>)}</article></section>
-      <section className="events-panel"><div className="panel-title"><div><h2>最近业绩事件</h2><p>入账后不可覆盖或删除</p></div><button onClick={onEnterOrders}>查看全部</button></div><table><thead><tr><th>订单编号</th><th>业务员</th><th>事件类型</th><th>记账月</th><th>金额</th><th>组织归属</th><th>状态</th></tr></thead><tbody>{data?.recent.map((event) => <tr key={`${event.orderNo}-${event.month}-${event.amount}`}><td>{event.orderNo}</td><td>{event.salespersonName}</td><td>{eventTypeName(event.eventType)}</td><td>{event.month}</td><td className={Number(event.amount)<0?"negative":""}>{formatMoney(event.amount)}</td><td>{event.groupName}</td><td>已入账</td></tr>)}</tbody></table></section>
-    </main>;
-}
-
-function AnalysisPage(){
-  const[month,setMonth]=useState(businessDateToday().slice(0,7));
-  const[data,setData]=useState<PerformanceAnalysis|null>(null);
-  const[error,setError]=useState("");
-  const[revision,setRevision]=useState(0);
-  const[selectedProvince,setSelectedProvince]=useState<AnalysisProvince|null>(null);
-  useEffect(()=>setSelectedProvince(null),[month]);
-  useEffect(()=>{const controller=new AbortController();setData(null);setError("");fetch(`/api/performance/analysis?month=${month}`,{signal:controller.signal}).then(async(response)=>{const result=await readResponseJson<PerformanceAnalysis&{message?:string}>(response,"分析服务响应无效");if(!response.ok)throw new Error(result.message??"分析加载失败");setData(result);}).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setError(failure instanceof Error?failure.message:"分析加载失败");});return()=>controller.abort();},[month,revision]);
-  return <main className="dashboard analysis-page"><section className="analysis-panel" aria-labelledby="analysis-title"><div className="analysis-header"><div><h1 id="analysis-title">地区与客户单位分析</h1><p>只按事件发生时的不可变分析维度快照汇总，不使用订单当前资料</p></div><label><span>分析月份</span><input type="month" value={month} onChange={(event)=>setMonth(event.target.value)}/></label></div>
-    {error?<div className="query-feedback"><p className="page-message" role="alert">{error}</p><button type="button" onClick={()=>setRevision((value)=>value+1)}>重试查询</button></div>:null}
-    {!data&&!error?<p className="analysis-loading" role="status">正在读取地区与客户单位分析…</p>:null}
-    {data?<><section className="metric-band analysis-metrics"><Metric label="授权范围总账" value={formatMoney(data.ledger.totalAmount)} note={`${data.ledger.eventCount} 条事件`}/><Metric label="已映射" value={formatMoney(data.mapped.totalAmount)} note={`${data.mapped.eventCount} 条可信维度事件`}/><Metric label="待补齐" value={formatMoney(data.pending.totalAmount)} note={`${data.pending.eventCount} 条缺少可信维度事件`} warning/></section><p className={`analysis-reconciliation ${data.reconciled?"":"analysis-reconciliation-error"}`}>{data.reconciled?"已映射金额 + 待补齐金额与授权范围总账完全对平。":"分析维度对账失败，请停止使用当前汇总。"}</p><div className="analysis-map-layout"><ChinaProvinceMap provinces={data.provinces} selectedRegionCode={selectedProvince?.regionCode??null} onSelect={setSelectedProvince}/><article className="analysis-card analysis-province-ranking"><h3>省份汇总</h3><div className="orders-table-wrap"><table aria-label="省份汇总"><thead><tr><th>省份</th><th>事件</th><th>金额</th></tr></thead><tbody>{data.provinces.length?data.provinces.map((item)=><tr key={item.regionCode}><td><button type="button" className="analysis-link" aria-pressed={selectedProvince?.regionCode===item.regionCode} onClick={()=>setSelectedProvince(item)}>{item.regionName}</button></td><td>{item.eventCount}</td><td className={Number(item.totalAmount)<0?"negative":""}>{formatMoney(item.totalAmount)}</td></tr>):<tr><td colSpan={3} className="empty-cell">本月没有已映射省份事件。</td></tr>}</tbody></table></div><div className="analysis-foreign"><div><strong>外贸（EXT-TRADE）</strong><span>独立区域，不进入省份统计</span></div><b className={Number(data.foreignTrade.totalAmount)<0?"negative":""}>{data.foreignTrade.eventCount} 条事件 · {formatMoney(data.foreignTrade.totalAmount)}</b></div></article></div><article className="analysis-card analysis-customer-summary"><h3>客户单位汇总</h3><div className="orders-table-wrap"><table aria-label="客户单位汇总"><thead><tr><th>区域</th><th>客户单位</th><th>事件</th><th>金额</th></tr></thead><tbody>{data.customers.length?data.customers.map((item)=><tr key={`${item.regionCode}:${item.customerUnit}`}><td>{item.regionName}</td><td>{item.customerUnit}</td><td>{item.eventCount}</td><td className={Number(item.totalAmount)<0?"negative":""}>{formatMoney(item.totalAmount)}</td></tr>):<tr><td colSpan={4} className="empty-cell">本月没有已映射客户单位事件。</td></tr>}</tbody></table></div></article>{selectedProvince?<AnalysisDrilldown key={`${month}:${selectedProvince.regionCode}`} province={selectedProvince} month={month}/>:null}</>:null}
-  </section></main>;
-}
-
-function ChinaProvinceMap({provinces,selectedRegionCode,onSelect}:{provinces:AnalysisProvince[];selectedRegionCode:string|null;onSelect:(province:AnalysisProvince)=>void}){
-  const[chinaMap,setChinaMap]=useState<ChinaMap|null>(null);
-  const[mapError,setMapError]=useState("");
-  useEffect(()=>{const controller=new AbortController();import("./china-map").then(({loadChinaMap})=>loadChinaMap(controller.signal)).then(setChinaMap).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setMapError(failure instanceof Error?failure.message:"中国省级地图加载失败");});return()=>controller.abort();},[]);
-  const byCode=new Map(provinces.map((province)=>[province.regionCode,province]));
-  const finiteTotals=provinces.map((province)=>Math.abs(Number(province.totalAmount))).filter(Number.isFinite);
-  const maximum=Math.max(1,...finiteTotals);
-  return <article className="analysis-card analysis-map-card"><h3>省份地图</h3><div className="analysis-map-wrap">{chinaMap?<svg className="analysis-map" viewBox={chinaMap.viewBox} role="group" aria-label="中国省份业绩地图">{chinaMap.locations.map((location)=>{
-    const regionCode=chinaMapRegionCodes[location.id];
-    const province=regionCode?byCode.get(regionCode):undefined;
-    if(!province)return <path key={location.id} d={location.path} className="analysis-map-empty" data-region-code={regionCode} aria-hidden="true"/>;
-    const selected=selectedRegionCode===province.regionCode;
-    const label=`地图选择${province.regionName}，${province.eventCount} 条事件，金额 ${formatMoney(province.totalAmount)}`;
-    return <path key={location.id} d={location.path} className={`analysis-map-region${Number(province.totalAmount)<0?" negative":""}${selected?" selected":""}`} style={{fillOpacity:.3+.7*Math.abs(Number(province.totalAmount))/maximum}} data-region-code={regionCode} role="button" tabIndex={0} aria-label={label} aria-pressed={selected} onClick={()=>onSelect(province)} onKeyDown={(event)=>{if(event.key==="Enter"||event.key===" "){event.preventDefault();onSelect(province);}}}/>;
-  })}</svg>:<p className={mapError?"page-message":"analysis-loading"} role={mapError?"alert":"status"}>{mapError||"正在读取中国省级地图…"}</p>}<p>颜色深浅只表示相对规模；金额以右侧省份汇总为准。</p><p><span>台湾省资料暂缺</span>；香港特别行政区、澳门特别行政区资料暂缺。</p><small>省级轮廓：<a href="https://www.tianditu.gov.cn/">天地图</a>来源；<a href="https://github.com/JayMuShui/chinese-global-compliant-geodata">chinese-global-compliant-geodata 1.0.0</a>（MIT）。</small></div></article>;
-}
-
-function AnalysisDrilldown({province,month}:{province:AnalysisProvince;month:string}){
-  const[customers,setCustomers]=useState<AnalysisCustomersDrilldown|null>(null);
-  const[customersError,setCustomersError]=useState("");
-  const[customersRevision,setCustomersRevision]=useState(0);
-  const[customerCursor,setCustomerCursor]=useState<string|null>(null);
-  const[customersLoadingMore,setCustomersLoadingMore]=useState(false);
-  const[selectedCustomer,setSelectedCustomer]=useState<AnalysisCustomer|null>(null);
-  const[months,setMonths]=useState<AnalysisMonthsDrilldown|null>(null);
-  const[monthsError,setMonthsError]=useState("");
-  const[monthsRevision,setMonthsRevision]=useState(0);
-  const[selectedMonth,setSelectedMonth]=useState<{month:string;eventCount:number;totalAmount:string}|null>(null);
-  const[events,setEvents]=useState<AnalysisEventsDrilldown|null>(null);
-  const[eventsError,setEventsError]=useState("");
-  const[eventsRevision,setEventsRevision]=useState(0);
-  const[eventCursor,setEventCursor]=useState<string|null>(null);
-  const[eventsLoadingMore,setEventsLoadingMore]=useState(false);
-
-  useEffect(()=>{
-    const append=customerCursor!==null;const controller=new AbortController();if(!append){setCustomers(null);setSelectedCustomer(null);setMonths(null);setSelectedMonth(null);setEvents(null);}else setCustomersLoadingMore(true);setCustomersError("");
-    const params=new URLSearchParams({level:"customers",regionCode:province.regionCode,month});if(customerCursor)params.set("cursor",customerCursor);
-    apiFetch(`/api/performance/analysis/drilldown?${params}`,{signal:controller.signal}).then(async(response)=>{const result=await readResponseJson<AnalysisCustomersDrilldown&{message?:string}>(response,"省份客户响应无效");if(!response.ok)throw new Error(result.message??"省份客户加载失败");setCustomers((current)=>append&&current?{...result,customers:[...current.customers,...result.customers]}:result);}).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setCustomersError(failure instanceof Error?failure.message:"省份客户加载失败");}).finally(()=>{if(!controller.signal.aborted)setCustomersLoadingMore(false);});
-    return()=>controller.abort();
-  },[province.regionCode,month,customerCursor,customersRevision]);
-
-  useEffect(()=>{
-    setMonths(null);setMonthsError("");setSelectedMonth(null);setEvents(null);setEventCursor(null);if(!selectedCustomer)return;
-    const controller=new AbortController();const params=new URLSearchParams({level:"months",regionCode:province.regionCode,customerUnit:selectedCustomer.customerUnit,year:month.slice(0,4)});
-    apiFetch(`/api/performance/analysis/drilldown?${params}`,{signal:controller.signal}).then(async(response)=>{const result=await readResponseJson<AnalysisMonthsDrilldown&{message?:string}>(response,"客户月份响应无效");if(!response.ok)throw new Error(result.message??"客户月份加载失败");setMonths(result);}).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setMonthsError(failure instanceof Error?failure.message:"客户月份加载失败");});
-    return()=>controller.abort();
-  },[province.regionCode,selectedCustomer,month,monthsRevision]);
-
-  useEffect(()=>{
-    const append=eventCursor!==null;if(!append)setEvents(null);setEventsError("");if(!selectedCustomer||!selectedMonth)return;
-    const controller=new AbortController();if(append)setEventsLoadingMore(true);const params=new URLSearchParams({level:"events",regionCode:province.regionCode,customerUnit:selectedCustomer.customerUnit,month:selectedMonth.month});if(eventCursor)params.set("cursor",eventCursor);
-    apiFetch(`/api/performance/analysis/drilldown?${params}`,{signal:controller.signal}).then(async(response)=>{const result=await readResponseJson<AnalysisEventsDrilldown&{message?:string}>(response,"订单事件响应无效");if(!response.ok)throw new Error(result.message??"订单事件加载失败");setEvents((current)=>{if(!append||!current)return result;const orders=new Map(current.orders.map((order)=>[order.orderId,{...order,events:[...order.events]}]));for(const order of result.orders){const existing=orders.get(order.orderId);if(existing)existing.events.push(...order.events);else orders.set(order.orderId,order);}return{...result,orders:[...orders.values()]};});}).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setEventsError(failure instanceof Error?failure.message:"订单事件加载失败");}).finally(()=>{if(!controller.signal.aborted)setEventsLoadingMore(false);});
-    return()=>controller.abort();
-  },[province.regionCode,selectedCustomer,selectedMonth,eventCursor,eventsRevision]);
-
-  const customerMatched=customers&&formatMoney(customers.totalAmount)===formatMoney(province.totalAmount)&&customers.eventCount===province.eventCount;
-  const monthRows=new Map(months?.months.map((item)=>[item.month,item])??[]);
-  const filledMonths=Array.from({length:12},(_,index)=>{const value=`${month.slice(0,4)}-${String(index+1).padStart(2,"0")}`;return monthRows.get(value)??{month:value,eventCount:0,totalAmount:"0.00"};});
-  const parentMonth=months&&selectedCustomer?monthRows.get(month):null;
-  const monthMatched=parentMonth&&selectedCustomer&&formatMoney(parentMonth.totalAmount)===formatMoney(selectedCustomer.totalAmount)&&parentMonth.eventCount===selectedCustomer.eventCount;
-  const eventsMatched=events&&selectedMonth&&formatMoney(events.totalAmount)===formatMoney(selectedMonth.totalAmount)&&events.eventCount===selectedMonth.eventCount;
-  const allEvents=events?.orders.flatMap((order)=>order.events)??[];
-  return <section className="analysis-drilldown" role="region" aria-label="分析穿透"><header><h3>{province.regionName}客户单位</h3><p>{province.eventCount} 条事件 · 省份汇总 <strong>{formatMoney(province.totalAmount)}</strong></p></header>
-    {customersError?<div className="query-feedback"><p className="page-message" role="alert">{customersError}</p><button type="button" disabled={customersLoadingMore} onClick={()=>setCustomersRevision((value)=>value+1)}>重试省份客户</button></div>:null}
-    {!customers&&!customersError?<p className="analysis-loading" role="status">正在读取{province.regionName}客户单位…</p>:null}
-    {customers?<><p className={`analysis-reconciliation ${customerMatched?"":"analysis-reconciliation-error"}`}>{customerMatched?`服务端客户事件数与金额 ${formatMoney(customers.totalAmount)} 均与省份汇总完全对平。`:"客户合计与省份汇总不一致，请停止使用当前穿透结果。"}</p><div className="orders-table-wrap"><table aria-label={`${province.regionName}客户单位`}><thead><tr><th>客户单位</th><th>事件</th><th>金额</th><th>穿透</th></tr></thead><tbody>{customers.customers.length?customers.customers.map((customer)=><tr key={customer.customerUnit}><td>{customer.customerUnit}</td><td>{customer.eventCount}</td><td className={Number(customer.totalAmount)<0?"negative":""}>{formatMoney(customer.totalAmount)}</td><td><button type="button" className="table-action" aria-pressed={selectedCustomer?.customerUnit===customer.customerUnit} aria-label={`查看${customer.customerUnit}月份趋势`} onClick={()=>setSelectedCustomer(customer)}>查看月份</button></td></tr>):<tr><td colSpan={4} className="empty-cell">该省份本月没有客户单位事件。</td></tr>}</tbody></table></div><nav className="table-pagination" aria-label="客户单位分页"><span>已加载 {customers.customers.length} / {customers.customerCount} 个客户单位</span>{customers.nextCursor?<button type="button" disabled={customersLoadingMore} onClick={()=>setCustomerCursor(customers.nextCursor)}>{customersLoadingMore?"正在加载…":"加载更多客户单位"}</button>:null}</nav></>:null}
-    {selectedCustomer?<section className="analysis-drilldown-stage"><header><h3>{selectedCustomer.customerUnit}月度趋势</h3><p>{month.slice(0,4)} 年 · 客户年度净额 {months?formatMoney(months.totalAmount):"读取中"}</p></header>{monthsError?<div className="query-feedback"><p className="page-message" role="alert">{monthsError}</p><button type="button" onClick={()=>setMonthsRevision((value)=>value+1)}>重试客户月份</button></div>:null}{!months&&!monthsError?<p className="analysis-loading" role="status">正在读取{selectedCustomer.customerUnit}月度趋势…</p>:null}{months?<><p className={`analysis-reconciliation ${monthMatched?"":"analysis-reconciliation-error"}`}>{monthMatched?`${month.replace("-","年")}月金额与上级客户行完全对平。`:`${month.replace("-","年")}月金额与上级客户行不一致，请停止使用当前穿透结果。`}</p><div className="analysis-months" role="list" aria-label={`${selectedCustomer.customerUnit}${months.year}年月度趋势`}>{filledMonths.map((item)=><div key={item.month} role="listitem"><button type="button" className={selectedMonth?.month===item.month?"selected":""} aria-pressed={selectedMonth?.month===item.month} aria-label={`查看${Number(item.month.slice(0,4))}年${Number(item.month.slice(5))}月订单事件，${item.eventCount} 条事件，金额 ${formatMoney(item.totalAmount)}`} onClick={()=>{setEventCursor(null);setSelectedMonth(item);}}><span>{Number(item.month.slice(5))}月</span><strong className={Number(item.totalAmount)<0?"negative":""}>{formatMoney(item.totalAmount)}</strong><small>{item.eventCount} 条事件</small></button></div>)}</div></>:null}</section>:null}
-    {selectedMonth&&selectedCustomer?<section className="analysis-drilldown-stage"><header><h3>{Number(selectedMonth.month.slice(0,4))}年{Number(selectedMonth.month.slice(5))}月订单与事件</h3><p>{selectedCustomer.customerUnit} · 上级月份净额 {formatMoney(selectedMonth.totalAmount)}</p></header>{eventsError?<div className="query-feedback"><p className="page-message" role="alert">{eventsError}</p><button type="button" disabled={eventsLoadingMore} onClick={()=>setEventsRevision((value)=>value+1)}>重试订单事件</button></div>:null}{!events&&!eventsError?<p className="analysis-loading" role="status">正在读取订单与不可变事件…</p>:null}{events?<><p className={`analysis-reconciliation ${eventsMatched?"":"analysis-reconciliation-error"}`}>{eventsMatched?`服务端全部订单事件合计 ${formatMoney(events.totalAmount)} 与上级月份完全对平。`:"订单事件合计与上级月份不一致，请停止使用当前穿透结果。"}</p><p className="analysis-loading">已加载 {allEvents.length} / {events.eventCount} 条事件</p><div className="orders-table-wrap"><table aria-label="订单对账"><thead><tr><th>订单</th><th>客户</th><th>事件</th><th>净额</th></tr></thead><tbody>{events.orders.length?events.orders.map((order)=><tr key={order.orderId}><td>{order.orderNo}</td><td>{order.customerName}</td><td>{order.eventCount}</td><td className={Number(order.totalAmount)<0?"negative":""}>{formatMoney(order.totalAmount)}</td></tr>):<tr><td colSpan={4} className="empty-cell">该月份没有订单事件。</td></tr>}</tbody></table></div>{allEvents.length?<div className="orders-table-wrap"><table aria-label="不可变事件明细"><thead><tr><th>序号</th><th>事件</th><th>金额</th><th>业务日 / 记账月</th><th>发生时分析维度</th><th>责任归属</th><th>状态 / 原因</th></tr></thead><tbody>{allEvents.map((event)=><tr key={event.id}><td>第 {event.sequence} 条</td><td>{eventTypeName(event.eventType)}</td><td className={Number(event.deltaAmount)<0?"negative":""}>{formatMoney(event.deltaAmount)}</td><td>{event.occurredOn} / {event.accountingMonth}</td><td>{province.regionName} / {event.customerUnit}<small className="table-secondary">{event.businessRegionSourceText}</small></td><td>{event.salespersonName}<small className="table-secondary">{[event.departmentName,event.groupName].filter(Boolean).join(" / ")||"—"}</small></td><td>{event.resultingLifecycleState?<Status state={event.resultingLifecycleState}/>:"原始状态未推断"}<small className="table-secondary">{event.reason??"—"}</small></td></tr>)}</tbody></table></div>:null}{events.nextCursor?<nav className="table-pagination" aria-label="事件分页"><button type="button" disabled={eventsLoadingMore} onClick={()=>setEventCursor(events.nextCursor)}>{eventsLoadingMore?"正在加载…":"加载更多事件"}</button></nav>:null}</>:null}</section>:null}
-  </section>;
-}
-
-function OrdersPage({ user }: { user: User }) {
-  const canEdit=user.capabilities.editPerformance;
-  const canExport=user.capabilities.exportPerformance;
-  const [orders, setOrders] = useState<Order[]>([]);
-  const [selected, setSelected] = useState<Order | null>(null);
-  const [showCreate, setShowCreate] = useState(false);
-  const [showImport, setShowImport] = useState(false);
-  const initialUrlState=useRef(readOrderUrlState()).current;
-  const [draftFilters,setDraftFilters]=useState<OrderFilters>(initialUrlState.filters);
-  const [committedFilters,setCommittedFilters]=useState<OrderFilters>(initialUrlState.filters);
-  const [isComposing, setIsComposing] = useState(false);
-  const [loadState,setLoadState]=useState<"loading"|"ready"|"error"|"forbidden">("loading");
-  const [loadError,setLoadError]=useState("");
-  const [refreshVersion, setRefreshVersion] = useState(0);
-  const [pageCursor, setPageCursor] = useState<string | null>(initialUrlState.cursor);
-  const [previousCursor, setPreviousCursor] = useState<string | null>(null);
-  const [nextCursor, setNextCursor] = useState<string | null>(null);
-  const searchRef = useRef<HTMLInputElement>(null);
-  const commitFilters=useCallback((value:OrderFilters,historyMode:"push"|"replace"="push",updateDraft=true)=>{
-    const normalized:OrderFilters={
-      search:value.search.trim(),month:value.month.trim(),status:value.status.trim(),salesperson:value.salesperson.trim(),
-      department:value.department.trim(),group:value.group.trim(),region:value.region.trim(),customerUnit:value.customerUnit.trim(),
-    };
-    writeOrderUrlState(normalized,null,historyMode);
-    if(updateDraft)setDraftFilters(normalized);
-    setPageCursor(null);
-    setCommittedFilters(normalized);
-  },[]);
-  useEffect(()=>{
-    const restore=()=>{const restored=readOrderUrlState();setDraftFilters(restored.filters);setCommittedFilters(restored.filters);setPageCursor(restored.cursor);};
-    window.addEventListener("popstate",restore);return()=>window.removeEventListener("popstate",restore);
-  },[]);
-  useEffect(()=>{
-    if(isComposing||draftFilters.search.trim()===committedFilters.search)return;
-    const timer=window.setTimeout(()=>commitFilters({...committedFilters,search:draftFilters.search},"push",false),300);
-    return()=>window.clearTimeout(timer);
-  },[draftFilters.search,isComposing,committedFilters,commitFilters]);
-  const filterRequestKey=JSON.stringify(committedFilters);
-  const lastFilterRequestKey=useRef<string|null>(null);
-  useEffect(() => {
-    if(lastFilterRequestKey.current!==filterRequestKey){lastFilterRequestKey.current=filterRequestKey;setOrders([]);setPreviousCursor(null);setNextCursor(null);}
-    const controller=new AbortController();setLoadState("loading");setLoadError("");
-    const params=new URLSearchParams();for(const [key,value] of Object.entries(committedFilters))if(value)params.set(key,value);if(pageCursor)params.set("cursor",pageCursor);
-    fetch(`/api/performance/orders?${params.toString()}`,{signal:controller.signal}).then(async(response)=>{
-      const data=await readResponseJson<{orders?:Order[];previousCursor?:string|null;nextCursor?:string|null;message?:string}>(response,"订单服务响应无效，请重试");
-      if(response.status===403){setOrders([]);setPreviousCursor(null);setNextCursor(null);setLoadState("forbidden");return;}
-      if(!response.ok)throw new Error(data.message??"订单加载失败");
-      setOrders(data.orders??[]);
-      setPreviousCursor(data.previousCursor??null);setNextCursor(data.nextCursor??null);
-      setLoadState("ready");
-    }).catch((error)=>{if(error instanceof DOMException&&error.name==="AbortError")return;setLoadError(error instanceof Error?error.message:"订单加载失败");setLoadState("error");});
-    return()=>controller.abort();
-  }, [filterRequestKey,pageCursor,refreshVersion]);
-  const loading=loadState==="loading";
-  const hasFilters=Object.values(committedFilters).some(Boolean);
-  async function refresh() { writeOrderUrlState(committedFilters,null,"replace");setPageCursor(null);setRefreshVersion((value)=>value+1); }
-  function movePage(cursor:string|null){if(!cursor)return;if(cursor===pageCursor){setRefreshVersion((value)=>value+1);return;}writeOrderUrlState(committedFilters,cursor);setPageCursor(cursor);}
-  function clearSearch(){setDraftFilters((current)=>({...current,search:""}));commitFilters({...committedFilters,search:""},"push",false);window.requestAnimationFrame(()=>searchRef.current?.focus());}
-  function updateFilter(key:keyof OrderFilters,value:string){setDraftFilters((current)=>({...current,[key]:value}));}
-  function exportOrders(){const params=new URLSearchParams();for(const [key,value] of Object.entries(committedFilters))if(value)params.set(key,value);const query=params.toString();window.location.href=`/api/exports/performance.csv${query?`?${query}`:""}`;}
-  const emptyMessage=loadState==="forbidden"?"无可显示订单。":loadState==="error"?"订单加载失败，可重试。":hasFilters?"没有符合当前组合条件的订单。":"暂无订单数据。";
-  return <main className="dashboard orders-page"><header><div><h1>订单业绩</h1><p>按订单编号维护不可变业绩事件；已入账记录不能覆盖或删除</p></div>{canExport||canEdit ? <div className="header-actions">{canExport?<button className="secondary-action" onClick={exportOrders}>导出全部匹配订单</button>:null}{canEdit?<><button className="secondary-action" onClick={() => setShowImport(true)}><FileUp size={16}/>Excel 导入</button><button className="primary-action" onClick={() => setShowCreate(true)}><Plus size={16}/>录入新订单</button></>:null}</div> : null}</header>
-    {loadState==="error"?<div className="query-feedback"><p className="page-message" role="alert">{loadError}</p><button type="button" onClick={()=>setRefreshVersion((value)=>value+1)}>重试查询</button></div>:null}
-    {loadState==="forbidden"?<div className="permission-note"><ShieldCheck size={18}/>当前账号没有订单查看权限。</div>:null}
-    {!canEdit ? <div className="permission-note"><ShieldCheck size={18}/>当前角色仅可查看。只有销售助理及销售助理组长可以录入或调整业绩。</div> : null}
-    <LedgerGovernancePanel user={user} orders={orders} onChanged={refresh}/>
-    <section className="orders-card" aria-labelledby="orders-table-title"><div className="orders-toolbar"><div><h2 id="orders-table-title">订单台账</h2><span role="status">{loading?"正在查询…":`本页 ${orders.length} 笔订单`}</span></div><button className="icon-action" onClick={() => refresh()} aria-label="刷新订单"><RefreshCw size={17}/></button></div>
-      <form className="order-search" role="search" noValidate onSubmit={(event)=>{event.preventDefault();if(!isComposing)commitFilters({...committedFilters,search:draftFilters.search},"push",false);}}><label htmlFor="order-search-input">定位订单</label><div><Search size={17} aria-hidden="true"/><input ref={searchRef} id="order-search-input" type="search" value={draftFilters.search} maxLength={100} placeholder="输入订单编号、客户名称、客户单位或业务员" onChange={(event)=>updateFilter("search",event.target.value)} onCompositionStart={()=>setIsComposing(true)} onCompositionEnd={(event)=>{setIsComposing(false);updateFilter("search",event.currentTarget.value);}}/>{draftFilters.search?<button type="button" onClick={clearSearch} aria-label="清除订单搜索"><X size={16}/></button>:null}</div><button type="submit">搜索</button></form>
-      <form className="order-filters" noValidate onSubmit={(event)=>{event.preventDefault();commitFilters(draftFilters);}}><div className="order-filter-grid">
-        <label className="field"><span>订单月份</span><input type="month" value={draftFilters.month} onChange={(event)=>updateFilter("month",event.target.value)}/></label>
-        <label className="field"><span>订单状态</span><select value={draftFilters.status} onChange={(event)=>updateFilter("status",event.target.value)}><option value="">全部状态</option><option value="active">正向计入</option><option value="paused">已暂停</option><option value="zero">零金额</option><option value="historical_review_required">待历史核对</option><option value="draft">草稿</option></select></label>
-        <label className="field"><span>业务员筛选</span><input value={draftFilters.salesperson} maxLength={300} onChange={(event)=>updateFilter("salesperson",event.target.value)}/></label>
-        <label className="field"><span>部门筛选</span><input value={draftFilters.department} maxLength={300} onChange={(event)=>updateFilter("department",event.target.value)}/></label>
-        <label className="field"><span>小组筛选</span><input value={draftFilters.group} maxLength={300} onChange={(event)=>updateFilter("group",event.target.value)}/></label>
-        <label className="field"><span>标准业务区域筛选</span><select value={draftFilters.region} onChange={(event)=>updateFilter("region",event.target.value)}><option value="">全部区域</option>{standardBusinessRegions.map(([code,name])=><option key={code} value={code}>{name}</option>)}</select></label>
-        <label className="field"><span>客户单位筛选</span><input value={draftFilters.customerUnit} maxLength={300} onChange={(event)=>updateFilter("customerUnit",event.target.value)}/></label>
-      </div><div className="order-filter-actions"><button type="button" onClick={()=>commitFilters({...emptyOrderFilters})}>清除筛选</button><button type="submit">应用筛选</button></div></form>
-      <div className="orders-table-wrap"><table><thead><tr><th scope="col">订单编号</th><th scope="col">客户</th><th scope="col">业务员</th><th scope="col">当前营业额</th><th scope="col">计入业绩</th><th scope="col">状态</th><th scope="col">操作</th></tr></thead><tbody>{!loading&&orders.length === 0 ? <tr><td colSpan={7} className="empty-cell">{emptyMessage}</td></tr> : orders.map((order) => <tr key={order.id}><td>{order.orderNo}</td><td>{order.customerName}</td><td>{order.salespersonName}</td><td>{formatMoney(order.currentRevenue)}</td><td>{formatMoney(order.countedAmount)}</td><td><Status state={order.lifecycleState}/></td><td><button className="table-action" onClick={() => setSelected(order)}>查看 / 调整</button></td></tr>)}</tbody></table></div>
-      <nav className="table-pagination" aria-label="订单分页"><button type="button" disabled={loading||!previousCursor} onClick={()=>movePage(previousCursor)}>上一页</button><button type="button" disabled={loading||!nextCursor} onClick={()=>movePage(nextCursor)}>下一页</button></nav>
-    </section>
-    {showCreate ? <CreateOrder onClose={() => setShowCreate(false)} onSaved={async () => { setShowCreate(false); await refresh(); }} /> : null}
-    {showImport ? <ExcelImportDialog user={user} onClose={() => setShowImport(false)} onImported={async () => { setShowImport(false); await refresh(); }} /> : null}
-    {selected ? <AdjustOrder order={selected} canEdit={canEdit} onClose={() => setSelected(null)} onSaved={async () => { setSelected(null); await refresh(); }} /> : null}
-  </main>;
-}
-
-type ImportConfig = { id:string; name:string; version:number; status:"draft"|"approved"|"retired"; sheetName:string; fixedEventType:"legacy_adjustment"|null };
-type ImportIssue = { rowNumber:number; code:string; severity:"blocking"|"warning"|"info"; message:string };
-type ImportReconciliationSummary = { rows:number; orders:number; events:number; totalAmount:number; monthly:Array<{month:string;events:number;totalAmount:number}> };
-type BackfillAmount = { events:number; totalAmount:number };
-type ImportReport = { batchId:string; status:"blocked"|"preflight_ready"; sourceSha256:string; issues:ImportIssue[]; summary:{rows:number;orders?:number;events?:number;readyMapped?:number;pending?:number;alreadyMapped?:number;totalAmount:number;blocking:number;warnings?:number}; reconciliation:{actual?:ImportReconciliationSummary;expected?:ImportReconciliationSummary|null;source?:BackfillAmount;readyMapped?:BackfillAmount;pending?:BackfillAmount;alreadyMapped?:BackfillAmount;blocked?:BackfillAmount;matched:boolean|null}; mappings?:Array<{rowNumber:number;eventId:string;sourceKey:string;status:"ready"|"already_mapped";businessRegionCode:string;businessRegionSourceText:string;customerUnit:string}> };
-
-function fileAsBase64(file:File):Promise<string>{return new Promise((resolve,reject)=>{const reader=new FileReader();reader.onerror=()=>reject(new Error("文件读取失败，请重新选择。"));reader.onload=()=>{const result=String(reader.result??"");const comma=result.indexOf(",");if(comma<0){reject(new Error("文件内容无效。"));return;}resolve(result.slice(comma+1));};reader.readAsDataURL(file);});}
-
-function ExcelImportDialog({user,onClose,onImported}:{user:User;onClose:()=>void;onImported:()=>Promise<void>}){
-  const[configs,setConfigs]=useState<ImportConfig[]>([]);const[configId,setConfigId]=useState("");const[file,setFile]=useState<File|null>(null);const[report,setReport]=useState<ImportReport|null>(null);const[mode,setMode]=useState<"ledger"|"dimension_backfill">("ledger");const[confirmedWarnings,setConfirmedWarnings]=useState<Set<string>>(new Set());const[error,setError]=useState("");const[loading,setLoading]=useState(true);const[busy,setBusy]=useState(false);const isLeader=user.roles.includes("sales_assistant_leader");
-  useEffect(()=>{const controller=new AbortController();fetch("/api/imports/configs",{signal:controller.signal}).then(async(response)=>{const data=await response.json() as {configs?:ImportConfig[];message?:string};if(!response.ok)throw new Error(data.message??"导入配置加载失败");const approved=(data.configs??[]).filter((item)=>item.status==="approved");setConfigs(approved);setConfigId(approved[0]?.id??"");}).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setError(failure instanceof Error?failure.message:"导入配置加载失败");}).finally(()=>{if(!controller.signal.aborted)setLoading(false);});return()=>controller.abort();},[]);
-  const availableConfigs=mode==="dimension_backfill"?configs.filter((item)=>item.fixedEventType==="legacy_adjustment"):configs;
-  function changeMode(next:"ledger"|"dimension_backfill"){setMode(next);setReport(null);setConfirmedWarnings(new Set());setError("");const candidates=next==="dimension_backfill"?configs.filter((item)=>item.fixedEventType==="legacy_adjustment"):configs;setConfigId(candidates[0]?.id??"");}
-  function chooseFile(selected:File|null){setReport(null);setConfirmedWarnings(new Set());setError("");if(!selected){setFile(null);return;}if(!selected.name.toLowerCase().endsWith(".xlsx")){setFile(null);setError("只接受 .xlsx 工作簿。");return;}if(selected.size===0||selected.size>20*1024*1024){setFile(null);setError("文件必须大于 0 且不超过 20 MB。");return;}setFile(selected);}
-  async function preflight(event:FormEvent){event.preventDefault();if(busy)return;if(!configId||!file){setError("请选择已批准的导入配置和一个 .xlsx 文件。");return;}setBusy(true);setError("");setReport(null);try{const contentBase64=await fileAsBase64(file);const endpoint=mode==="dimension_backfill"?"/api/imports/dimension-backfills/preflight":"/api/imports/preflight";const response=await apiFetch(endpoint,{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({configId:Number(configId),fileName:file.name,contentBase64})});const data=await response.json() as ImportReport&{message?:string};if(!response.ok)throw new Error(data.message??"预检失败");setReport(data);}catch(failure){setError(failure instanceof Error?failure.message:"预检失败，请重试。");}finally{setBusy(false);}}
-  async function confirm(){if(!report||busy)return;setBusy(true);setError("");try{const endpoint=mode==="dimension_backfill"?`/api/imports/dimension-backfills/${report.batchId}/confirm`:`/api/imports/batches/${report.batchId}/confirm`;const response=await apiFetch(endpoint,{method:"POST",headers:{"content-type":"application/json"},body:mode==="dimension_backfill"?"{}":JSON.stringify({confirmedWarnings:[...confirmedWarnings]})});const data=await response.json() as {message?:string};if(!response.ok)throw new Error(data.message??(mode==="dimension_backfill"?"确认补齐失败":"确认入账失败"));await onImported();}catch(failure){setError(failure instanceof Error?failure.message:"确认失败，请重试。");}finally{setBusy(false);}}
-  const warningKeys=report?.issues.filter((issue)=>issue.severity==="warning").map((issue)=>`${issue.rowNumber}:${issue.code}`)??[];const allWarningsConfirmed=warningKeys.every((key)=>confirmedWarnings.has(key));
-  return <Modal title="Excel 批量导入" note="先完整预检，只有销售助理组长可以确认无阻断批次" onClose={onClose}><form noValidate className="business-form import-form" onSubmit={preflight}>
-    <fieldset className="import-mode"><legend>操作类型</legend><label><input type="radio" name="import-mode" checked={mode==="ledger"} disabled={busy} onChange={()=>changeMode("ledger")}/>业绩入账</label><label><input type="radio" name="import-mode" checked={mode==="dimension_backfill"} disabled={busy} onChange={()=>changeMode("dimension_backfill")}/>历史分析维度补齐</label></fieldset>
-    <div className="import-guidance">{mode==="ledger"?<a href="/SampleFlow标准业绩导入模板.xlsx" download>下载标准业绩模板</a>:null}<span>{mode==="dimension_backfill"?"上传原始受控工作簿；只按文件哈希、工作表和行号匹配旧事件，不会修改订单或事件。":"仅接收固定值 .xlsx；公式、宏、外部链接和未知表头会被阻断。"}</span></div>
-    <label className="field"><span>已批准导入配置</span><select value={configId} disabled={loading||busy} onChange={(event)=>{setConfigId(event.target.value);setReport(null);}}><option value="">{loading?"正在加载…":"请选择"}</option>{availableConfigs.map((item)=><option key={item.id} value={item.id}>{item.name} · v{item.version}</option>)}</select></label>
-    <label className="import-file"><FileUp size={22}/><span>{file?file.name:"选择一个 .xlsx 文件"}</span><small>{file?`${(file.size/1024).toFixed(1)} KB`:"单文件上限 20 MB；不会显示或记录本地路径"}</small><input type="file" accept=".xlsx,application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" disabled={busy} onChange={(event)=>chooseFile(event.currentTarget.files?.[0]??null)}/></label>
-    {!loading&&availableConfigs.length===0?<p className="permission-note">{mode==="dimension_backfill"?"暂无已批准的专用历史配置。":"暂无已批准配置。销售助理组长需先创建草稿，并由人事批准业务区域及人员精确映射。"}</p>:null}{error?<p className="form-error" role="alert">{error}</p>:null}
-    {report?<section className={`import-report import-${report.status}`} aria-labelledby="import-report-title"><h3 id="import-report-title">{report.status==="blocked"?"预检未通过":"预检通过，等待确认"}</h3><dl><div><dt>来源 SHA-256</dt><dd>{report.sourceSha256}</dd></div>{mode==="dimension_backfill"?<><div><dt>来源行 / 可补齐映射 / 待补齐</dt><dd>{report.summary.rows} / {report.summary.readyMapped} / {report.summary.pending}</dd></div><div><dt>已补齐 / 阻断</dt><dd>{report.summary.alreadyMapped} / {report.summary.blocking}</dd></div></>:<><div><dt>来源行 / 订单 / 待入账事件</dt><dd>{report.summary.rows} / {report.summary.orders} / {report.summary.events}</dd></div><div><dt>阻断 / 警告</dt><dd>{report.summary.blocking} / {report.summary.warnings}</dd></div></>}<div><dt>金额合计</dt><dd>{formatMoney(report.summary.totalAmount)}</dd></div></dl>
-      {mode==="dimension_backfill"?<div className="import-reconciliation"><h4>来源对账</h4><p>受控来源 {report.reconciliation.source?.events??0} 笔 / {formatMoney(report.reconciliation.source?.totalAmount??0)}；可补齐映射 {report.reconciliation.readyMapped?.events??0} 笔 / {formatMoney(report.reconciliation.readyMapped?.totalAmount??0)}。</p><p>待补齐 {report.reconciliation.pending?.events??0} 笔 / {formatMoney(report.reconciliation.pending?.totalAmount??0)}；已补齐 {report.reconciliation.alreadyMapped?.events??0} 笔 / {formatMoney(report.reconciliation.alreadyMapped?.totalAmount??0)}；冲突 {report.reconciliation.blocked?.events??0} 笔 / {formatMoney(report.reconciliation.blocked?.totalAmount??0)}。</p><p>{report.reconciliation.matched?"四类数量和金额与受控来源总账对平。":"分类对账失败，批次已阻断。"}</p>{report.mappings?.length?<div className="orders-table-wrap"><table><thead><tr><th>来源行</th><th>状态</th><th>标准区域</th><th>区域原文</th><th>客户单位</th></tr></thead><tbody>{report.mappings.map((mapping)=><tr key={mapping.sourceKey}><td>{mapping.rowNumber}</td><td>{mapping.status==="ready"?"可补齐":"已补齐"}</td><td>{mapping.businessRegionCode}</td><td>{mapping.businessRegionSourceText}</td><td>{mapping.customerUnit}</td></tr>)}</tbody></table></div>:null}</div>:<div className="import-reconciliation"><h4>逐月对账</h4><div className="orders-table-wrap"><table><thead><tr><th>月份</th><th>实际事件</th><th>实际金额</th><th>预期事件</th><th>预期金额</th></tr></thead><tbody>{report.reconciliation.actual?.monthly.map((month)=>{const expected=report.reconciliation.expected?.monthly.find((item)=>item.month===month.month);return <tr key={month.month}><td>{month.month}</td><td>{month.events}</td><td>{formatMoney(month.totalAmount)}</td><td>{expected?.events??"—"}</td><td>{expected?formatMoney(expected.totalAmount):"—"}</td></tr>;})}</tbody></table></div><p>{report.reconciliation.matched===true?"整体与逐月对账一致。":report.reconciliation.matched===false?"整体或逐月对账不一致，批次已阻断。":"当前配置未设置预期基准，仅展示实际逐月汇总。"}</p></div>}
-      {report.issues.length?<ul>{report.issues.map((issue,index)=>{const warningKey=`${issue.rowNumber}:${issue.code}`;const issueLabel=issue.severity==="blocking"?"阻断":issue.severity==="warning"?"警告":"提示";return <li key={`${warningKey}-${index}`}><strong>{issue.rowNumber===0?"批次级":`第 ${issue.rowNumber} 行`} · {issueLabel}</strong><span>{issue.message}</span>{issue.severity==="warning"?<label><input type="checkbox" checked={confirmedWarnings.has(warningKey)} onChange={(event)=>setConfirmedWarnings((current)=>{const next=new Set(current);if(event.target.checked)next.add(warningKey);else next.delete(warningKey);return next;})}/>我已核对并确认此条警告</label>:null}</li>;})}</ul>:<p>未发现阻断、警告或提示。</p>}</section>:null}
-    <div className="modal-actions"><button type="button" onClick={onClose}>取消</button>{!report?<button type="submit" disabled={busy||loading||!file||!configId} aria-busy={busy}>{busy?"正在预检…":"运行只读预检"}</button>:report.status==="preflight_ready"&&isLeader?<button type="button" className="import-confirm" disabled={busy||!allWarningsConfirmed} aria-busy={busy} onClick={confirm}>{busy?"正在确认…":mode==="dimension_backfill"?"确认补齐分析维度":"确认整批入账"}</button>:report.status==="preflight_ready"?<span className="import-handoff">请交由销售助理组长确认</span>:<button type="button" onClick={()=>setReport(null)}>重新选择</button>}</div>
-  </form></Modal>;
-}
-
-const standardBusinessRegions = [
-  ["CN-BJ","北京市"],["CN-TJ","天津市"],["CN-HE","河北省"],["CN-SX","山西省"],["CN-NM","内蒙古自治区"],
-  ["CN-LN","辽宁省"],["CN-JL","吉林省"],["CN-HL","黑龙江省"],["CN-SH","上海市"],["CN-JS","江苏省"],
-  ["CN-ZJ","浙江省"],["CN-AH","安徽省"],["CN-FJ","福建省"],["CN-JX","江西省"],["CN-SD","山东省"],
-  ["CN-HA","河南省"],["CN-HB","湖北省"],["CN-HN","湖南省"],["CN-GD","广东省"],["CN-GX","广西壮族自治区"],
-  ["CN-HI","海南省"],["CN-CQ","重庆市"],["CN-SC","四川省"],["CN-GZ","贵州省"],["CN-YN","云南省"],
-  ["CN-XZ","西藏自治区"],["CN-SN","陕西省"],["CN-GS","甘肃省"],["CN-QH","青海省"],["CN-NX","宁夏回族自治区"],
-  ["CN-XJ","新疆维吾尔自治区"],
-  ["EXT-TRADE","外贸"],
-] as const;
-
-const chinaMapRegionCodes:Record<string,string>={
-  "110000":"CN-BJ","120000":"CN-TJ","130000":"CN-HE","140000":"CN-SX","150000":"CN-NM",
-  "210000":"CN-LN","220000":"CN-JL","230000":"CN-HL","310000":"CN-SH","320000":"CN-JS",
-  "330000":"CN-ZJ","340000":"CN-AH","350000":"CN-FJ","360000":"CN-JX","370000":"CN-SD",
-  "410000":"CN-HA","420000":"CN-HB","430000":"CN-HN","440000":"CN-GD","450000":"CN-GX",
-  "460000":"CN-HI","500000":"CN-CQ","510000":"CN-SC","520000":"CN-GZ","530000":"CN-YN",
-  "540000":"CN-XZ","610000":"CN-SN","620000":"CN-GS","630000":"CN-QH","640000":"CN-NX",
-  "650000":"CN-XJ","710000":"CN-TW","810000":"CN-HK","820000":"CN-MO",
-};
-
-const initialOrder = { orderNo: "", customerName: "", customerUnit: "", businessRegionSourceText:"", businessRegionCode: "", salespersonPersonId: "", serviceType: "", sourceReceivedOn: businessDateToday(), amount: "", reason: "首次录入" };
-
-function previousBusinessMonth():string{const [year,month]=businessDateToday().slice(0,7).split("-").map(Number);return new Date(Date.UTC(year!,month!-2,1)).toISOString().slice(0,7);}
-
-function LedgerGovernancePanel({user,orders,onChanged}:{user:User;orders:Order[];onChanged:()=>Promise<void>}){
-  const isLeader=user.roles.includes("sales_assistant_leader");
-  const isHr=user.roles.includes("hr");
-  const[periods,setPeriods]=useState<AccountingPeriod[]>([]);
-  const[corrections,setCorrections]=useState<AccountingCorrection[]>([]);
-  const[reviews,setReviews]=useState<HistoricalReview[]>([]);
-  const[month,setMonth]=useState(previousBusinessMonth());
-  const[periodNote,setPeriodNote]=useState("");
-  const[decisionNote,setDecisionNote]=useState("");
-  const[message,setMessage]=useState("");
-  const[busy,setBusy]=useState(false);
-  const[correction,setCorrection]=useState({orderId:"",eventType:"revenue_change",occurredOn:`${previousBusinessMonth()}-01`,reason:"",businessRegionCode:"",businessRegionSourceText:"",customerUnit:"",analysisDimensionEvidence:""});
-  const[review,setReview]=useState({orderId:"",lifecycleState:"active",currentRevenue:"",conclusion:"",evidence:"",reason:""});
-  const[execution,setExecution]=useState<AccountingCorrection|null>(null);
-  const[executionAmount,setExecutionAmount]=useState("");
-  const[executionReason,setExecutionReason]=useState("");
-  const load=useCallback(async()=>{
-    if(!isLeader&&!isHr)return;
-    const responses=await Promise.all([fetch("/api/accounting-periods"),fetch("/api/accounting-corrections"),fetch("/api/historical-order-reviews")]);
-    const data=await Promise.all(responses.map((response)=>response.json()));
-    const failed=responses.findIndex((response)=>!response.ok);
-    if(failed>=0)throw new Error((data[failed] as {message?:string}).message??"账本治理数据加载失败");
-    setPeriods((data[0] as {periods:AccountingPeriod[]}).periods??[]);
-    setCorrections((data[1] as {corrections:AccountingCorrection[]}).corrections??[]);
-    setReviews((data[2] as {reviews:HistoricalReview[]}).reviews??[]);
-  },[isHr,isLeader]);
-  useEffect(()=>{load().catch((error)=>setMessage(error instanceof Error?error.message:"账本治理数据加载失败"));},[load]);
-  useEffect(()=>{if(orders[0]){setCorrection((current)=>current.orderId?current:{...current,orderId:orders[0]!.id});setReview((current)=>current.orderId?current:{...current,orderId:orders[0]!.id});}},[orders]);
-  if(!isLeader&&!isHr)return null;
-  async function post(url:string,payload:unknown){
-    setBusy(true);setMessage("");
-    try{const response=await apiFetch(url,{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify(payload)});const data=await response.json() as {message?:string};if(!response.ok)throw new Error(data.message??"操作失败");await load();await onChanged();setMessage("操作已记录并刷新。");return true;}
-    catch(error){setMessage(error instanceof Error?error.message:"操作失败");return false;}
-    finally{setBusy(false);}
-  }
-  async function submitCorrection(event:FormEvent){event.preventDefault();const ok=await post("/api/accounting-corrections",{periodMonth:month,orderId:Number(correction.orderId),eventType:correction.eventType,occurredOn:correction.occurredOn,reason:correction.reason,businessRegionCode:correction.businessRegionCode,businessRegionSourceText:correction.businessRegionSourceText,customerUnit:correction.customerUnit,analysisDimensionEvidence:correction.analysisDimensionEvidence});if(ok)setCorrection((current)=>({...current,reason:"",analysisDimensionEvidence:""}));}
-  async function submitReview(event:FormEvent){event.preventDefault();const ok=await post("/api/historical-order-reviews",{orderId:Number(review.orderId),lifecycleState:review.lifecycleState,currentRevenue:Number(review.currentRevenue),conclusion:review.conclusion,evidence:review.evidence,reason:review.reason});if(ok)setReview((current)=>({...current,currentRevenue:"",conclusion:"",evidence:"",reason:""}));}
-  async function executeCorrection(event:FormEvent){event.preventDefault();if(!execution)return;const payload={type:execution.eventType,reason:executionReason,idempotencyKey:crypto.randomUUID(),correctionRequestId:Number(execution.id),...(execution.eventType==="revenue_change"?{newAmount:Number(executionAmount)}:{}),...(execution.eventType==="first_include"?{amount:Number(executionAmount)}:{})};const ok=await post(`/api/performance/orders/${execution.orderId}/events`,payload);if(ok){setExecution(null);setExecutionAmount("");setExecutionReason("");}}
-  const reviewableOrders=orders.filter((order)=>order.lifecycleState==="historical_review_required");
-  return <section className="governance-card" aria-labelledby="ledger-governance-title"><div className="orders-toolbar"><div><h2 id="ledger-governance-title">记账治理工作台</h2><span>关账、更正与历史核对均保留职责分离和审计</span></div><button className="icon-action" onClick={()=>load().catch((error)=>setMessage(error instanceof Error?error.message:"刷新失败"))} aria-label="刷新记账治理"><RefreshCw size={17}/></button></div>
-    {message?<p className="page-message" role="status">{message}</p>:null}
-    <div className="governance-grid"><form noValidate onSubmit={(event)=>event.preventDefault()}><h3>{isLeader&&isHr?"期间核对与关账":isHr?"人事关账":"组长核对确认"}</h3><label className="field"><span>记账月份</span><input type="month" value={month} onChange={(event)=>{setMonth(event.target.value);setCorrection((current)=>({...current,occurredOn:`${event.target.value}-01`}));}}/></label><Field label={isLeader&&isHr?"核对或关账说明":isHr?"关账说明":"核对说明"} value={periodNote} onChange={setPeriodNote}/>{isLeader?<button type="button" disabled={busy} onClick={()=>post(`/api/accounting-periods/${month}/confirm-close`,{note:periodNote})}>提交核对确认</button>:null}{isHr?<button type="button" disabled={busy} onClick={()=>post(`/api/accounting-periods/${month}/close`,{note:periodNote})}>关闭记账期间</button>:null}</form>
-      {isLeader?<form noValidate onSubmit={submitCorrection}><h3>申请关闭月更正</h3><OrderSelect orders={orders} value={correction.orderId} onChange={(orderId)=>setCorrection((current)=>({...current,orderId}))}/><label className="field"><span>更正类型</span><select value={correction.eventType} onChange={(event)=>setCorrection((current)=>({...current,eventType:event.target.value}))}><option value="revenue_change">营业额修改</option><option value="pause">整单暂停</option><option value="restart">订单重启</option><option value="first_include">首次计入</option></select></label><Field label="原业务日期" type="date" value={correction.occurredOn} onChange={(occurredOn)=>setCorrection((current)=>({...current,occurredOn}))}/><label className="field"><span>发生时标准业务区域</span><select required value={correction.businessRegionCode} onChange={(event)=>setCorrection((current)=>({...current,businessRegionCode:event.target.value}))}><option value="">请选择</option>{standardBusinessRegions.map(([code,name])=><option key={code} value={code}>{name}</option>)}</select></label><Field label="发生时来源区域原文" value={correction.businessRegionSourceText} onChange={(businessRegionSourceText)=>setCorrection((current)=>({...current,businessRegionSourceText}))}/><Field label="发生时客户单位" value={correction.customerUnit} onChange={(customerUnit)=>setCorrection((current)=>({...current,customerUnit}))}/><Field label="分析维度证据" value={correction.analysisDimensionEvidence} onChange={(analysisDimensionEvidence)=>setCorrection((current)=>({...current,analysisDimensionEvidence}))}/><Field label="申请原因" value={correction.reason} onChange={(reason)=>setCorrection((current)=>({...current,reason}))}/><button type="submit" disabled={busy}>提交更正申请</button></form>:null}
-      {isLeader?<form noValidate onSubmit={submitReview}><h3>提交历史订单核对</h3><OrderSelect orders={reviewableOrders} value={review.orderId} onChange={(orderId)=>setReview((current)=>({...current,orderId}))}/><label className="field"><span>核对后状态</span><select value={review.lifecycleState} onChange={(event)=>setReview((current)=>({...current,lifecycleState:event.target.value}))}><option value="active">正向计入</option><option value="paused">已暂停</option><option value="zero">零金额</option></select></label><Field label="核对后当前营业额" type="number" value={review.currentRevenue} onChange={(currentRevenue)=>setReview((current)=>({...current,currentRevenue}))}/><Field label="核对结论" value={review.conclusion} onChange={(conclusion)=>setReview((current)=>({...current,conclusion}))}/><Field label="核对依据" value={review.evidence} onChange={(evidence)=>setReview((current)=>({...current,evidence}))}/><Field label="核对原因" value={review.reason} onChange={(reason)=>setReview((current)=>({...current,reason}))}/><button type="submit" disabled={busy||!reviewableOrders.length}>提交人事审批</button></form>:null}</div>
-    {isHr?<label className="field governance-decision"><span>审批意见</span><input value={decisionNote} onChange={(event)=>setDecisionNote(event.target.value)} placeholder="批准、驳回或撤销前填写"/></label>:null}
-    <div className="governance-lists"><div><h3>记账期间</h3>{periods.length?<ul>{periods.map((period)=><li key={period.periodMonth}><strong>{period.periodMonth.slice(0,7)}</strong><span>{period.status==="closed"?`已关闭 · 版本 ${period.version}`:"开放"}{period.needsReclose?" · 待重新关账":""}</span></li>)}</ul>:<p>尚无期间治理记录。</p>}</div><div><h3>更正申请</h3>{corrections.length?<ul>{corrections.map((item)=><li key={item.id}><strong>{item.orderNo} · {eventTypeName(item.eventType)}</strong><span>{item.periodMonth.slice(0,7)} · {item.status} · 申请人 {item.requestedBy}</span><span>{item.businessRegionCode?`${businessRegionName(item.businessRegionCode)}（${item.businessRegionCode}） · ${item.businessRegionSourceText} · ${item.customerUnit} · 证据：${item.analysisDimensionEvidence}`:"缺少发生时分析维度证据，需重新提交"}</span><div className="row-actions">{isHr&&item.status==="pending"?<><button onClick={()=>post(`/api/accounting-corrections/${item.id}/approve`,{note:decisionNote})}>批准</button><button onClick={()=>post(`/api/accounting-corrections/${item.id}/reject`,{note:decisionNote})}>驳回</button></>:null}{isHr&&item.status==="approved"?<button onClick={()=>post(`/api/accounting-corrections/${item.id}/revoke`,{note:decisionNote})}>撤销</button>:null}{isLeader&&item.status==="approved"?<button onClick={()=>setExecution(item)}>执行更正</button>:null}</div></li>)}</ul>:<p>暂无更正申请。</p>}</div><div><h3>历史核对</h3>{reviews.length?<ul>{reviews.map((item)=><li key={item.id}><strong>{item.orderNo} · {item.conclusion}</strong><span>{item.status} · 核对人 {item.requestedBy} · 依据 {item.evidence}</span>{isHr&&item.status==="pending"?<div className="row-actions"><button onClick={()=>post(`/api/historical-order-reviews/${item.id}/approve`,{note:decisionNote})}>批准并解析</button><button onClick={()=>post(`/api/historical-order-reviews/${item.id}/reject`,{note:decisionNote})}>驳回</button></div>:null}</li>)}</ul>:<p>暂无历史核对记录。</p>}</div></div>
-    {execution?<Modal title={`执行更正 · ${execution.orderNo}`} note={`获批范围：${execution.periodMonth.slice(0,7)} / ${eventTypeName(execution.eventType)}`} onClose={()=>setExecution(null)}><form noValidate className="business-form" onSubmit={executeCorrection}>{["revenue_change","first_include"].includes(execution.eventType)?<Field label={execution.eventType==="revenue_change"?"调整后营业额":"首次计入金额"} type="number" value={executionAmount} onChange={setExecutionAmount}/>:null}<Field label="执行原因" value={executionReason} onChange={setExecutionReason}/><div className="modal-actions"><button type="button" onClick={()=>setExecution(null)}>取消</button><button type="submit" disabled={busy}>确认追加更正事件</button></div></form></Modal>:null}
-  </section>;
-}
-
-function OrderSelect({orders,value,onChange}:{orders:Order[];value:string;onChange:(value:string)=>void}){return <label className="field"><span>订单</span><select aria-label="订单" value={orders.some((order)=>order.id===value)?value:""} onChange={(event)=>onChange(event.target.value)}><option value="">请选择当前列表中的订单</option>{orders.map((order)=><option key={order.id} value={order.id}>{order.orderNo} · {order.customerName}</option>)}</select></label>;}
-
-function CreateOrder({ onClose, onSaved }: { onClose: () => void; onSaved: () => Promise<void> }) {
-  const [form, setForm] = useState(initialOrder);
-  const [people, setPeople] = useState<Array<{ id:string; displayName:string }>>([]);
-  const [error, setError] = useState("");
-  const [saving, setSaving] = useState(false);
-  useEffect(() => { fetch("/api/performance/people").then(async (response) => {
-    const data = await response.json() as { people?:Array<{ id:string; displayName:string }>; message?:string };
-    if (!response.ok) throw new Error(data.message ?? "业务员列表加载失败");
-    setPeople(data.people ?? []);
-  }).catch((reason) => setError(reason instanceof Error ? reason.message : "业务员列表加载失败")); }, []);
-  function set(name: keyof typeof form, value: string) { setForm((current) => ({ ...current, [name]: value })); }
-  async function submit(event: FormEvent) {
-    event.preventDefault(); if (saving) return; setSaving(true); setError("");
-    try {
-      const response = await apiFetch("/api/performance/orders", { method: "POST", headers: { "content-type": "application/json" }, body: JSON.stringify({ ...form, salespersonPersonId: Number(form.salespersonPersonId), amount: Number(form.amount) }) });
-      const data = await readResponseJson<{ message?: string }>(response,"订单入账响应无效，请重试。");
-      if (!response.ok) throw new Error(data.message ?? "订单入账失败");
-      await onSaved();
-    } catch (reason) { setError(reason instanceof Error ? reason.message : "订单入账失败"); }
-    finally { setSaving(false); }
-  }
-  return <Modal title="录入订单业绩" note="组织归属和分析维度按发生时事实固化，后续只能追加更正事件" onClose={onClose} preventClose={saving}><form noValidate className="business-form" onSubmit={submit}><div className="form-grid"><Field label="订单编号" value={form.orderNo} onChange={(v) => set("orderNo",v)}/><Field label="收到日期" value={form.sourceReceivedOn} type="date" onChange={(v) => set("sourceReceivedOn",v)}/><Field label="客户名称" value={form.customerName} onChange={(v) => set("customerName",v)}/><Field label="客户单位" value={form.customerUnit} onChange={(v) => set("customerUnit",v)}/><Field label="来源区域原文" value={form.businessRegionSourceText} onChange={(v)=>set("businessRegionSourceText",v)}/><label className="field"><span>标准业务区域</span><select required value={form.businessRegionCode} onChange={(event) => set("businessRegionCode",event.target.value)}><option value="">请选择</option>{standardBusinessRegions.map(([code,name]) => <option key={code} value={code}>{name}</option>)}</select></label><label className="field"><span>业务员</span><select required value={form.salespersonPersonId} onChange={(event) => set("salespersonPersonId",event.target.value)}><option value="">请选择</option>{people.map((person) => <option key={person.id} value={person.id}>{person.displayName}</option>)}</select></label><Field label="服务类型" value={form.serviceType} onChange={(v) => set("serviceType",v)}/><Field label="营业额" value={form.amount} type="number" onChange={(v) => set("amount",v)}/><Field label="入账原因" value={form.reason} onChange={(v) => set("reason",v)}/></div>{error ? <p className="form-error" role="alert">{error}</p> : null}<div className="modal-actions"><button type="button" onClick={onClose} disabled={saving}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving ? "正在入账…" : "确认入账"}</button></div></form></Modal>;
-}
-
-function AdjustOrder({ order, canEdit, onClose, onSaved }: { order: Order; canEdit: boolean; onClose: () => void; onSaved: () => Promise<void> }) {
-  const [events,setEvents]=useState<PerformanceEvent[]>([]);
-  const [allowed,setAllowed]=useState<string[]>([]);
-  const [lifecycle,setLifecycle]=useState<OrderLifecycle>(order.lifecycleState);
-  const [loading,setLoading]=useState(true);
-  const [type, setType] = useState("");
-  const [amount, setAmount] = useState("");
-  const [reason, setReason] = useState("");
-  const [error, setError] = useState("");
-  const [saving,setSaving]=useState(false);
-  const idempotencyKey=useRef(crypto.randomUUID());
-  useEffect(()=>{
-    const controller=new AbortController();
-    fetch(`/api/performance/orders/${order.id}/events`,{signal:controller.signal}).then(async(response)=>{
-      const data=await response.json() as {events?:PerformanceEvent[];lifecycleState?:OrderLifecycle;allowedActions?:string[];message?:string};
-      if(!response.ok)throw new Error(data.message??"事件链加载失败");
-      const nextAllowed=data.allowedActions??[];setEvents(data.events??[]);setLifecycle(data.lifecycleState??order.lifecycleState);setAllowed(nextAllowed);setType((current)=>nextAllowed.includes(current)?current:(nextAllowed[0]??""));
-    }).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setError(failure instanceof Error?failure.message:"事件链加载失败");}).finally(()=>{if(!controller.signal.aborted)setLoading(false);});
-    return()=>controller.abort();
-  },[order.id,order.lifecycleState]);
-  async function submit(event: FormEvent) {
-    event.preventDefault();if(saving)return;setSaving(true);setError("");
-    const payload = { type, reason, idempotencyKey:idempotencyKey.current, ...(type === "revenue_change" ? { newAmount: Number(amount) } : {}), ...(type === "first_include" ? { amount: Number(amount) } : {}) };
-    try{
-      const response = await apiFetch(`/api/performance/orders/${order.id}/events`, { method: "POST", headers: { "content-type": "application/json" }, body: JSON.stringify(payload) });
-      const data = await readResponseJson<{ message?: string }>(response,"调整入账响应无效，请重试。");
-      if (!response.ok) { setError(data.message ?? "调整入账失败"); return; }
-      await onSaved();
-    }catch{setError("网络异常，结果尚未确认；可安全重试，系统不会重复入账。");}finally{setSaving(false);}
-  }
-  return <Modal title={order.orderNo} note={`${order.customerName} · 当前营业额 ${formatMoney(order.currentRevenue)} · 计入 ${formatMoney(order.countedAmount)}`} onClose={onClose} preventClose={saving}><section className="event-ledger" aria-labelledby="event-ledger-title"><div className="event-ledger-heading"><div><h3 id="event-ledger-title">不可变事件链</h3><p>{loading?"正在读取事件…":`${events.length} 条事件 · 按服务端账本序号排列`}</p></div><Status state={lifecycle}/></div>{!loading&&events.length?<ol>{events.map((item)=><li key={item.id}><span className="event-sequence" aria-label={`第 ${item.sequence} 条事件`}>{item.sequence}</span><div className="event-content"><div className="event-summary"><strong>{eventTypeName(item.eventType)}</strong><b className={Number(item.deltaAmount)<0?"negative":""}>{Number(item.deltaAmount)>0?"+":""}{formatMoney(item.deltaAmount)}</b>{item.resultingLifecycleState?<Status state={item.resultingLifecycleState}/>:<span className="legacy-semantic-note">原始状态未推断</span>}</div><dl><div><dt>业务日 / 记账月</dt><dd>{item.occurredOn} / {item.accountingMonth.slice(0,7)}</dd></div><div><dt>操作时间</dt><dd>{formatOperationTime(item.occurredAt)}</dd></div><div><dt>投影结果</dt><dd>营业额 {formatMoney(item.resultingCurrentRevenue)} · 计入 {formatMoney(item.resultingCountedAmount)}</dd></div><div><dt>原因 / 操作者</dt><dd>{item.reason||"—"} / {item.actorName||"历史导入"}</dd></div><div><dt>组织快照</dt><dd>{[item.departmentName,item.groupName].filter(Boolean).join(" / ")||"—"} · 组长 {item.leaderName||"—"} · 主管 {item.supervisorName||"—"}</dd></div><div><dt>分析维度快照</dt><dd>{item.businessRegionCode?`${businessRegionName(item.businessRegionCode)} (${item.businessRegionCode}) · 来源 ${item.businessRegionSourceText} · 客户单位 ${item.customerUnit}`:"待补齐（无受控来源证据）"}</dd></div></dl></div></li>)}</ol>:!loading&&!error?<p className="event-empty">没有可显示的事件。</p>:null}</section>{error ? <p className="form-error event-error" role="alert">{error}</p> : null}{!loading&&canEdit && allowed.length ? <form noValidate className="business-form event-form" onSubmit={submit}><div className="event-options">{allowed.includes("revenue_change") ? <button type="button" className={type==="revenue_change"?"selected":""} aria-pressed={type==="revenue_change"} onClick={() => setType("revenue_change")}><RefreshCw size={17}/>修改营业额</button> : null}{allowed.includes("pause") ? <button type="button" className={type==="pause"?"selected":""} aria-pressed={type==="pause"} onClick={() => setType("pause")}><PauseCircle size={17}/>整单暂停</button> : null}{allowed.includes("restart") ? <button type="button" className={type==="restart"?"selected":""} aria-pressed={type==="restart"} onClick={() => setType("restart")}><PlayCircle size={17}/>订单重启</button> : null}{allowed.includes("first_include") ? <button type="button" className={type==="first_include"?"selected":""} aria-pressed={type==="first_include"} onClick={() => setType("first_include")}><Plus size={17}/>首次计入</button> : null}</div><p className="form-note">操作时间和记账月由服务器确定，组织快照按该业务日的有效任职自动解析。</p><div className="form-grid">{type === "revenue_change" || type === "first_include" ? <Field label={type === "revenue_change" ? "调整后营业额" : "首次计入金额"} value={amount} type="number" onChange={setAmount}/> : null}<Field label="原因（必填）" value={reason} onChange={setReason}/></div><div className="modal-actions"><button type="button" onClick={onClose} disabled={saving}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在追加…":"确认追加事件"}</button></div></form> : !loading?<div className="permission-note">{lifecycle==="historical_review_required"?"该历史订单需要先完成核对与人事批准，当前不能追加事件。":"当前订单状态没有可执行操作，或当前角色仅可查看。"}</div>:null}</Modal>;
-}
-
-function Field({ label, value, onChange, type="text" }: { label: string; value: string; onChange: (value: string) => void; type?: string }) { return <label className="field"><span>{label}</span><input required value={value} type={type} min={type === "number" ? "0" : undefined} step={type === "number" ? "0.01" : undefined} onChange={(event) => onChange(event.target.value)}/></label>; }
-function Modal({ title, note, onClose, preventClose=false, children }: { title: string; note: string; onClose: () => void; preventClose?: boolean; children: ReactNode }) { const titleId=useId();const noteId=useId();const dialogRef=useRef<HTMLElement>(null);const closeRef=useRef(onClose);const requestClose=()=>{if(!preventClose)onClose();};closeRef.current=requestClose;useEffect(()=>{const prior=document.activeElement instanceof HTMLElement?document.activeElement:null;const root=document.getElementById("root");const priorOverflow=document.body.style.overflow;root?.setAttribute("inert","");root?.setAttribute("aria-hidden","true");document.body.style.overflow="hidden";const dialog=dialogRef.current;dialog?.focus();function keydown(event:KeyboardEvent){if(event.key==="Escape"){event.preventDefault();closeRef.current();return;}if(event.key!=="Tab"||!dialog)return;const controls=[...dialog.querySelectorAll<HTMLElement>('button:not([disabled]),input:not([disabled]),select:not([disabled]),textarea:not([disabled]),[href],[tabindex]:not([tabindex="-1"])')].filter((item)=>!item.hasAttribute("hidden"));if(!controls.length){event.preventDefault();dialog.focus();return;}const first=controls[0]!;const last=controls.at(-1)!;if(event.shiftKey&&document.activeElement===first){event.preventDefault();last.focus();}else if(!event.shiftKey&&document.activeElement===last){event.preventDefault();first.focus();}}document.addEventListener("keydown",keydown);return()=>{document.removeEventListener("keydown",keydown);root?.removeAttribute("inert");root?.removeAttribute("aria-hidden");document.body.style.overflow=priorOverflow;prior?.focus();};},[]);return createPortal(<div className="modal-backdrop" role="presentation" onMouseDown={(event) => { if (event.target === event.currentTarget) requestClose(); }}><section ref={dialogRef} tabIndex={-1} className="modal" role="dialog" aria-modal="true" aria-labelledby={titleId} aria-describedby={noteId}><header><div><h2 id={titleId}>{title}</h2><p id={noteId}>{note}</p></div><button type="button" onClick={requestClose} disabled={preventClose} aria-label="关闭">×</button></header>{children}</section></div>,document.body); }
-function Status({ state }: { state: OrderLifecycle }) { const names:Record<OrderLifecycle,string> = { draft:"草稿", active:"正向计入", paused:"已暂停", zero:"零金额",historical_review_required:"待历史核对" }; return <span className={`status status-${state}`}>{names[state]}</span>; }
-function achievementReasonText(value:PersonalAchievement["calculationReason"],scope:string){return value===null?null:({PERIOD_IN_FUTURE:"未来月份不计算目标达成",TARGET_NOT_ACTIVE:`${scope}目标尚未生效`,TARGET_AMOUNT_NOT_POSITIVE:"生效目标金额必须大于零",TARGET_SCOPE_AMBIGUOUS:"存在多个生效的销售组织目标，暂不计算"} as const)[value];}
-function OrganizationAchievementCard({scope,title,countNote,achievement,onOpen}:{scope:"部门"|"销售组织";title:string;countNote:string;achievement:PersonalAchievement;onOpen:(metric:"actual"|"gap")=>void}){const reason=achievementReasonText(achievement.calculationReason,scope);return <article className="team-achievement orders-card" aria-label={title}><PanelTitle title={title} note={countNote}/><section className="metric-band personal-metrics"><Metric label={`${scope}生效目标`} value={achievement.targetAmount===null?"—":formatMoney(achievement.targetAmount)} note={reason??`当前月${scope}目标`}/><Metric label={`${scope}业绩`} value={formatMoney(achievement.actualAmount)} note={`${achievement.eventCount} 条事件`} onClick={()=>onOpen("actual")} actionLabel={`查看${scope}业绩构成`}/><Metric label={`${scope}差距`} value={achievement.gapAmount===null?"不计算":formatMoney(achievement.gapAmount)} note={reason??`生效目标减${scope}业绩`} onClick={achievement.gapAmount===null?undefined:()=>onOpen("gap")} actionLabel={achievement.gapAmount===null?undefined:`查看${scope}差距构成`}/><Metric label={`${scope}达成率`} value={achievement.achievementRate===null?"不计算":`${achievement.achievementRate}%`} note={reason??`${scope}业绩 ÷ 生效目标`}/><Metric label="自然日进度" value={achievement.timeProgressRate===null?"不计算":`${achievement.timeProgressRate}%`} note={reason??"按上海自然日计算"}/><Metric label="进度偏差" value={achievement.progressVariance===null?"不计算":`${achievement.progressVariance}%`} note={reason??"达成率减自然日进度"} warning={achievement.progressVariance!==null&&Number(achievement.progressVariance)<0}/></section></article>}
-function OrganizationHierarchy({departments}:{departments:DepartmentAchievementDrilldown[]}){
-  return <div className="group-drilldown">{departments.map((department)=>{
-    const departmentReason=achievementReasonText(department.calculationReason,"部门");
-    return <section className="orders-card group-member" aria-label={`${department.departmentName}部门构成`} key={department.departmentKey}>
-      <PanelTitle title={department.departmentName} note={`${department.groupCount} 个小组 · ${department.eventCount} 条事件 · 实际 ${formatMoney(department.actualAmount)} · ${departmentReason??`达成率 ${department.achievementRate}%`}`}/>
-      {department.groups.map((group)=>{
-        const groupReason=achievementReasonText(group.calculationReason,"小组");
-        return <article className="orders-card group-member" aria-label={`${group.groupName}小组构成`} key={group.groupKey}>
-          <PanelTitle title={group.groupName} note={`${group.memberCount} 位事件责任人 · ${group.eventCount} 条事件 · 实际 ${formatMoney(group.actualAmount)} · ${groupReason??`达成率 ${group.achievementRate}%`}`}/>
-          <AchievementMembers members={group.members}/>
-          {group.members.length===0?<p className="empty-cell">当前月份没有小组业绩事件。</p>:null}
-        </article>;
-      })}
-      {department.groups.length===0?<p className="empty-cell">当前月份没有部门小组。</p>:null}
-    </section>;
-  })}</div>;
-}
-function AchievementMembers({members}:{members:GroupAchievementMember[]}){
-  return <>{members.map((member)=><section className="orders-card group-member" key={member.personKey}>
-    <PanelTitle title={member.name} note={`${member.orders.length} 笔订单 · ${member.eventCount} 条事件 · 净额 ${formatMoney(member.actualAmount)}`}/>
-    {member.orders.map((order)=><div className="group-order" key={order.orderId}>
-      <h4>{order.orderNo} · {order.customerName}<small>{order.eventCount} 条事件 · 净额 {formatMoney(order.actualAmount)}</small></h4>
-      <div className="orders-table-wrap"><table><thead><tr><th>不可变事件</th><th>业务日 / 记账月</th><th>金额方向</th><th>状态 / 计入</th><th>组织归属</th><th>分析维度</th><th>原因</th></tr></thead><tbody>{order.events.map((event)=><tr key={event.id}><td>{eventTypeName(event.eventType)} · 第 {event.sequence} 条</td><td>{event.occurredOn} / {event.accountingMonth}</td><td className={event.deltaAmount.startsWith("-")?"negative":""}>{event.deltaAmount.startsWith("-")?"负向":"正向"} {formatMoney(event.deltaAmount.startsWith("-")?event.deltaAmount.slice(1):event.deltaAmount)}</td><td>{event.resultingLifecycleState?<><Status state={event.resultingLifecycleState}/><small className="table-secondary">计入至 {formatMoney(event.resultingCountedAmount)}</small></>:"原始状态未推断"}</td><td>{[event.departmentName,event.groupName].filter(Boolean).join(" / ")||"—"}</td><td>待补齐</td><td>{event.reason??"—"}</td></tr>)}</tbody></table></div>
-    </div>)}
-  </section>)}</>;
-}
-function formatMoney(value: string | number) {
-  if(typeof value==="string"){
-    const decimal=/^(-?)(\d+)(?:\.(\d{1,2}))?$/.exec(value);
-    if(decimal)return `${decimal[1]}¥${new Intl.NumberFormat("zh-CN").format(BigInt(decimal[2]!))}.${(decimal[3]??"").padEnd(2,"0")}`;
-  }
-  return new Intl.NumberFormat("zh-CN", { style: "currency", currency: "CNY", minimumFractionDigits: 2 }).format(Number(value));
-}
-function formatOperationTime(value:string){return new Intl.DateTimeFormat("zh-CN",{timeZone:"Asia/Shanghai",year:"numeric",month:"2-digit",day:"2-digit",hour:"2-digit",minute:"2-digit",second:"2-digit",hour12:false}).format(new Date(value));}
-function eventTypeName(type: string) { return ({ initial:"首次录入", revenue_change:"营业额修改", pause:"整单暂停", restart:"订单重启", first_include:"首次计入", legacy_adjustment:"历史迁移",historical_review_resolution:"历史核对解析" } as Record<string,string>)[type] ?? type; }
-function businessRegionName(code:string){return standardBusinessRegions.find(([value])=>value===code)?.[1]??code;}
-function goalLevelName(level: string) { return ({ sales_manager:"销售经理总目标", department:"部门目标", group:"小组目标", personal:"个人目标" } as Record<string,string>)[level] ?? level; }
-function goalStatusName(status: string) { return ({ draft:"草稿", pending_signature:"待责任人确认", pending_gm:"待总经理审批", pending_hr:"待人事审批", active:"已生效", rejected:"已拒绝", superseded:"已替代" } as Record<string,string>)[status] ?? status; }
-function workflowStatusName(status:string){return ({pending:"待处理",accepted:"已接受，待重新确认",rejected:"已拒绝",withdrawn:"已撤回",invalidated:"已失效",completed:"已完成"} as Record<string,string>)[status]??status;}
-function goalAuditName(action:string){return ({"goal.version_created":"创建目标版本","goal.version_confirmed":"责任人实名确认","goal.version_signed":"历史责任人确认","goal.approved":"审批批准","goal.rejected":"审批拒绝","goal.version_superseded":"替代旧版本","goal.change_requested":"提交修改申请","goal.change_accepted":"接受修改申请","goal.change_rejected":"拒绝修改申请","goal.change_withdrawn":"撤回修改申请","goal.change_invalidated":"修改申请失效","goal.change_completed":"修改流程完成","goal.linkage_requested":"生成目标联动待办","goal.linkage_decided":"完成目标联动选择"} as Record<string,string>)[action]??action;}
-
-function PanelTitle({ title, note }: { title: string; note: string }) { return <div className="panel-title"><div><h2>{title}</h2><p>{note}</p></div></div>; }
-function Metric({label,value,note,warning,negative,onClick,actionLabel}:{label:string;value:string;note:string;warning?:boolean;negative?:boolean;onClick?:(()=>void)|undefined;actionLabel?:string|undefined}) { return <div className="metric"><span>{label}</span>{onClick?<button type="button" className={`metric-value-action${negative?" negative":""}`} aria-label={actionLabel} onClick={onClick}>{value}</button>:<strong className={negative?"negative":""}>{value}</strong>}<small className={warning?"warning":""}>{note}</small></div>; }
-function TrendChart({year,monthly}:{year:string;monthly:Array<{month:string;total:string}>}) { const totals=new Map(monthly.map((item)=>[item.month,item.total]));const raw=Array.from({length:12},(_,index)=>totals.get(`${year}-${String(index+1).padStart(2,"0")}`)??"0");const filled=raw.map(Number);const max=Math.max(1,...filled);const x=(index:number)=>44+index*(526/11);const points=filled.map((value,index)=>`${x(index)},${190-(value/max)*150}`).join(" ");const description=raw.map((value,index)=>`${index+1}月 ${formatMoney(value)}`).join("；");return <div className="chart"><svg viewBox="0 0 590 220" role="img" aria-label={`${year}年1月至12月业绩折线图`}><desc>{description}</desc><g className="grid-lines">{[40,90,140,190].map((y)=><line key={y} x1="42" y1={y} x2="570" y2={y}/>)}</g><polyline points={points} fill="none" stroke="#2f6fed" strokeWidth="4" strokeLinecap="round" strokeLinejoin="round"/>{filled.map((value,index)=><circle key={index} cx={x(index)} cy={190-(value/max)*150} r="5" fill="#fff" stroke="#2f6fed" strokeWidth="3"/>)}</svg><div className="chart-labels">{Array.from({length:12},(_,index)=><span key={index}>{index+1}月</span>)}</div></div>; }
