+++ apps/web/src/app-types.ts
@@ -1,0 +1,49 @@
+export type Capabilities = { viewPerformance:boolean; viewGoals:boolean; viewAudits:boolean; viewOrganization:boolean; viewApprovals:boolean; editPerformance:boolean; exportPerformance:boolean; exportGoals:boolean; manageAccounts:boolean; manageOrganization:boolean };
+export type User = { id: string; personId:string; username: string; displayName: string; mustChangePassword: boolean; roles: string[]; capabilities:Capabilities };
+export type AuthState = { status: "loading" } | { status: "guest" } | { status: "authenticated"; user: User };
+export type OrderLifecycle = "draft" | "active" | "paused" | "zero" | "historical_review_required";
+export type Order = { id: string; orderNo: string; customerName: string; customerUnit: string; salespersonName: string; serviceType: string | null; sourceReceivedOn: string; originalAmount: string; currentRevenue: string; countedAmount: string; lifecycleState: OrderLifecycle; postedAt: string; departmentName:string; groupName:string; leaderName:string|null; supervisorName:string|null };
+export type OrderFilters = { search:string;month:string;status:string;salesperson:string;department:string;group:string;region:string;customerUnit:string };
+export type PerformanceEvent = { id:string; sequence:number; eventType:string; deltaAmount:string; resultingCurrentRevenue:string; resultingCountedAmount:string; resultingLifecycleState:OrderLifecycle|null; accountingMonth:string; occurredOn:string; occurredAt:string; reason:string|null; actorName:string|null; salespersonName:string; departmentName:string|null; groupName:string|null; leaderName:string|null; supervisorName:string|null; businessRegionCode:string|null; businessRegionSourceText:string|null; customerUnit:string|null };
+export type AccountingPeriod={periodMonth:string;status:"open"|"closed";version:number;needsReclose:boolean;verifiedAt:string|null;verifiedBy:string|null;closedAt:string|null;closedBy:string|null};
+export type AccountingCorrection={id:string;periodMonth:string;orderId:string;orderNo:string;eventType:string;occurredOn:string;reason:string;businessRegionCode:string|null;businessRegionSourceText:string|null;customerUnit:string|null;analysisDimensionEvidence:string|null;status:"pending"|"approved"|"rejected"|"consumed"|"revoked";requestedBy:string;reviewedBy:string|null;reviewNote:string|null;expiresAt:string|null};
+export type HistoricalReview={id:string;orderId:string;orderNo:string;lifecycleState:"active"|"paused"|"zero";currentRevenue:string;conclusion:string;evidence:string;reason:string;status:"pending"|"approved"|"rejected";requestedBy:string;reviewedBy:string|null;reviewNote:string|null};
+export type PersonalAchievementEvent = {id:string;orderId:string;orderNo:string;customerName:string;eventType:string;deltaAmount:string;accountingMonth:string;occurredOn:string;sequence:number;reason:string|null;resultingCountedAmount:string;resultingLifecycleState:"active"|"paused"|"zero"|null;departmentName:string|null;groupName:string|null};
+export type PersonalAchievement = { goalId:string|null;periodMonth:string;targetAmount:string|null;actualAmount:string;gapAmount:string|null;achievementRate:string|null;timeProgressRate:string|null;progressVariance:string|null;calculationReason:"PERIOD_IN_FUTURE"|"TARGET_NOT_ACTIVE"|"TARGET_AMOUNT_NOT_POSITIVE"|"TARGET_SCOPE_AMBIGUOUS"|null;eventCount:number };
+export type PersonalAchievementDrilldown = PersonalAchievement&{events:PersonalAchievementEvent[]};
+export type GroupAchievement = PersonalAchievement&{groupId:string;groupName:string;memberCount:number};
+export type GroupAchievementOrder = {orderId:string;orderNo:string;customerName:string;actualAmount:string;eventCount:number;events:PersonalAchievementEvent[]};
+export type GroupAchievementMember = {personId:string|null;personKey:string;name:string;actualAmount:string;eventCount:number;orders:GroupAchievementOrder[]};
+export type GroupAchievementDrilldown = GroupAchievement&{members:GroupAchievementMember[]};
+export type OrganizationGroupAchievement = PersonalAchievement&{groupId:string|null;groupKey:string;groupName:string;memberCount:number;members:GroupAchievementMember[]};
+export type DepartmentAchievement = PersonalAchievement&{departmentId:string|null;departmentKey:string;departmentName:string;groupCount:number};
+export type DepartmentAchievementDrilldown = DepartmentAchievement&{groups:OrganizationGroupAchievement[]};
+export type SalesAchievement = PersonalAchievement&{departmentCount:number};
+export type SalesAchievementDrilldown = SalesAchievement&{departments:DepartmentAchievementDrilldown[]};
+export type DashboardData = { month: string; metrics: { total: string; eventCount: number; negativeTotal: string; pendingApprovals: number }; monthly: Array<{ month: string; total: string }>; groups: Array<{ id:string;name:string;total:string }>; recent: Array<{ orderNo: string; salespersonName: string; eventType: string; month: string; amount: string; groupName: string }>; personalAchievement:PersonalAchievement|null;groupAchievements:GroupAchievement[];departmentAchievements:DepartmentAchievement[];salesAchievement:SalesAchievement|null };
+export type AnalysisAmount = { eventCount:number;totalAmount:string };
+export type AnalysisProvince = {regionCode:string;regionName:string;eventCount:number;totalAmount:string};
+export type AnalysisCustomer = {customerUnit:string;eventCount:number;totalAmount:string};
+export type PerformanceAnalysis = { month:string;ledger:AnalysisAmount;mapped:AnalysisAmount;pending:AnalysisAmount;reconciled:boolean;provinces:AnalysisProvince[];foreignTrade:{regionCode:"EXT-TRADE";regionName:string;eventCount:number;totalAmount:string};customers:Array<AnalysisProvince&{customerUnit:string}> };
+export type AnalysisCustomersDrilldown = {level:"customers";regionCode:string;regionName:string;month:string;eventCount:number;totalAmount:string;customerCount:number;nextCursor:string|null;pageSize:number;customers:AnalysisCustomer[]};
+export type AnalysisMonthsDrilldown = {level:"months";regionCode:string;regionName:string;customerUnit:string;year:string;eventCount:number;totalAmount:string;months:Array<{month:string;eventCount:number;totalAmount:string}>};
+export type AnalysisDrilldownEvent = {id:string;eventType:string;deltaAmount:string;resultingCurrentRevenue:string;resultingCountedAmount:string;resultingLifecycleState:"active"|"paused"|"zero"|null;accountingMonth:string;occurredOn:string;reason:string|null;salespersonName:string;departmentName:string|null;groupName:string|null;sequence:number;businessRegionCode:string;businessRegionSourceText:string;customerUnit:string};
+export type AnalysisDrilldownOrder = {orderId:string;orderNo:string;customerName:string;eventCount:number;totalAmount:string;events:AnalysisDrilldownEvent[]};
+export type AnalysisEventsDrilldown = {level:"events";regionCode:string;regionName:string;customerUnit:string;month:string;eventCount:number;totalAmount:string;nextCursor:string|null;pageSize:number;orders:AnalysisDrilldownOrder[]};
+export type GoalLevel="sales_manager"|"department"|"group"|"personal";
+export type Goal = { id:string;periodMonth:string;level:GoalLevel;ownerUsername:string|null;ownerName:string;ownerPersonId:string;orgUnitId:string|null;orgUnitName:string|null;parentGoalId:string|null;versionId:string;versionNo:string;amount:string;effectiveAmount:string|null;status:string;signatureText:string|null;signedAt:string|null;changeReason:string;allocatedAmount:string;allocationDifference:string;allocationType:"unallocated"|"overallocated"|"balanced";allocationRatio:string|null };
+export type GoalOption={personId:string;name:string;orgUnitId:string|null;orgUnitName:string|null};
+export type ParentGoalOption={id:string;ownerName:string;orgUnitId:string|null;orgUnitName:string|null;amount:string};
+export type GoalChangeRequest={id:string;goalId:string;level:GoalLevel;ownerName:string;requestedAmount:string|null;reason:string;status:string;outcomeComment:string|null;createdVersionId:string|null;createdAt:string;currentAmount:string|null;newAmount:string|null;amountDifference:string|null;requestedByName:string;handledByName:string|null;canHandle:boolean;canWithdraw:boolean};
+export type GoalLinkageDecision={id:string;parentGoalId:string;triggeringChildVersionId:string;decision:"keep_parent"|"adjust_parent"|null;status:string;reason:string|null;generatedChangeRequestId:string|null;childGoalId:string;childOwnerName:string;childAmount:string;decidedAt:string|null;parentLevel:GoalLevel;parentAmount:string;canDecide:boolean};
+export type GoalWorkflows={changeRequests:GoalChangeRequest[];linkageDecisions:GoalLinkageDecision[]};
+export type GoalHistory={versions:Array<{id:string;versionNo:string;amount:string;previousAmount:string|null;amountDifference:string|null;status:string;createdByName:string;signedByName:string|null;signatureText:string|null;signedAt:string|null;changeReason:string;createdAt:string}>;approvals:Array<{id:string;goalVersionId:string;stage:string;decision:string;decidedByName:string;comment:string|null;decidedAt:string}>;changeRequests:Array<{id:string;requestedAmount:string|null;reason:string;status:string;outcomeComment:string|null;createdAt:string;requestedByName:string;handledByName:string|null}>;audit:Array<{id:string;action:string;entityType:string;entityId:string;actorName:string|null;beforeData:unknown;afterData:unknown;createdAt:string}>};
+export type FormalReport = { goalId:string; periodMonth:string; level:Goal["level"]; ownerName:string; targetAmount:string; actualAmount:string; gapAmount:string; achievementRate:string|null };
+export type AdminUser = { id:string; username:string; displayName:string; isActive:boolean; mustChangePassword:boolean; roles:string[] };
+export type PersonOption = { id:string; displayName:string; username:string|null };
+export type RolePermission = { code:string; name:string; businessScope:"none"|"self"|"group"|"department"|"all"; businessOperations:readonly string[]; targetResponsibilities:string; exportPermission:string; forbidden:readonly string[] };
+export type OrgUnit = { id:string; name:string; unitType:"department"|"group"; parentId:string|null; parentName:string|null; isActive:boolean };
+export type Assignment = { id:string; username:string; displayName:string; departmentName:string|null; groupName:string|null; effectiveFrom:string; effectiveTo:string|null };
+export type Responsibility = { id:string; personId:string; username:string|null; displayName:string; unitId:string; unitName:string; unitType:"department"|"group"; responsibilityType:"leader"|"supervisor"; effectiveFrom:string; effectiveTo:string|null };
+export type AuditRow = { id:string; actorPersonId:string|null; actorUsername:string|null; actorDisplayName:string|null; action:string; entityType:string; entityId:string|null; beforeData:unknown; afterData:unknown; createdAt:string };
+export type AuditFilters = { person:string; action:string; entityType:string; entityId:string; from:string; to:string };
