+++ apps/web/src/pages/accounts-page.tsx
@@ -1,0 +1,34 @@
+import { type FormEvent, useCallback, useEffect, useRef, useState } from "react";
+import { Plus, Search, ShieldCheck, X } from "lucide-react";
+import { apiFetch, readResponseJson, roleNames } from "../app-api";
+import type { AdminUser, PersonOption, RolePermission, User } from "../app-types";
+import { Field, Modal } from "../shared-ui";
+
+export function AccountsPage({user}:{user:User}){
+  const initialSearch=useRef(new URLSearchParams(window.location.search).get("accountSearch")??"").current;
+  const [users,setUsers]=useState<AdminUser[]>([]);const [roles,setRoles]=useState<Array<{code:string;name:string}>>([]);const [permissionMatrix,setPermissionMatrix]=useState<RolePermission[]>([]);const [message,setMessage]=useState("");const [loadError,setLoadError]=useState("");const [loading,setLoading]=useState(false);const [searchInput,setSearchInput]=useState(initialSearch);const [search,setSearch]=useState(initialSearch);const [isComposing,setIsComposing]=useState(false);const [cursors,setCursors]=useState<Array<string|null>>([null]);const [pageIndex,setPageIndex]=useState(0);const [nextCursor,setNextCursor]=useState<string|null>(null);const [revision,setRevision]=useState(0);const [showCreate,setShowCreate]=useState(false);const [roleTarget,setRoleTarget]=useState<AdminUser|null>(null);const [temporaryCredential,setTemporaryCredential]=useState<{username:string;password:string;expiresAt:string}|null>(null);const searchRef=useRef<HTMLInputElement>(null);const isAdmin=user.capabilities.manageAccounts;const cursor=cursors[pageIndex]??null;
+  const commitSearch=useCallback((value:string,historyMode:"push"|"replace"="push",updateInput=true)=>{const normalized=value.trim();const params=new URLSearchParams(window.location.search);if(normalized)params.set("accountSearch",normalized);else params.delete("accountSearch");window.history[historyMode==="push"?"pushState":"replaceState"]({},"",`${window.location.pathname}${params.size?`?${params.toString()}`:""}${window.location.hash}`);if(updateInput)setSearchInput(normalized);setSearch(normalized);setCursors([null]);setPageIndex(0);setRevision((current)=>current+1);},[]);
+  useEffect(()=>{const restore=()=>{const restored=new URLSearchParams(window.location.search).get("accountSearch")??"";setSearchInput(restored);setSearch(restored);setCursors([null]);setPageIndex(0);setRevision((current)=>current+1);};window.addEventListener("popstate",restore);return()=>window.removeEventListener("popstate",restore);},[]);
+  useEffect(()=>{if(isComposing||searchInput.trim()===search)return;const timer=window.setTimeout(()=>commitSearch(searchInput,"push",false),300);return()=>window.clearTimeout(timer);},[searchInput,isComposing,search,commitSearch]);
+  const requestKey=JSON.stringify([search,cursor]);const lastRequestKey=useRef<string|null>(null);
+  useEffect(()=>{if(!isAdmin)return;if(lastRequestKey.current!==requestKey){lastRequestKey.current=requestKey;setUsers([]);}const controller=new AbortController();const params=new URLSearchParams();if(search)params.set("search",search);if(cursor)params.set("cursor",cursor);setLoading(true);setLoadError("");setNextCursor(null);fetch(`/api/admin/users${params.size?`?${params.toString()}`:""}`,{signal:controller.signal}).then(async(response)=>{const data=await readResponseJson<{users?:AdminUser[];roles?:Array<{code:string;name:string}>;permissionMatrix?:RolePermission[];nextCursor?:string|null;message?:string}>(response,"账号响应无效，请重试。");if(!response.ok)throw new Error(data.message??"账号加载失败");setUsers(data.users??[]);setRoles(data.roles??[]);setPermissionMatrix(data.permissionMatrix??[]);setNextCursor(data.nextCursor??null);}).catch((error)=>{if(error instanceof DOMException&&error.name==="AbortError")return;setLoadError(error instanceof Error?error.message:"账号加载失败");}).finally(()=>{if(!controller.signal.aborted)setLoading(false);});return()=>controller.abort();},[isAdmin,search,cursor,requestKey,revision]);
+  function refresh(){setRevision((value)=>value+1);}
+  function restart(){setCursors([null]);setPageIndex(0);setRevision((value)=>value+1);}
+  function submitSearch(event:FormEvent){event.preventDefault();if(isComposing)return;setMessage("");commitSearch(searchInput);}
+  function clearSearch(){setSearchInput("");commitSearch("","push",false);window.requestAnimationFrame(()=>searchRef.current?.focus());}
+  function nextPage(){if(!nextCursor)return;setCursors((current)=>[...current.slice(0,pageIndex+1),nextCursor]);setPageIndex((value)=>value+1);}
+  async function toggle(item:AdminUser){const response=await apiFetch(`/api/admin/users/${item.id}/status`,{method:"PATCH",headers:{"content-type":"application/json"},body:JSON.stringify({isActive:!item.isActive})});const data=await readResponseJson<{message?:string}>(response,"状态修改响应无效，请重试。");if(!response.ok){setMessage(data.message??"状态修改失败");return;}refresh();}
+  async function resetPassword(item:AdminUser){const response=await apiFetch(`/api/admin/users/${item.id}/reset-password`,{method:"POST",headers:{"content-type":"application/json"},body:"{}"});const data=await readResponseJson<{message?:string;temporaryPassword?:string;temporaryPasswordExpiresAt?:string}>(response,"密码重置响应无效，请重试。");if(!response.ok||!data.temporaryPassword||!data.temporaryPasswordExpiresAt){setMessage(data.message??"密码重置失败");return;}setTemporaryCredential({username:item.username,password:data.temporaryPassword,expiresAt:data.temporaryPasswordExpiresAt});refresh();}
+  return <main className="dashboard"><header><div><h1>账号管理</h1><p>系统管理权限与业务权限分离；业务角色必须显式分配</p></div>{isAdmin?<button className="primary-action" onClick={()=>setShowCreate(true)}><Plus size={16}/>创建账号</button>:null}</header>{!isAdmin?<div className="permission-note"><ShieldCheck size={18}/>仅独立系统管理员可以维护账号和角色。</div>:null}{message?<p className="page-message" role="status">{message}</p>:null}{isAdmin?<><section className="orders-card" aria-labelledby="accounts-table-title"><div className="orders-toolbar"><div><h2 id="accounts-table-title">系统账号</h2><span role="status">{loading?"正在查询…":`第 ${pageIndex+1} 页 · 本页 ${users.length} 个账号`}</span></div><form noValidate className="account-search" role="search" onSubmit={submitSearch}><div className="account-search-field"><Search size={16}/><input ref={searchRef} aria-label="搜索账号" type="search" value={searchInput} onCompositionStart={()=>setIsComposing(true)} onCompositionEnd={()=>setIsComposing(false)} onChange={(event)=>setSearchInput(event.target.value)} placeholder="账号或姓名"/>{searchInput?<button type="button" className="account-search-clear" aria-label="清除账号搜索" onClick={clearSearch}><X size={15}/></button>:null}</div><button type="submit">搜索账号</button></form></div>{loadError?<div className="query-feedback"><p className="page-message" role="alert">{loadError}</p><button type="button" onClick={refresh}>重试查询</button></div>:null}<div className="orders-table-wrap"><table><thead><tr><th>账号</th><th>姓名</th><th>角色</th><th>状态</th><th>操作</th></tr></thead><tbody>{!loading&&!loadError&&users.length===0?<tr><td colSpan={5} className="empty-cell">{search?"没有符合条件的账号。":"暂无账号数据。"}</td></tr>:users.map((item)=><tr key={item.id}><td>{item.username}</td><td>{item.displayName}</td><td>{item.roles.map((role)=>roleNames[role]??role).join("、")}</td><td><span className={`status ${item.isActive?"status-active":"status-paused"}`}>{item.isActive?"启用":"停用"}</span></td><td><div className="table-actions"><button className="table-action" onClick={()=>setRoleTarget(item)}>修改角色</button><button className="table-action" onClick={()=>resetPassword(item)}>重置密码</button><button className="table-action" onClick={()=>toggle(item)}>{item.isActive?"停用":"启用"}</button></div></td></tr>)}</tbody></table></div><nav className="table-pagination" aria-label="账号分页"><span>固定按账号创建顺序分页</span><button type="button" disabled={loading||pageIndex===0} onClick={()=>setPageIndex((value)=>value-1)}>上一页</button><button type="button" disabled={loading||!nextCursor} onClick={nextPage}>下一页</button></nav></section><section className="orders-card permission-matrix-card" aria-labelledby="permission-matrix-title"><div className="orders-toolbar"><div><h2 id="permission-matrix-title">角色权限说明</h2><span>多角色账号取各角色权限并集；系统管理员角色本身不增加业务权限</span></div></div><div className="orders-table-wrap" tabIndex={0} role="region" aria-label="角色权限矩阵，可横向滚动"><table className="permission-matrix-table"><thead><tr><th>角色</th><th>数据范围</th><th>业务操作</th><th>目标职责</th><th>导出权限</th><th>明确禁止</th></tr></thead><tbody>{permissionMatrix.length===0?<tr><td colSpan={6} className="empty-cell">暂无角色权限定义</td></tr>:permissionMatrix.map((item)=><tr key={item.code}><td><strong>{item.name}</strong></td><td><span className={`scope-badge scope-${item.businessScope}`}>{scopeName(item.businessScope)}</span></td><td>{item.businessOperations.join("；")}</td><td>{item.targetResponsibilities}</td><td>{item.exportPermission}</td><td>{item.forbidden.join("；")}</td></tr>)}</tbody></table></div></section></>:null}{showCreate?<CreateAccount roles={roles} onClose={()=>setShowCreate(false)} onUncertain={(username)=>{setShowCreate(false);setMessage("创建结果不确定，已重新查询该账号；如账号已存在，请重置密码生成新的临时密码。");commitSearch(username);}} onSaved={async()=>{setShowCreate(false);restart();}}/>:null}{roleTarget?<EditAccountRoles item={roleTarget} roles={roles} onClose={()=>setRoleTarget(null)} onSaved={async()=>{const changedSelf=roleTarget.id===user.id;setRoleTarget(null);setMessage(`${roleTarget.username} 的角色已更新。`);if(changedSelf)window.location.reload();else refresh();}}/>:null}{temporaryCredential?<Modal title="临时密码已生成" note="该密码只显示一次，24 小时后失效" onClose={()=>setTemporaryCredential(null)}><div className="temporary-password-result"><p>账号：{temporaryCredential.username}</p><strong>{temporaryCredential.password}</strong><span>失效时间：{new Date(temporaryCredential.expiresAt).toLocaleString("zh-CN")}</span><div className="modal-actions"><button type="button" onClick={()=>setTemporaryCredential(null)}>我已安全保存</button></div></div></Modal>:null}</main>;
+}
+
+function scopeName(scope:RolePermission["businessScope"]):string{return{none:"无业务范围",self:"仅本人",group:"本人及所负责小组",department:"本人及所负责部门",all:"全公司 / 销售组织"}[scope];}
+
+function EditAccountRoles({item,roles,onClose,onSaved}:{item:AdminUser;roles:Array<{code:string;name:string}>;onClose:()=>void;onSaved:()=>Promise<void>}){const[selected,setSelected]=useState<string[]>(item.roles);const[error,setError]=useState("");const[saving,setSaving]=useState(false);function toggle(role:string){setSelected((current)=>current.includes(role)?current.filter((item)=>item!==role):[...current,role]);}async function submit(event:FormEvent){event.preventDefault();if(saving)return;if(selected.length===0){setError("至少保留一个固定角色。");return;}setSaving(true);setError("");try{const response=await apiFetch(`/api/admin/users/${item.id}/roles`,{method:"PATCH",headers:{"content-type":"application/json"},body:JSON.stringify({roles:selected})});const data=await readResponseJson<{message?:string}>(response,"角色修改响应无效，请重试。");if(!response.ok){setError(data.message??"角色修改失败");return;}await onSaved();}catch(failure){setError(failure instanceof Error?failure.message:"网络异常，角色修改失败，请重试。");}finally{setSaving(false);}}return <Modal title={`修改 ${item.username} 的角色`} note="只能选择固定角色；系统管理员角色本身不增加业务查看或导出权限" onClose={onClose}><form noValidate className="business-form" onSubmit={submit}><fieldset className="role-options"><legend>固定角色</legend>{roles.map((role)=><label key={role.code}><input type="checkbox" checked={selected.includes(role.code)} onChange={()=>toggle(role.code)}/><span>{role.name}</span></label>)}</fieldset>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving||selected.length===0} aria-busy={saving}>{saving?"正在保存…":"保存角色"}</button></div></form></Modal>}
+
+function CreateAccount({roles,onClose,onUncertain,onSaved}:{roles:Array<{code:string;name:string}>;onClose:()=>void;onUncertain:(username:string)=>void;onSaved:()=>Promise<void>}){
+  const[username,setUsername]=useState("");const[displayName,setDisplayName]=useState("");const[role,setRole]=useState(roles[0]?.code??"salesperson");const[personId,setPersonId]=useState("");const[people,setPeople]=useState<PersonOption[]>([]);const[error,setError]=useState("");const[created,setCreated]=useState<{temporaryPassword:string;temporaryPasswordExpiresAt:string}|null>(null);const[saving,setSaving]=useState(false);
+  useEffect(()=>{fetch("/api/admin/people").then(async(response)=>{const data=await response.json() as {people?:PersonOption[];message?:string};if(!response.ok)throw new Error(data.message??"人员列表加载失败");setPeople((data.people??[]).filter((person)=>!person.username));}).catch((reason)=>setError(reason instanceof Error?reason.message:"人员列表加载失败"));},[]);
+  async function submit(event:FormEvent){event.preventDefault();if(saving)return;setSaving(true);setError("");try{const response=await apiFetch("/api/admin/users",{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({username,displayName,roles:[role],personId:personId?Number(personId):null})});if(!response.ok){const data=await response.json().catch(()=>({})) as {message?:string};setError(data.message??"创建账号失败，请重试。");return;}const data=await readResponseJson<{temporaryPassword?:string;temporaryPasswordExpiresAt?:string}>(response,"创建账号响应无效，请重试。");if(!data.temporaryPassword||!data.temporaryPasswordExpiresAt){onUncertain(username);return;}setCreated({temporaryPassword:data.temporaryPassword,temporaryPasswordExpiresAt:data.temporaryPasswordExpiresAt});}catch{onUncertain(username);}finally{setSaving(false);}}
+  return <Modal title="创建系统账号" note="可绑定已有人员身份；临时密码只显示一次，24 小时内必须完成首次改密" onClose={onClose} preventClose={saving}>{created?<div className="temporary-password-result"><p>请立即安全保存临时密码，关闭后无法再次查看。</p><strong>{created.temporaryPassword}</strong><span>失效时间：{new Date(created.temporaryPasswordExpiresAt).toLocaleString("zh-CN")}</span><div className="modal-actions"><button type="button" onClick={async()=>{await onSaved();}}>我已安全保存</button></div></div>:<form noValidate className="business-form" onSubmit={submit}><div className="form-grid"><Field label="登录账号" value={username} onChange={setUsername}/><Field label="账号显示姓名" value={displayName} onChange={setDisplayName}/><label className="field"><span>绑定已有人员（可选）</span><select value={personId} onChange={(event)=>setPersonId(event.target.value)}><option value="">新建人员身份</option>{people.map((person)=><option key={person.id} value={person.id}>{person.displayName}</option>)}</select></label><label className="field"><span>角色</span><select value={role} onChange={(event)=>setRole(event.target.value)}>{roles.map((item)=><option key={item.code} value={item.code}>{item.name}</option>)}</select></label></div>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose} disabled={saving}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在创建…":"创建账号"}</button></div></form>}</Modal>;
+}
