+++ apps/web/src/pages/goal-workspace.tsx
@@ -1,0 +1,55 @@
+import { type FormEvent, type ReactNode, useCallback, useEffect, useRef, useState } from "react";
+import { Plus } from "lucide-react";
+import { GOAL_CONFIRMATION_STATEMENT, apiFetch, businessDateToday, formatMoney, formatOperationTime, goalAuditName, goalLevelName, goalStatusName, readResponseJson, workflowStatusName } from "../app-api";
+import type { FormalReport, Goal, GoalChangeRequest, GoalHistory, GoalLevel, GoalLinkageDecision, GoalOption, GoalWorkflows, ParentGoalOption, User } from "../app-types";
+import { Field, Modal } from "../shared-ui";
+
+type GoalDialog={kind:"sign"|"approve"|"reject"|"request";goal:Goal};
+type WorkflowDialog={kind:"accept"|"reject";request:GoalChangeRequest}|{kind:"linkage";linkage:GoalLinkageDecision};
+
+function allocationText(goal:Goal){if(goal.allocationType==="balanced")return "已完全分配";const label=goal.allocationType==="overallocated"?"超分配":"未分配";return `${label} ${formatMoney(Math.abs(Number(goal.allocationDifference)))}${goal.allocationRatio?`（${goal.allocationRatio}%）`:""}`;}
+function differenceText(value:string|null){if(value===null)return "—";const number=Number(value);return `${number>0?"+":""}${formatMoney(number)}`;}
+function auditDataText(value:unknown){return value===null||value===undefined?"—":JSON.stringify(value);}
+async function fetchGoalHistory(goalId:string,signal?:AbortSignal){const response=await fetch(`/api/goals/${goalId}/history`,signal?{signal}:{});const result=await readResponseJson<GoalHistory&{message?:string}>(response,"目标历史响应无效，请重试。");if(!response.ok)throw new Error(result.message??"目标历史加载失败");return result;}
+
+export function GoalWorkspace({ user, pendingOnly }: { user: User; pendingOnly: boolean }) {
+  const [goals,setGoals]=useState<Goal[]>([]);const[workflows,setWorkflows]=useState<GoalWorkflows>({changeRequests:[],linkageDecisions:[]});
+  const[showCreate,setShowCreate]=useState(false);const[message,setMessage]=useState("");const[loading,setLoading]=useState(true);
+  const[formalGoal,setFormalGoal]=useState<Goal|null>(null);const[historyGoal,setHistoryGoal]=useState<Goal|null>(null);
+  const[goalDialog,setGoalDialog]=useState<GoalDialog|null>(null);const[workflowDialog,setWorkflowDialog]=useState<WorkflowDialog|null>(null);const[busyWorkflowId,setBusyWorkflowId]=useState<string|null>(null);
+  const loadRevision=useRef(0);
+  const canCreate=user.roles.some((role)=>["sales_manager","sales_supervisor","sales_leader"].includes(role));
+  const load=useCallback(async()=>{const revision=++loadRevision.current;setLoading(true);setGoals([]);setWorkflows({changeRequests:[],linkageDecisions:[]});try{const [goalResponse,workflowResponse]=await Promise.all([fetch(pendingOnly?"/api/goals?pendingOnly=true":"/api/goals"),pendingOnly?fetch("/api/goal-workflows"):Promise.resolve(null)]);const goalData=await readResponseJson<{goals?:Goal[];message?:string}>(goalResponse,"目标响应无效，请重试。");if(!goalResponse.ok)throw new Error(goalData.message??"目标加载失败");let workflowData:GoalWorkflows={changeRequests:[],linkageDecisions:[]};if(workflowResponse){const parsed=await readResponseJson<GoalWorkflows&{message?:string}>(workflowResponse,"目标待办响应无效，请重试。");if(!workflowResponse.ok)throw new Error(parsed.message??"目标待办加载失败");workflowData={changeRequests:parsed.changeRequests??[],linkageDecisions:parsed.linkageDecisions??[]};}if(revision!==loadRevision.current)return;setGoals(goalData.goals??[]);setWorkflows(workflowData);}catch(error){if(revision!==loadRevision.current)return;setGoals([]);setWorkflows({changeRequests:[],linkageDecisions:[]});throw error;}finally{if(revision===loadRevision.current)setLoading(false);}},[pendingOnly]);
+  useEffect(()=>{load().catch((error)=>setMessage(error instanceof Error?error.message:"目标加载失败"));},[load]);
+  const refreshed=async(note:string)=>{setGoalDialog(null);setWorkflowDialog(null);try{await load();setMessage(note);}catch{setMessage(`${note} 列表刷新失败，请手动刷新页面。`);}};
+  async function withdraw(item:GoalChangeRequest){if(busyWorkflowId)return;setBusyWorkflowId(item.id);setMessage("");try{const response=await apiFetch(`/api/goal-change-requests/${item.id}/withdraw`,{method:"POST",headers:{"content-type":"application/json"},body:"{}"});const data=await readResponseJson<{message?:string}>(response,"撤回响应无效，请重试。");if(!response.ok){if(response.status===403||response.status===409){await refreshed("状态已变化，已重新读取权威状态。");return;}setMessage(data.message??"撤回失败");return;}await refreshed("修改申请已撤回，可以基于当前生效目标重新申请。");}catch{await refreshed("操作结果不确定，已重新读取权威状态。");}finally{setBusyWorkflowId(null);}}
+  return <main className="dashboard goals-page"><header><div><h1>{pendingOnly?"审批中心":"目标管理"}</h1><p>{pendingOnly?"处理目标审批、修改申请和逐级联动选择":"按组织责任逐级下达，实名确认和审批均绑定具体版本"}</p></div><div className="header-actions">{user.capabilities.exportGoals?<button className="secondary-action" onClick={()=>{window.location.href="/api/exports/goals.csv"}}>导出目标</button>:null}{canCreate&&!pendingOnly?<button className="primary-action" onClick={()=>setShowCreate(true)}><Plus size={16}/>下达目标</button>:null}</div></header>{message?<p className="page-message" role="status">{message}</p>:null}<section className="orders-card" aria-labelledby="goal-table-title"><div className="orders-toolbar"><div><h2 id="goal-table-title">{pendingOnly?"待确认与待审批目标":"目标责任台账"}</h2><span role="status">{loading?"正在读取…":`${goals.length} 条记录`}</span></div></div><div className="orders-table-wrap"><table><thead><tr><th>月份</th><th>层级 / 范围</th><th>责任人</th><th>目标</th><th>直接下级</th><th>分配差额</th><th>状态</th><th>操作</th></tr></thead><tbody>{!loading&&goals.length===0?<tr><td colSpan={8} className="empty-cell">{pendingOnly?"当前没有待确认或待审批目标。":"暂无目标；有下达权限的负责人可创建第一条目标。"}</td></tr>:goals.map((goal)=><tr key={goal.id}><td>{goal.periodMonth}</td><td><strong>{goalLevelName(goal.level)}</strong>{goal.orgUnitName?<small className="table-secondary">{goal.orgUnitName}</small>:null}</td><td>{goal.ownerName}</td><td>{formatMoney(goal.amount)}{goal.effectiveAmount!==null&&goal.effectiveAmount!==goal.amount?<small className="table-secondary">当前生效 {formatMoney(goal.effectiveAmount)}</small>:null}</td><td>{formatMoney(goal.allocatedAmount)}</td><td className={goal.allocationType==="overallocated"?"warning":""}>{allocationText(goal)}</td><td><span className={`status goal-${goal.status}`}>{goalStatusName(goal.status)}</span></td><td><div className="row-actions">{goal.status==="active"?<button onClick={()=>setFormalGoal(goal)}>查看正式报表</button>:null}<button onClick={()=>setHistoryGoal(goal)}>版本与记录</button>{goal.ownerPersonId===user.personId&&goal.status==="pending_signature"?<button onClick={()=>setGoalDialog({kind:"sign",goal})}>确认目标</button>:null}{goal.ownerPersonId===user.personId&&goal.level!=="sales_manager"&&goal.status==="active"?<button onClick={()=>setGoalDialog({kind:"request",goal})}>申请修改</button>:null}{goal.status==="pending_gm"&&user.roles.includes("general_manager")?<><button onClick={()=>setGoalDialog({kind:"approve",goal})}>批准</button><button className="danger-action" onClick={()=>setGoalDialog({kind:"reject",goal})}>拒绝</button></>:null}{goal.status==="pending_hr"&&user.roles.includes("hr")?<><button onClick={()=>setGoalDialog({kind:"approve",goal})}>批准</button><button className="danger-action" onClick={()=>setGoalDialog({kind:"reject",goal})}>拒绝</button></>:null}</div></td></tr>)}</tbody></table></div></section>{pendingOnly?<><WorkflowTable title="目标修改申请" empty="当前没有相关修改申请。" rows={workflows.changeRequests.map((item)=><tr key={item.id}><td>{goalLevelName(item.level)} · {item.ownerName}</td><td>{item.requestedByName}</td><td>{item.currentAmount?formatMoney(item.currentAmount):"—"}</td><td>{item.requestedAmount?formatMoney(item.requestedAmount):"由直属上级填写"}</td><td>{item.newAmount?formatMoney(item.newAmount):"尚未生成"}</td><td>{differenceText(item.amountDifference)}</td><td className="wrap-cell">{item.reason}{item.outcomeComment?<small className="table-secondary">处理意见：{item.outcomeComment}</small>:null}</td><td><span className={`status workflow-${item.status}`}>{workflowStatusName(item.status)}</span></td><td><div className="row-actions">{item.canHandle?<><button onClick={()=>setWorkflowDialog({kind:"accept",request:item})}>接受并填金额</button><button className="danger-action" onClick={()=>setWorkflowDialog({kind:"reject",request:item})}>拒绝</button></>:null}{item.canWithdraw?<button disabled={busyWorkflowId!==null} aria-busy={busyWorkflowId===item.id} onClick={()=>withdraw(item)}>{busyWorkflowId===item.id?"正在撤回…":"撤回"}</button>:null}</div></td></tr>)} headers={["目标","申请人","当前生效","建议金额","候选版本","版本差异","原因 / 意见","状态","操作"]}/><WorkflowTable title="目标联动选择" empty="当前没有待处理的目标联动。" rows={workflows.linkageDecisions.map((item)=><tr key={item.id}><td>{goalLevelName(item.parentLevel)}</td><td>{item.childOwnerName}</td><td>{formatMoney(item.childAmount)}</td><td>{formatMoney(item.parentAmount)}</td><td><span className={`status workflow-${item.status}`}>{workflowStatusName(item.status)}</span></td><td><div className="row-actions">{item.canDecide?<button onClick={()=>setWorkflowDialog({kind:"linkage",linkage:item})}>选择是否调整</button>:null}</div></td></tr>)} headers={["本级目标","变化的下级责任人","下级新目标","本级生效目标","状态","操作"]}/></>:null}{showCreate?<CreateGoal user={user} onClose={()=>setShowCreate(false)} onSaved={async()=>{setShowCreate(false);try{await load();setMessage("目标已下达，等待责任人确认。");}catch{setMessage("目标已下达，但列表刷新失败，请手动刷新页面。");}}}/>:null}{goalDialog?<GoalActionDialog dialog={goalDialog} onClose={()=>setGoalDialog(null)} onSaved={()=>refreshed(goalDialog.kind==="request"?"修改申请已提交。":"目标状态已更新。")} onStale={()=>refreshed("状态已变化，已重新读取权威状态。")} onUncertain={()=>refreshed("操作结果不确定，已重新读取权威状态。")}/>:null}{workflowDialog?<WorkflowActionDialog dialog={workflowDialog} onClose={()=>setWorkflowDialog(null)} onSaved={()=>refreshed("目标待办已处理。")} onStale={()=>refreshed("状态已变化，已重新读取权威状态。")} onUncertain={()=>refreshed("操作结果不确定，已重新读取权威状态。")}/>:null}{historyGoal?<GoalHistoryDialog goal={historyGoal} onClose={()=>setHistoryGoal(null)}/>:null}{formalGoal?<FormalReportDialog goal={formalGoal} onClose={()=>setFormalGoal(null)}/>:null}</main>;
+}
+
+function WorkflowTable({title,empty,headers,rows}:{title:string;empty:string;headers:string[];rows:ReactNode[]}){return <section className="orders-card workflow-card"><div className="orders-toolbar"><div><h2>{title}</h2><span>{rows.length} 条记录</span></div></div><div className="orders-table-wrap"><table><thead><tr>{headers.map((header)=><th key={header}>{header}</th>)}</tr></thead><tbody>{rows.length?rows:<tr><td colSpan={headers.length} className="empty-cell">{empty}</td></tr>}</tbody></table></div></section>}
+
+function GoalActionDialog({dialog,onClose,onSaved,onStale,onUncertain}:{dialog:GoalDialog;onClose:()=>void;onSaved:()=>Promise<void>;onStale:()=>Promise<void>;onUncertain:()=>Promise<void>}){
+  const[reason,setReason]=useState("");const[requestedAmount,setRequestedAmount]=useState("");const[comment,setComment]=useState("");const[error,setError]=useState("");const[saving,setSaving]=useState(false);
+  const[history,setHistory]=useState<GoalHistory|null>(null);const[historyError,setHistoryError]=useState("");const[historyRevision,setHistoryRevision]=useState(0);
+  const needsHistory=dialog.kind==="approve"||dialog.kind==="reject";
+  const title=dialog.kind==="sign"?"确认目标版本":dialog.kind==="request"?"申请修改目标":dialog.kind==="approve"?"批准目标":"拒绝目标";
+  useEffect(()=>{if(!needsHistory)return;const controller=new AbortController();setHistory(null);setHistoryError("");fetchGoalHistory(dialog.goal.id,controller.signal).then(setHistory).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setHistoryError(failure instanceof Error?failure.message:"目标历史加载失败");});return()=>controller.abort();},[dialog.goal.id,historyRevision,needsHistory]);
+  const currentVersion=history?.versions.find((version)=>version.id===dialog.goal.versionId)??null;
+  const previousVersion=currentVersion?history?.versions.find((version)=>Number(version.versionNo)===Number(currentVersion.versionNo)-1)??null:null;
+  const currentApprovals=currentVersion?history?.approvals.filter((approval)=>approval.goalVersionId===currentVersion.id)??[]:[];
+  const latestRequest=history?.changeRequests[0]??null;
+  async function submit(event:FormEvent){event.preventDefault();if(saving||needsHistory&&!currentVersion)return;setSaving(true);setError("");let url=`/api/goals/${dialog.goal.id}/decision`;let body:Record<string,unknown>={expectedVersionId:dialog.goal.versionId,decision:dialog.kind==="reject"?"rejected":"approved",comment};if(dialog.kind==="sign"){url=`/api/goal-versions/${dialog.goal.versionId}/confirm`;body={};}else if(dialog.kind==="request"){url=`/api/goals/${dialog.goal.id}/change-requests`;body={reason,...(requestedAmount?{requestedAmount:Number(requestedAmount)}:{})};}try{const response=await apiFetch(url,{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify(body)});const data=await readResponseJson<{message?:string}>(response,"服务响应无效，请重试。");if(!response.ok){if(response.status===403||response.status===409){await onStale();return;}setError(data.message??"操作失败");return;}await onSaved();}catch{await onUncertain();}finally{setSaving(false);}}
+  return <Modal title={title} note={`${dialog.goal.periodMonth} · ${goalLevelName(dialog.goal.level)} · 版本 ${dialog.goal.versionNo}`} onClose={onClose}><form noValidate className="business-form" onSubmit={submit}>{needsHistory?historyError?<section className="approval-evidence"><p className="form-error" role="alert">{historyError}</p><button type="button" onClick={()=>setHistoryRevision((value)=>value+1)}>重新读取审批依据</button></section>:!history?<p role="status">正在读取审批依据…</p>:!currentVersion?<section className="approval-evidence"><p className="form-error" role="alert">当前待办版本已变化，请返回审批中心重新处理。</p><button type="button" onClick={onStale}>重新读取审批中心</button></section>:<><section className="approval-evidence" aria-labelledby="approval-evidence-title"><h3 id="approval-evidence-title">审批依据</h3><div className="approval-facts"><p>原目标 <strong>{currentVersion.previousAmount===null?"—":formatMoney(currentVersion.previousAmount)}</strong></p><p>当前目标 <strong>{formatMoney(currentVersion.amount)}</strong></p><p>差异 <strong>{differenceText(currentVersion.amountDifference)}</strong></p><p>状态 <strong>{previousVersion?goalStatusName(previousVersion.status):"初始版本"} → {goalStatusName(currentVersion.status)}</strong></p></div><div className="approval-confirmation"><strong>目标确认</strong><p>确认人 {currentVersion.signedByName??"尚未确认"}{currentVersion.signedAt?` · ${formatOperationTime(currentVersion.signedAt)}`:""}</p><p>{currentVersion.signatureText??"未留下确认声明"}</p></div><div className="approval-opinions"><h4>原因与既有意见</h4><p>版本原因：{currentVersion.changeReason}</p>{latestRequest?<p>修改申请：{latestRequest.reason}{latestRequest.outcomeComment?` · 处理意见 ${latestRequest.outcomeComment}`:""}</p>:null}{currentApprovals.map((approval)=><p key={approval.id}>{approval.stage==="general_manager"?"总经理":"人事部"} {approval.decidedByName}：{approval.comment||"未填写意见"}</p>)}</div></section><label className="field"><span>审批意见</span><textarea className="resize-none" rows={4} value={comment} onChange={(event)=>setComment(event.target.value)} placeholder={dialog.kind==="reject"?"拒绝时必须说明原因":"填写本次审批意见"}/></label></>:dialog.kind==="sign"?<section className="form-note" aria-labelledby="goal-confirmation-statement"><strong id="goal-confirmation-statement">固定确认声明</strong><p>{GOAL_CONFIRMATION_STATEMENT}</p><p>责任人：{dialog.goal.ownerName} · 目标金额：{formatMoney(dialog.goal.amount)}</p></section>:<div className="form-grid"><Field label="修改原因" value={reason} onChange={setReason}/><label className="field"><span>建议金额（可选）</span><input value={requestedAmount} type="number" min="0" step="0.01" onChange={(event)=>setRequestedAmount(event.target.value)}/></label></div>}{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving||needsHistory&&!currentVersion} aria-busy={saving}>{saving?"正在提交…":dialog.kind==="sign"?"确认目标":dialog.kind==="request"?"提交修改申请":dialog.kind==="approve"?"确认批准":"确认拒绝"}</button></div></form></Modal>;
+}
+
+function WorkflowActionDialog({dialog,onClose,onSaved,onStale,onUncertain}:{dialog:WorkflowDialog;onClose:()=>void;onSaved:()=>Promise<void>;onStale:()=>Promise<void>;onUncertain:()=>Promise<void>}){const isLinkage=dialog.kind==="linkage";const[decision,setDecision]=useState<"keep_parent"|"adjust_parent">("keep_parent");const[newAmount,setNewAmount]=useState(!isLinkage&&dialog.kind==="accept"?(dialog.request.requestedAmount??""):"");const[reason,setReason]=useState("");const[error,setError]=useState("");const[saving,setSaving]=useState(false);async function submit(event:FormEvent){event.preventDefault();if(saving)return;setSaving(true);setError("");let url:string;let body:Record<string,unknown>;if(isLinkage){url=`/api/goal-linkage-decisions/${dialog.linkage.id}/decide`;body={decision,reason,...(decision==="adjust_parent"&&dialog.linkage.parentLevel==="sales_manager"?{newAmount:Number(newAmount)}:{})};}else{url=`/api/goal-change-requests/${dialog.request.id}/${dialog.kind}`;body=dialog.kind==="accept"?{newAmount:Number(newAmount),comment:reason}:{comment:reason};}try{const response=await apiFetch(url,{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify(body)});const data=await readResponseJson<{message?:string}>(response,"服务响应无效，请重试。");if(!response.ok){if(response.status===403||response.status===409){await onStale();return;}setError(data.message??"待办处理失败");return;}await onSaved();}catch{await onUncertain();}finally{setSaving(false);}}const title=isLinkage?"处理目标联动":dialog.kind==="accept"?"接受修改申请":"拒绝修改申请";return <Modal title={title} note={isLinkage?"不调整会保留本级确认；调整会逐级发起申请，绝不自动改金额":"直属目标下达人填写新金额，责任人需要重新确认"} onClose={onClose}><form noValidate className="business-form" onSubmit={submit}>{isLinkage?<><label className="field"><span>处理方式</span><select value={decision} onChange={(event)=>setDecision(event.target.value as "keep_parent"|"adjust_parent")}><option value="keep_parent">不调整本级目标，保留原确认</option><option value="adjust_parent">调整本级目标，发起上一级流程</option></select></label>{decision==="adjust_parent"&&dialog.linkage.parentLevel==="sales_manager"?<Field label="销售经理总目标新金额" value={newAmount} type="number" onChange={setNewAmount}/>:null}</>:dialog.kind==="accept"?<Field label="新目标金额" value={newAmount} type="number" onChange={setNewAmount}/>:null}<label className="field"><span>{isLinkage?"联动原因":"处理意见"}</span><textarea className="resize-none" rows={4} value={reason} onChange={(event)=>setReason(event.target.value)}/></label>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在提交…":isLinkage?"确认联动选择":dialog.kind==="accept"?"接受并创建新版本":"确认拒绝"}</button></div></form></Modal>}
+
+function GoalHistoryDialog({goal,onClose}:{goal:Goal;onClose:()=>void}){const[data,setData]=useState<GoalHistory|null>(null);const[error,setError]=useState("");useEffect(()=>{const controller=new AbortController();fetchGoalHistory(goal.id,controller.signal).then(setData).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setError(failure instanceof Error?failure.message:"目标历史加载失败");});return()=>controller.abort();},[goal.id]);return <Modal title="目标版本与操作记录" note={`${goal.periodMonth} · ${goal.ownerName} · 每次确认和审批均绑定具体版本`} onClose={onClose}>{error?<p className="form-error history-content" role="alert">{error}</p>:!data?<p className="history-content">正在读取版本与审计记录…</p>:<div className="history-content"><section><h3>版本与差异</h3><ol className="history-list">{data.versions.map((version)=><li key={version.id}><div><strong>版本 {version.versionNo} · {formatMoney(version.amount)}</strong><span className={`status goal-${version.status}`}>{goalStatusName(version.status)}</span></div><p>{version.previousAmount===null?"初始版本":`基准 ${formatMoney(version.previousAmount)} · 差异 ${differenceText(version.amountDifference)}`} · {version.changeReason} · 录入 {version.createdByName} · 确认者 {version.signedByName??"尚未确认"}</p></li>)}</ol></section><section><h3>修改申请</h3>{data.changeRequests.length?<ol className="history-list">{data.changeRequests.map((request)=><li key={request.id}><div><strong>{request.requestedByName} · {workflowStatusName(request.status)}</strong><time>{formatOperationTime(request.createdAt)}</time></div><p>{request.reason} · 建议金额 {request.requestedAmount?formatMoney(request.requestedAmount):"未填写"} · 处理人 {request.handledByName??"尚未处理"}{request.outcomeComment?` · 意见 ${request.outcomeComment}`:""}</p></li>)}</ol>:<p>暂无修改申请。</p>}</section><section><h3>审批</h3>{data.approvals.length?<ol className="history-list">{data.approvals.map((approval)=><li key={approval.id}><div><strong>{approval.stage==="general_manager"?"总经理":"人事部"} · {approval.decision==="approved"?"批准":"拒绝"}</strong><time>{formatOperationTime(approval.decidedAt)}</time></div><p>{approval.decidedByName} · {approval.comment}</p></li>)}</ol>:<p>暂无审批记录。</p>}</section><section><h3>审计轨迹</h3><ol className="history-list">{data.audit.map((item)=><li key={item.id}><div><strong>{goalAuditName(item.action)}</strong><time>{formatOperationTime(item.createdAt)}</time></div><p>{item.actorName??"系统"} · 变更前 {auditDataText(item.beforeData)} · 变更后 {auditDataText(item.afterData)}</p></li>)}</ol></section></div>}</Modal>}
+
+function FormalReportDialog({goal,onClose}:{goal:Goal;onClose:()=>void}){
+  const[report,setReport]=useState<FormalReport|null>(null);const[error,setError]=useState("");
+  useEffect(()=>{fetch(`/api/performance/formal-reports/${goal.id}`).then(async(response)=>{const data=await response.json() as FormalReport&{message?:string};if(!response.ok)throw new Error(data.message??"正式报表加载失败");setReport(data);}).catch((reason)=>setError(reason instanceof Error?reason.message:"正式报表加载失败"));},[goal.id]);
+  return <Modal title="正式业绩报表" note="目标已生效，达成率按该层级目标独立计算" onClose={onClose}>{error?<p className="form-error formal-report-content">{error}</p>:report?<div className="formal-report-content"><div><span>目标月份</span><strong>{report.periodMonth}</strong></div><div><span>责任层级</span><strong>{goalLevelName(report.level)}</strong></div><div><span>生效目标</span><strong>{formatMoney(report.targetAmount)}</strong></div><div><span>实际业绩</span><strong>{formatMoney(report.actualAmount)}</strong></div><div><span>目标差距</span><strong>{formatMoney(report.gapAmount)}</strong></div><div><span>达成率</span><strong>{report.achievementRate===null?"不计算":`${report.achievementRate}%`}</strong></div><div className="modal-actions"><button type="button" onClick={()=>{window.location.href=`/api/exports/formal-reports/${goal.id}.csv`;}}>导出正式报表</button></div></div>:<p className="formal-report-content">正在计算正式报表…</p>}</Modal>;
+}
+
+function CreateGoal({user,onClose,onSaved}:{user:User;onClose:()=>void;onSaved:()=>Promise<void>}){const available:GoalLevel[]=user.roles.includes("sales_manager")?["sales_manager","department"]:user.roles.includes("sales_supervisor")?["group"]:["personal"];const[periodMonth,setPeriodMonth]=useState(businessDateToday().slice(0,7));const[level,setLevel]=useState<GoalLevel>(available[0]!);const[parentGoalId,setParentGoalId]=useState("");const[parentGoals,setParentGoals]=useState<ParentGoalOption[]>([]);const[owners,setOwners]=useState<GoalOption[]>([]);const[ownerKey,setOwnerKey]=useState("");const[amount,setAmount]=useState("");const[changeReason,setChangeReason]=useState("目标下达");const[error,setError]=useState("");const[loading,setLoading]=useState(true);const[saving,setSaving]=useState(false);useEffect(()=>{const controller=new AbortController();setLoading(true);setError("");const params=new URLSearchParams({periodMonth,level});if(parentGoalId)params.set("parentGoalId",parentGoalId);fetch(`/api/goals/options?${params}`,{signal:controller.signal}).then(async(response)=>{const data=await readResponseJson<{parentGoals?:ParentGoalOption[];owners?:GoalOption[];message?:string}>(response,"目标选项响应无效，请重试。");if(!response.ok)throw new Error(data.message??"目标选项加载失败");const nextParents=data.parentGoals??[];const nextOwners=data.owners??[];setParentGoals(nextParents);setOwners(nextOwners);if(level!=="sales_manager"&&!parentGoalId&&nextParents[0]){setParentGoalId(nextParents[0].id);return;}const keys=nextOwners.map((owner)=>`${owner.personId}:${owner.orgUnitId??""}`);setOwnerKey((current)=>keys.includes(current)?current:(keys[0]??""));}).catch((failure)=>{if(failure instanceof DOMException&&failure.name==="AbortError")return;setError(failure instanceof Error?failure.message:"目标选项加载失败");}).finally(()=>{if(!controller.signal.aborted)setLoading(false);});return()=>controller.abort();},[periodMonth,level,parentGoalId]);const selectedOwner=owners.find((owner)=>`${owner.personId}:${owner.orgUnitId??""}`===ownerKey);async function submit(event:FormEvent){event.preventDefault();if(saving)return;if(!selectedOwner||!amount||!changeReason.trim()){setError("请选择合法责任人，并填写目标金额和下达原因。");return;}setSaving(true);setError("");try{const response=await apiFetch("/api/goals",{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({periodMonth,level,ownerPersonId:Number(selectedOwner.personId),orgUnitId:selectedOwner.orgUnitId?Number(selectedOwner.orgUnitId):null,parentGoalId:parentGoalId?Number(parentGoalId):null,amount:Number(amount),changeReason})});const data=await readResponseJson<{message?:string}>(response,"服务响应无效，请重试。");if(!response.ok){setError(data.message??"目标下达失败");return;}await onSaved();}catch(failure){setError(failure instanceof Error?failure.message:"网络异常，目标下达失败，请重试。");}finally{setSaving(false);}}return <Modal title="下达目标" note="责任人从当前月有效组织关系中选择；服务端再次校验直属关系" onClose={onClose}><form noValidate className="business-form" onSubmit={submit}><div className="form-grid"><Field label="目标月份" value={periodMonth} type="month" onChange={(value)=>{setPeriodMonth(value);setParentGoalId("");setOwnerKey("");}}/><label className="field"><span>目标层级</span><select value={level} onChange={(event)=>{setLevel(event.target.value as GoalLevel);setParentGoalId("");setOwnerKey("");}}>{available.map((item)=><option key={item} value={item}>{goalLevelName(item)}</option>)}</select></label>{level!=="sales_manager"?<label className="field"><span>直属上级目标</span><select value={parentGoalId} onChange={(event)=>{setParentGoalId(event.target.value);setOwnerKey("");}}><option value="">请选择已生效的直属上级目标</option>{parentGoals.map((parent)=><option key={parent.id} value={parent.id}>{parent.ownerName} · {parent.orgUnitName??goalLevelName(level)} · {formatMoney(parent.amount)}</option>)}</select></label>:null}<label className="field"><span>目标责任人</span><select value={ownerKey} disabled={loading||owners.length===0} onChange={(event)=>setOwnerKey(event.target.value)}><option value="">{loading?"正在加载合法责任人…":"请选择责任人"}</option>{owners.map((owner)=><option key={`${owner.personId}:${owner.orgUnitId??""}`} value={`${owner.personId}:${owner.orgUnitId??""}`}>{owner.name}{owner.orgUnitName?` · ${owner.orgUnitName}`:""}</option>)}</select></label><Field label="目标金额" value={amount} type="number" onChange={setAmount}/><Field label="下达原因" value={changeReason} onChange={setChangeReason}/></div>{level!=="sales_manager"&&parentGoals.length===0&&!loading?<p className="form-note">当前月份没有可用的已生效直属上级目标，请先完成上一级审批。</p>:null}{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving||loading} aria-busy={saving}>{saving?"正在提交…":"提交待确认"}</button></div></form></Modal>}
