+++ apps/web/src/pages/organization-page.tsx
@@ -1,0 +1,42 @@
+import { type FormEvent, useCallback, useEffect, useRef, useState } from "react";
+import { Plus } from "lucide-react";
+import { apiFetch, businessDateToday, readResponseJson } from "../app-api";
+import type { Assignment, OrgUnit, PersonOption, Responsibility, User } from "../app-types";
+import { Field, Modal } from "../shared-ui";
+
+export function OrganizationPage({user}:{user:User}){
+  const[units,setUnits]=useState<OrgUnit[]>([]);const[assignments,setAssignments]=useState<Assignment[]>([]);const[responsibilities,setResponsibilities]=useState<Responsibility[]>([]);
+  const[message,setMessage]=useState("");const[dialog,setDialog]=useState<"unit"|"assignment"|"transfer"|null>(null);const[closing,setClosing]=useState<Assignment|null>(null);const[replacing,setReplacing]=useState<Responsibility|null>(null);
+  const loadSequence=useRef(0);
+  const isAdmin=user.capabilities.manageOrganization;
+  const load=useCallback(async()=>{const sequence=++loadSequence.current;try{const response=await fetch("/api/organization");const data=await readResponseJson<{units?:OrgUnit[];assignments?:Assignment[];responsibilities?:Responsibility[];message?:string}>(response,"组织架构响应无效，请重试。");if(!response.ok)throw new Error(data.message??"组织架构加载失败");if(sequence!==loadSequence.current)return;setUnits(data.units??[]);setAssignments(data.assignments??[]);setResponsibilities(data.responsibilities??[]);setMessage("");}catch(error){if(sequence===loadSequence.current)setMessage(error instanceof Error?error.message:"组织架构加载失败");throw error;}},[]);
+  useEffect(()=>{load().catch(()=>undefined);},[load]);
+  return <main className="dashboard"><header><div><h1>组织架构</h1><p>组织与人员任职按生效日期管理；业绩入账时固化组织快照</p></div>{isAdmin?<div className="header-actions"><button className="secondary-action" onClick={()=>setDialog("transfer")}>办理组织异动</button><button className="secondary-action" onClick={()=>setDialog("assignment")}>新增任职</button><button className="primary-action" onClick={()=>setDialog("unit")}><Plus size={16}/>新增组织</button></div>:null}</header>{message?<p className="page-message" role="alert">{message}</p>:null}<section className="org-grid"><article className="orders-card"><div className="orders-toolbar"><div><h2>部门与小组</h2><span>{units.length} 个组织单元</span></div></div><div className="compact-list">{units.map((unit)=><div key={unit.id}><span>{unit.unitType==="department"?"部门":"小组"}</span><strong>{unit.name}</strong><small>{unit.parentName??"顶层"}</small></div>)}</div></article><article className="orders-card"><div className="orders-toolbar"><div><h2>人员任职</h2><span>{assignments.length} 条有效期记录</span></div></div><div className="compact-list">{assignments.length?assignments.map((item)=><div key={item.id}><span>{item.effectiveFrom}{item.effectiveTo?` 至 ${item.effectiveTo}`:" 起"}</span><strong>{item.displayName}</strong><div className="compact-list-detail"><small>{[item.departmentName,item.groupName].filter(Boolean).join(" / ")}</small>{isAdmin&&!item.effectiveTo?<button type="button" onClick={()=>setClosing(item)}>关闭任职</button>:null}</div></div>):<p className="empty-cell">暂无人员任职记录</p>}</div></article>{isAdmin?<article className="orders-card responsibility-card"><div className="orders-toolbar"><div><h2>负责人职责</h2><span>{responsibilities.length} 条有效期记录</span></div></div><div className="compact-list responsibility-list">{responsibilities.length?responsibilities.map((item)=><div key={item.id}><span>{item.effectiveFrom}{item.effectiveTo?` 至 ${item.effectiveTo}`:" 起"}</span><strong>{item.displayName}</strong><div className="compact-list-detail"><small>{item.unitName} · {item.responsibilityType==="leader"?"小组负责人":"部门主管"}</small>{!item.effectiveTo?<button type="button" onClick={()=>setReplacing(item)}>更换负责人</button>:null}</div></div>):<p className="empty-cell">暂无负责人职责记录</p>}</div></article>:null}</section>{dialog==="unit"?<CreateOrgUnit units={units} onClose={()=>setDialog(null)} onSaved={async()=>{await load();setDialog(null);}}/>:null}{dialog==="assignment"||dialog==="transfer"?<CreateAssignment units={units} closePrevious={dialog==="transfer"} onClose={()=>setDialog(null)} onSaved={async()=>{await load();setDialog(null);}}/>:null}{closing?<CloseAssignment assignment={closing} onClose={()=>setClosing(null)} onSaved={async(effectiveTo)=>{setAssignments((current)=>current.map((item)=>item.id===closing.id?{...item,effectiveTo}:item));setClosing(null);try{await load();}catch{setMessage("任职已关闭，但组织列表刷新失败，请刷新页面重试。");}}}/>:null}{replacing?<ReplaceResponsibility responsibility={replacing} onClose={()=>setReplacing(null)} onSaved={async(predecessorEffectiveTo)=>{setResponsibilities((current)=>current.map((item)=>item.id===replacing.id?{...item,effectiveTo:predecessorEffectiveTo}:item));setReplacing(null);try{await load();}catch{setMessage("负责人已更换，但组织列表刷新失败，请刷新页面重试。");}}}/>:null}</main>
+}
+
+function CreateOrgUnit({units,onClose,onSaved}:{units:OrgUnit[];onClose:()=>void;onSaved:()=>Promise<void>}){
+  const[name,setName]=useState("");const[unitType,setUnitType]=useState<"department"|"group">("department");const[parentId,setParentId]=useState("");const[error,setError]=useState("");const[saving,setSaving]=useState(false);
+  async function submit(event:FormEvent){event.preventDefault();if(saving)return;setSaving(true);setError("");try{const response=await apiFetch("/api/admin/organization/units",{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({name,unitType,parentId:parentId?Number(parentId):null})});const data=await readResponseJson<{message?:string}>(response,"新增组织响应无效，请重试。");if(!response.ok){setError(data.message??"新增组织失败");return;}await onSaved();}catch(reason){setError(reason instanceof Error?reason.message:"网络异常，新增组织失败，请重试。");}finally{setSaving(false);}}
+  return <Modal title="新增组织单元" note="小组必须归属于一个部门" onClose={onClose} preventClose={saving}><form noValidate className="business-form" onSubmit={submit}><Field label="名称" value={name} onChange={setName}/><label className="field"><span>类型</span><select value={unitType} onChange={(e)=>setUnitType(e.target.value as "department"|"group")}><option value="department">部门</option><option value="group">小组</option></select></label>{unitType==="group"?<label className="field"><span>所属部门</span><select required value={parentId} onChange={(e)=>setParentId(e.target.value)}><option value="">请选择</option>{units.filter((u)=>u.unitType==="department").map((u)=><option key={u.id} value={u.id}>{u.name}</option>)}</select></label>:null}{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose} disabled={saving}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在保存…":"保存组织"}</button></div></form></Modal>;
+}
+
+function CreateAssignment({units,closePrevious=false,onClose,onSaved}:{units:OrgUnit[];closePrevious?:boolean;onClose:()=>void;onSaved:()=>Promise<void>}){
+  const[people,setPeople]=useState<PersonOption[]>([]);const[personId,setPersonId]=useState("");const[departmentId,setDepartmentId]=useState("");const[groupId,setGroupId]=useState("");const[leaderPersonId,setLeaderPersonId]=useState("");const[supervisorPersonId,setSupervisorPersonId]=useState("");const[effectiveFrom,setEffectiveFrom]=useState(businessDateToday);const[error,setError]=useState("");const[saving,setSaving]=useState(false);
+  useEffect(()=>{fetch("/api/admin/people").then(async(response)=>{const data=await response.json() as {people?:PersonOption[];message?:string};if(!response.ok)throw new Error(data.message??"人员列表加载失败");setPeople(data.people??[]);}).catch((reason)=>setError(reason instanceof Error?reason.message:"人员列表加载失败"));},[]);
+  async function submit(event:FormEvent){event.preventDefault();if(saving)return;setSaving(true);setError("");try{const response=await apiFetch("/api/admin/organization/assignments",{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({personId:Number(personId),departmentId:Number(departmentId),groupId:Number(groupId),leaderPersonId:Number(leaderPersonId),supervisorPersonId:Number(supervisorPersonId),effectiveFrom,effectiveTo:null,closePrevious})});const data=await readResponseJson<{message?:string}>(response,"任职保存响应无效，请重试。");if(!response.ok){setError(data.message??"任职保存失败");return;}await onSaved();}catch(reason){setError(reason instanceof Error?reason.message:"网络异常，任职保存失败，请重试。");}finally{setSaving(false);}}
+  const personSelect=(label:string,value:string,onChange:(value:string)=>void)=><label className="field"><span>{label}</span><select required value={value} onChange={(event)=>onChange(event.target.value)}><option value="">请选择</option>{people.map((item)=><option key={item.id} value={item.id}>{item.displayName}{item.username?`（${item.username}）`:"（无登录账号）"}</option>)}</select></label>;
+  return <Modal title={closePrevious?"办理组织异动":"新增人员任职"} note={closePrevious?"系统将于新任职生效日前一天结束原任职；历史业绩归属不会重写":"成员归属与负责人职责均按生效日期保存；新任职不改写历史事件"} onClose={onClose} preventClose={saving}><form noValidate className="business-form" onSubmit={submit}><div className="form-grid">{personSelect("任职人员",personId,setPersonId)}<Field label="生效日期" value={effectiveFrom} type="date" onChange={setEffectiveFrom}/><label className="field"><span>部门</span><select required value={departmentId} onChange={(event)=>{setDepartmentId(event.target.value);setGroupId("");}}><option value="">请选择</option>{units.filter((unit)=>unit.unitType==="department").map((unit)=><option key={unit.id} value={unit.id}>{unit.name}</option>)}</select></label><label className="field"><span>小组</span><select required value={groupId} onChange={(event)=>setGroupId(event.target.value)}><option value="">请选择</option>{units.filter((unit)=>unit.unitType==="group"&&unit.parentId===departmentId).map((unit)=><option key={unit.id} value={unit.id}>{unit.name}</option>)}</select></label>{personSelect("小组负责人",leaderPersonId,setLeaderPersonId)}{personSelect("部门主管",supervisorPersonId,setSupervisorPersonId)}</div>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose} disabled={saving}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在保存…":closePrevious?"确认异动":"保存任职"}</button></div></form></Modal>;
+}
+
+function CloseAssignment({assignment,onClose,onSaved}:{assignment:Assignment;onClose:()=>void;onSaved:(effectiveTo:string)=>Promise<void>}){
+  const[effectiveOn,setEffectiveOn]=useState(businessDateToday);const[error,setError]=useState("");const[saving,setSaving]=useState(false);
+  async function submit(event:FormEvent){event.preventDefault();setSaving(true);setError("");try{const response=await apiFetch(`/api/admin/organization/memberships/${assignment.id}/close`,{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({effectiveOn})});const data=await readResponseJson<{effectiveTo?:string;message?:string}>(response,"任职关闭响应无效，请重试。");if(!response.ok){setError(data.message??"任职关闭失败");return;}if(!data.effectiveTo)throw new Error("任职关闭响应无效，请重试。");await onSaved(data.effectiveTo);}catch(reason){setError(reason instanceof Error?reason.message:"网络异常，任职关闭失败");}finally{setSaving(false);}}
+  return <Modal title="关闭人员任职" note={`${assignment.displayName} · ${[assignment.departmentName,assignment.groupName].filter(Boolean).join(" / ")}；生效日起不再具有该成员归属，历史事件不会改写`} onClose={onClose}><form noValidate className="business-form" onSubmit={submit}><Field label="离任生效日期" value={effectiveOn} type="date" onChange={setEffectiveOn}/>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在关闭…":"确认关闭"}</button></div></form></Modal>;
+}
+
+function ReplaceResponsibility({responsibility,onClose,onSaved}:{responsibility:Responsibility;onClose:()=>void;onSaved:(predecessorEffectiveTo:string)=>Promise<void>}){
+  const[people,setPeople]=useState<PersonOption[]>([]);const[successorPersonId,setSuccessorPersonId]=useState("");const[effectiveOn,setEffectiveOn]=useState(businessDateToday);const[error,setError]=useState("");const[saving,setSaving]=useState(false);
+  useEffect(()=>{fetch("/api/admin/people").then(async(response)=>{const data=await readResponseJson<{people?:PersonOption[];message?:string}>(response,"人员列表响应无效，请重试。");if(!response.ok)throw new Error(data.message??"人员列表加载失败");setPeople(data.people??[]);}).catch((reason)=>setError(reason instanceof Error?reason.message:"人员列表加载失败"));},[]);
+  async function submit(event:FormEvent){event.preventDefault();setSaving(true);setError("");try{const response=await apiFetch(`/api/admin/organization/responsibilities/${responsibility.id}/replace`,{method:"POST",headers:{"content-type":"application/json"},body:JSON.stringify({successorPersonId,effectiveOn})});const data=await readResponseJson<{predecessorEffectiveTo?:string;message?:string}>(response,"负责人继任响应无效，请重试。");if(!response.ok){setError(data.message??"负责人继任失败");return;}if(!data.predecessorEffectiveTo)throw new Error("负责人继任响应无效，请重试。");await onSaved(data.predecessorEffectiveTo);}catch(reason){setError(reason instanceof Error?reason.message:"网络异常，负责人继任失败");}finally{setSaving(false);}}
+  return <Modal title="更换负责人" note={`${responsibility.unitName} · 现任 ${responsibility.displayName}；旧职责于继任生效日前一天结束，不允许负责人空档`} onClose={onClose}><form noValidate className="business-form" onSubmit={submit}><label className="field"><span>继任负责人</span><select required value={successorPersonId} onChange={(event)=>setSuccessorPersonId(event.target.value)}><option value="">请选择</option>{people.filter((person)=>person.id!==responsibility.personId).map((person)=><option key={person.id} value={person.id}>{person.displayName}{person.username?`（${person.username}）`:"（无登录账号）"}</option>)}</select></label><Field label="继任生效日期" value={effectiveOn} type="date" onChange={setEffectiveOn}/>{error?<p className="form-error" role="alert">{error}</p>:null}<div className="modal-actions"><button type="button" onClick={onClose}>取消</button><button type="submit" disabled={saving} aria-busy={saving}>{saving?"正在更换…":"确认继任"}</button></div></form></Modal>;
+}
