+++ apps/web/src/pages/overview-page.tsx
@@ -1,0 +1,101 @@
+import { useEffect, useRef, useState } from "react";
+import { eventTypeName, formatMoney, readResponseJson } from "../app-api";
+import type { DashboardData, DepartmentAchievement, DepartmentAchievementDrilldown, GroupAchievement, GroupAchievementDrilldown, GroupAchievementMember, PersonalAchievement, PersonalAchievementDrilldown, SalesAchievement, SalesAchievementDrilldown } from "../app-types";
+import { Metric, Modal, Status } from "../shared-ui";
+
+export function Overview({ canEdit, canExport, onEnterOrders }: { canEdit: boolean; canExport: boolean; onEnterOrders: () => void }) {
+  const [data, setData] = useState<DashboardData | null>(null);
+  const [showPersonalDetails,setShowPersonalDetails]=useState(false);
+  const [personalDetailKind,setPersonalDetailKind]=useState<"actual"|"gap">("actual");
+  const [personalDetails,setPersonalDetails]=useState<PersonalAchievementDrilldown|null>(null);
+  const [personalDetailsError,setPersonalDetailsError]=useState("");
+  const [selectedGroup,setSelectedGroup]=useState<GroupAchievement|null>(null);
+  const [groupDetailKind,setGroupDetailKind]=useState<"actual"|"gap">("actual");
+  const [groupDetails,setGroupDetails]=useState<GroupAchievementDrilldown|null>(null);
+  const [groupDetailsError,setGroupDetailsError]=useState("");
+  const groupDetailsRequest=useRef(0);
+  const [organizationDialog,setOrganizationDialog]=useState<{kind:"sales"|"department";metric:"actual"|"gap";achievement:SalesAchievement|DepartmentAchievement}|null>(null);
+  const [organizationDetails,setOrganizationDetails]=useState<SalesAchievementDrilldown|DepartmentAchievementDrilldown|null>(null);
+  const [organizationDetailsError,setOrganizationDetailsError]=useState("");
+  const organizationDetailsRequest=useRef(0);
+  useEffect(() => { fetch("/api/performance/dashboard").then(async (response) => { if (!response.ok) throw new Error("总览加载失败"); setData(await response.json() as DashboardData); }).catch(() => setData(null)); }, []);
+  const maxGroup = Math.max(1, ...(data?.groups.map((group) => Number(group.total)) ?? [1]));
+  const personal=data?.personalAchievement;
+  const reason=personal?achievementReasonText(personal.calculationReason,"个人"):null;
+  async function openPersonalDetails(kind:"actual"|"gap"){if(!personal)return;setPersonalDetailKind(kind);setShowPersonalDetails(true);setPersonalDetails(null);setPersonalDetailsError("");try{const response=await fetch(`/api/performance/personal-achievement/events?month=${personal.periodMonth}`);const result=await readResponseJson<PersonalAchievementDrilldown&{message?:string}>(response,"个人业绩构成响应无效");if(!response.ok)throw new Error(result.message??"个人业绩构成加载失败");const{events,...achievement}=result;if(kind==="gap"&&result.gapAmount===null)setPersonalDetailKind("actual");setPersonalDetails(result);setData((current)=>current?{...current,personalAchievement:achievement}:current);}catch(error){setPersonalDetailsError(error instanceof Error?error.message:"个人业绩构成加载失败");}}
+  async function openGroupDetails(group:GroupAchievement,kind:"actual"|"gap"){const requestId=++groupDetailsRequest.current;setSelectedGroup(group);setGroupDetailKind(kind);setGroupDetails(null);setGroupDetailsError("");try{const response=await fetch(`/api/performance/group-achievement/events?month=${group.periodMonth}&groupId=${group.groupId}`);const result=await readResponseJson<GroupAchievementDrilldown&{message?:string}>(response,"小组业绩构成响应无效");if(requestId!==groupDetailsRequest.current)return;if(!response.ok)throw new Error(result.message??"小组业绩构成加载失败");const{members,...achievement}=result;if(kind==="gap"&&result.gapAmount===null)setGroupDetailKind("actual");setGroupDetails(result);setData((current)=>current?{...current,groupAchievements:current.groupAchievements.map((item)=>item.groupId===result.groupId?achievement:item)}:current);}catch(error){if(requestId===groupDetailsRequest.current)setGroupDetailsError(error instanceof Error?error.message:"小组业绩构成加载失败");}}
+  async function openOrganizationDetails(kind:"sales"|"department",achievement:SalesAchievement|DepartmentAchievement,metric:"actual"|"gap"){
+    const requestId=++organizationDetailsRequest.current;
+    setOrganizationDialog({kind,achievement,metric});setOrganizationDetails(null);setOrganizationDetailsError("");
+    try{
+      const url=kind==="sales"
+        ? `/api/performance/sales-achievement/events?month=${achievement.periodMonth}`
+        : (achievement as DepartmentAchievement).departmentId===null
+          ? `/api/performance/sales-achievement/events?month=${achievement.periodMonth}`
+          : `/api/performance/department-achievement/events?month=${achievement.periodMonth}&departmentId=${(achievement as DepartmentAchievement).departmentId}`;
+      const response=await fetch(url);
+      const result=await readResponseJson<(SalesAchievementDrilldown|DepartmentAchievementDrilldown)&{message?:string}>(response,"组织业绩构成响应无效");
+      if(requestId!==organizationDetailsRequest.current)return;
+      if(!response.ok)throw new Error(result.message??"组织业绩构成加载失败");
+      const details=kind==="department"&&"departments" in result
+        ? result.departments.find((item)=>item.departmentKey===(achievement as DepartmentAchievement).departmentKey)??null
+        : result;
+      if(!details)throw new Error("部门业绩构成不存在");
+      setOrganizationDetails(details);
+      if(metric==="gap"&&details.gapAmount===null)setOrganizationDialog({kind,achievement,metric:"actual"});
+    }catch(error){if(requestId===organizationDetailsRequest.current)setOrganizationDetailsError(error instanceof Error?error.message:"组织业绩构成加载失败");}
+  }
+  const sales=data?.salesAchievement;
+  const departments=data?.departmentAchievements??[];
+  const organizationDepartments=organizationDetails?("departments" in organizationDetails?organizationDetails.departments:[organizationDetails]):[];
+  const dashboardDescription=sales?"销售组织目标与不可变业绩事件":departments.length?"个人、部门目标与不可变业绩事件":personal?"个人目标与不可变业绩事件":"原始账本，不代表正式绩效结果";
+  return <main className="dashboard"><header><div><h1>业绩账本总览</h1><p>{data ? `${data.month.replace("-", " 年 ")} 月 · ${dashboardDescription}` : "正在加载真实业绩账本…"}</p></div><div className="header-actions">{canExport?<button className="secondary-action" onClick={()=>{window.location.href="/api/exports/performance.csv"}}>导出全部订单</button>:null}{canEdit?<button className="primary-action" onClick={onEnterOrders}>录入订单业绩</button>:null}</div></header>
+      {personal?<><section className="metric-band personal-metrics" aria-label="个人目标达成"><Metric label="生效目标" value={personal.targetAmount===null?"—":formatMoney(personal.targetAmount)} note={reason??"当前月个人目标"}/><Metric label="个人业绩" value={formatMoney(personal.actualAmount)} note={`${personal.eventCount} 条本人事件`} onClick={()=>openPersonalDetails("actual")} actionLabel="查看个人业绩构成"/><Metric label="差距" value={personal.gapAmount===null?"不计算":formatMoney(personal.gapAmount)} note={reason??"生效目标减个人业绩"} onClick={personal.gapAmount===null?undefined:()=>openPersonalDetails("gap")} actionLabel={personal.gapAmount===null?undefined:"查看个人差距构成"}/><Metric label="达成率" value={personal.achievementRate===null?"不计算":`${personal.achievementRate}%`} note={reason??"个人业绩 ÷ 生效目标"}/><Metric label="自然日进度" value={personal.timeProgressRate===null?"不计算":`${personal.timeProgressRate}%`} note={reason??"按上海自然日计算"}/><Metric label="进度偏差" value={personal.progressVariance===null?"不计算":`${personal.progressVariance}%`} note={reason??"达成率减自然日进度"} warning={personal.progressVariance!==null&&Number(personal.progressVariance)<0}/></section>{showPersonalDetails?<Modal title={personalDetailKind==="gap"?"个人差距构成":"个人业绩构成"} note={personalDetails?(personalDetailKind==="gap"?`目标 ${formatMoney(personalDetails.targetAmount!)} − 事件净额 ${formatMoney(personalDetails.actualAmount)} = 差距 ${formatMoney(personalDetails.gapAmount!)}`:`${personalDetails.events.length} 条事件 · 净额 ${formatMoney(personalDetails.actualAmount)}`):`正在读取 ${personal.periodMonth} 个人业绩事件`} onClose={()=>setShowPersonalDetails(false)}>{personalDetailsError?<p className="form-error personal-drilldown" role="alert">{personalDetailsError}</p>:personalDetails?<div className="personal-drilldown"><div className="orders-table-wrap"><table><thead><tr><th>订单</th><th>客户</th><th>不可变事件</th><th>业务日 / 记账月</th><th>金额方向</th><th>状态 / 计入</th><th>组织归属</th><th>订单当前分析资料</th><th>原因</th></tr></thead><tbody>{personalDetails.events.map((event)=><tr key={event.id}><td>{event.orderNo}</td><td>{event.customerName}</td><td>{eventTypeName(event.eventType)} · 第 {event.sequence} 条</td><td>{event.occurredOn} / {event.accountingMonth}</td><td className={event.deltaAmount.startsWith("-")?"negative":""}>{event.deltaAmount.startsWith("-")?"负向":"正向"} {formatMoney(event.deltaAmount.startsWith("-")?event.deltaAmount.slice(1):event.deltaAmount)}</td><td>{event.resultingLifecycleState?<><Status state={event.resultingLifecycleState}/><small className="table-secondary">计入至 {formatMoney(event.resultingCountedAmount)}</small></>:"原始状态未推断"}</td><td>{[event.departmentName,event.groupName].filter(Boolean).join(" / ")||"—"}</td><td>分析维度待补齐</td><td>{event.reason??"—"}</td></tr>)}</tbody></table></div>{personalDetails.events.length===0?<p className="empty-cell">当前月份没有个人业绩事件。</p>:null}</div>:<p className="personal-drilldown">正在读取个人业绩构成…</p>}</Modal>:null}</>:null}
+      {sales?<OrganizationAchievementCard scope="销售组织" title="销售组织目标达成" countNote={`${sales.departmentCount} 个部门 · 当前仅一个销售组织`} achievement={sales} onOpen={(metric)=>openOrganizationDetails("sales",sales,metric)}/>:null}
+      {departments.map((department)=><OrganizationAchievementCard key={department.departmentKey} scope="部门" title={`部门目标达成 · ${department.departmentName}`} countNote={`${department.groupCount} 个小组 · 按事件发生时组织归属`} achievement={department} onOpen={(metric)=>openOrganizationDetails("department",department,metric)}/>)}
+      {organizationDialog?<Modal title={organizationDialog.kind==="sales"?`销售组织 · ${organizationDialog.metric==="gap"?"差距构成":"业绩构成"}`:`${(organizationDialog.achievement as DepartmentAchievement).departmentName} · ${organizationDialog.metric==="gap"?"部门差距构成":"部门业绩构成"}`} note={organizationDetails?(organizationDialog.metric==="gap"?`目标 ${formatMoney(organizationDetails.targetAmount!)} − 事件净额 ${formatMoney(organizationDetails.actualAmount)} = 差距 ${formatMoney(organizationDetails.gapAmount!)}`:`${organizationDetails.eventCount} 条事件 · 净额 ${formatMoney(organizationDetails.actualAmount)}`):"正在读取部门、小组、人员、订单和事件"} onClose={()=>{organizationDetailsRequest.current+=1;setOrganizationDialog(null);}}>{organizationDetailsError?<p className="form-error group-drilldown" role="alert">{organizationDetailsError}</p>:organizationDetails?<OrganizationHierarchy departments={organizationDepartments}/>:<p className="group-drilldown">正在读取组织业绩构成…</p>}</Modal>:null}
+      {data?.groupAchievements.map((group)=>{const groupReason=achievementReasonText(group.calculationReason,"小组");return <article className="team-achievement orders-card" aria-label={`${group.groupName}目标达成`} key={group.groupId}><PanelTitle title={`小组目标达成 · ${group.groupName}`} note={`${group.memberCount} 位事件责任人 · 按事件发生时组织归属`}/><section className="metric-band personal-metrics"><Metric label="小组生效目标" value={group.targetAmount===null?"—":formatMoney(group.targetAmount)} note={groupReason??"当前月小组目标"}/><Metric label="小组业绩" value={formatMoney(group.actualAmount)} note={`${group.eventCount} 条小组事件`} onClick={()=>openGroupDetails(group,"actual")} actionLabel={`查看${group.groupName}业绩构成`}/><Metric label="小组差距" value={group.gapAmount===null?"不计算":formatMoney(group.gapAmount)} note={groupReason??"生效目标减小组业绩"} onClick={group.gapAmount===null?undefined:()=>openGroupDetails(group,"gap")} actionLabel={group.gapAmount===null?undefined:`查看${group.groupName}差距构成`}/><Metric label="小组达成率" value={group.achievementRate===null?"不计算":`${group.achievementRate}%`} note={groupReason??"小组业绩 ÷ 生效目标"}/><Metric label="自然日进度" value={group.timeProgressRate===null?"不计算":`${group.timeProgressRate}%`} note={groupReason??"按上海自然日计算"}/><Metric label="进度偏差" value={group.progressVariance===null?"不计算":`${group.progressVariance}%`} note={groupReason??"达成率减自然日进度"} warning={group.progressVariance!==null&&Number(group.progressVariance)<0}/></section></article>})}
+      {selectedGroup?<Modal
+        title={`${selectedGroup.groupName} · ${groupDetailKind==="gap"?"小组差距构成":"小组业绩构成"}`}
+        note={groupDetails?(groupDetailKind==="gap"?`目标 ${formatMoney(groupDetails.targetAmount!)} − 事件净额 ${formatMoney(groupDetails.actualAmount)} = 差距 ${formatMoney(groupDetails.gapAmount!)}`:`${groupDetails.memberCount} 位事件责任人 · ${groupDetails.eventCount} 条事件 · 净额 ${formatMoney(groupDetails.actualAmount)}`):`正在读取 ${selectedGroup.groupName} 小组业绩事件`}
+        onClose={()=>{groupDetailsRequest.current+=1;setSelectedGroup(null);}}
+      >{groupDetailsError?<p className="form-error group-drilldown" role="alert">{groupDetailsError}</p>:groupDetails?<div className="group-drilldown"><AchievementMembers members={groupDetails.members}/></div>:<p className="group-drilldown">正在读取小组业绩构成…</p>}</Modal>:null}
+      <section className="metric-band"><Metric label="账本净额" value={data ? formatMoney(data.metrics.total) : "—"} note={`${data?.metrics.eventCount ?? 0} 条授权范围事件`}/><Metric label="正式报表" value="从生效目标进入" note="未生效不计算达成率"/><Metric label="待处理审批" value={String(data?.metrics.pendingApprovals ?? 0)} note="目标确认与变更" warning/><Metric label="负向调整" value={data ? formatMoney(data.metrics.negativeTotal) : "—"} note="暂停与金额变更" negative/></section>
+      <section className="dashboard-grid"><article className="trend-panel"><PanelTitle title="月度业绩趋势" note={`${data?.month.slice(0,4)??"当前"} 年授权范围业绩净额`}/><TrendChart year={data?.month.slice(0,4)??""} monthly={data?.monthly??[]}/></article><article className="ranking-panel"><PanelTitle title="小组业绩" note="按事件发生时组织归属"/>{data?.groups.map((group,i) => <div className="rank-row" key={group.id}><span>{i+1}</span><div><strong>{group.name}</strong><span><i style={{width:`${Math.max(0, Number(group.total)) / maxGroup * 100}%`}}/></span></div><b>{(Number(group.total)/10000).toFixed(1)}万</b></div>)}</article></section>
+      <section className="events-panel"><div className="panel-title"><div><h2>最近业绩事件</h2><p>入账后不可覆盖或删除</p></div><button onClick={onEnterOrders}>查看全部</button></div><table><thead><tr><th>订单编号</th><th>业务员</th><th>事件类型</th><th>记账月</th><th>金额</th><th>组织归属</th><th>状态</th></tr></thead><tbody>{data?.recent.map((event) => <tr key={`${event.orderNo}-${event.month}-${event.amount}`}><td>{event.orderNo}</td><td>{event.salespersonName}</td><td>{eventTypeName(event.eventType)}</td><td>{event.month}</td><td className={Number(event.amount)<0?"negative":""}>{formatMoney(event.amount)}</td><td>{event.groupName}</td><td>已入账</td></tr>)}</tbody></table></section>
+    </main>;
+}
+
+
+function achievementReasonText(value:PersonalAchievement["calculationReason"],scope:string){return value===null?null:({PERIOD_IN_FUTURE:"未来月份不计算目标达成",TARGET_NOT_ACTIVE:`${scope}目标尚未生效`,TARGET_AMOUNT_NOT_POSITIVE:"生效目标金额必须大于零",TARGET_SCOPE_AMBIGUOUS:"存在多个生效的销售组织目标，暂不计算"} as const)[value];}
+function OrganizationAchievementCard({scope,title,countNote,achievement,onOpen}:{scope:"部门"|"销售组织";title:string;countNote:string;achievement:PersonalAchievement;onOpen:(metric:"actual"|"gap")=>void}){const reason=achievementReasonText(achievement.calculationReason,scope);return <article className="team-achievement orders-card" aria-label={title}><PanelTitle title={title} note={countNote}/><section className="metric-band personal-metrics"><Metric label={`${scope}生效目标`} value={achievement.targetAmount===null?"—":formatMoney(achievement.targetAmount)} note={reason??`当前月${scope}目标`}/><Metric label={`${scope}业绩`} value={formatMoney(achievement.actualAmount)} note={`${achievement.eventCount} 条事件`} onClick={()=>onOpen("actual")} actionLabel={`查看${scope}业绩构成`}/><Metric label={`${scope}差距`} value={achievement.gapAmount===null?"不计算":formatMoney(achievement.gapAmount)} note={reason??`生效目标减${scope}业绩`} onClick={achievement.gapAmount===null?undefined:()=>onOpen("gap")} actionLabel={achievement.gapAmount===null?undefined:`查看${scope}差距构成`}/><Metric label={`${scope}达成率`} value={achievement.achievementRate===null?"不计算":`${achievement.achievementRate}%`} note={reason??`${scope}业绩 ÷ 生效目标`}/><Metric label="自然日进度" value={achievement.timeProgressRate===null?"不计算":`${achievement.timeProgressRate}%`} note={reason??"按上海自然日计算"}/><Metric label="进度偏差" value={achievement.progressVariance===null?"不计算":`${achievement.progressVariance}%`} note={reason??"达成率减自然日进度"} warning={achievement.progressVariance!==null&&Number(achievement.progressVariance)<0}/></section></article>}
+function OrganizationHierarchy({departments}:{departments:DepartmentAchievementDrilldown[]}){
+  return <div className="group-drilldown">{departments.map((department)=>{
+    const departmentReason=achievementReasonText(department.calculationReason,"部门");
+    return <section className="orders-card group-member" aria-label={`${department.departmentName}部门构成`} key={department.departmentKey}>
+      <PanelTitle title={department.departmentName} note={`${department.groupCount} 个小组 · ${department.eventCount} 条事件 · 实际 ${formatMoney(department.actualAmount)} · ${departmentReason??`达成率 ${department.achievementRate}%`}`}/>
+      {department.groups.map((group)=>{
+        const groupReason=achievementReasonText(group.calculationReason,"小组");
+        return <article className="orders-card group-member" aria-label={`${group.groupName}小组构成`} key={group.groupKey}>
+          <PanelTitle title={group.groupName} note={`${group.memberCount} 位事件责任人 · ${group.eventCount} 条事件 · 实际 ${formatMoney(group.actualAmount)} · ${groupReason??`达成率 ${group.achievementRate}%`}`}/>
+          <AchievementMembers members={group.members}/>
+          {group.members.length===0?<p className="empty-cell">当前月份没有小组业绩事件。</p>:null}
+        </article>;
+      })}
+      {department.groups.length===0?<p className="empty-cell">当前月份没有部门小组。</p>:null}
+    </section>;
+  })}</div>;
+}
+function AchievementMembers({members}:{members:GroupAchievementMember[]}){
+  return <>{members.map((member)=><section className="orders-card group-member" key={member.personKey}>
+    <PanelTitle title={member.name} note={`${member.orders.length} 笔订单 · ${member.eventCount} 条事件 · 净额 ${formatMoney(member.actualAmount)}`}/>
+    {member.orders.map((order)=><div className="group-order" key={order.orderId}>
+      <h4>{order.orderNo} · {order.customerName}<small>{order.eventCount} 条事件 · 净额 {formatMoney(order.actualAmount)}</small></h4>
+      <div className="orders-table-wrap"><table><thead><tr><th>不可变事件</th><th>业务日 / 记账月</th><th>金额方向</th><th>状态 / 计入</th><th>组织归属</th><th>分析维度</th><th>原因</th></tr></thead><tbody>{order.events.map((event)=><tr key={event.id}><td>{eventTypeName(event.eventType)} · 第 {event.sequence} 条</td><td>{event.occurredOn} / {event.accountingMonth}</td><td className={event.deltaAmount.startsWith("-")?"negative":""}>{event.deltaAmount.startsWith("-")?"负向":"正向"} {formatMoney(event.deltaAmount.startsWith("-")?event.deltaAmount.slice(1):event.deltaAmount)}</td><td>{event.resultingLifecycleState?<><Status state={event.resultingLifecycleState}/><small className="table-secondary">计入至 {formatMoney(event.resultingCountedAmount)}</small></>:"原始状态未推断"}</td><td>{[event.departmentName,event.groupName].filter(Boolean).join(" / ")||"—"}</td><td>待补齐</td><td>{event.reason??"—"}</td></tr>)}</tbody></table></div>
+    </div>)}
+  </section>)}</>;
+}
+
+function PanelTitle({ title, note }: { title: string; note: string }) { return <div className="panel-title"><div><h2>{title}</h2><p>{note}</p></div></div>; }
+
+function TrendChart({year,monthly}:{year:string;monthly:Array<{month:string;total:string}>}) { const totals=new Map(monthly.map((item)=>[item.month,item.total]));const raw=Array.from({length:12},(_,index)=>totals.get(`${year}-${String(index+1).padStart(2,"0")}`)??"0");const filled=raw.map(Number);const max=Math.max(1,...filled);const x=(index:number)=>44+index*(526/11);const points=filled.map((value,index)=>`${x(index)},${190-(value/max)*150}`).join(" ");const description=raw.map((value,index)=>`${index+1}月 ${formatMoney(value)}`).join("；");return <div className="chart"><svg viewBox="0 0 590 220" role="img" aria-label={`${year}年1月至12月业绩折线图`}><desc>{description}</desc><g className="grid-lines">{[40,90,140,190].map((y)=><line key={y} x1="42" y1={y} x2="570" y2={y}/>)}</g><polyline points={points} fill="none" stroke="#2f6fed" strokeWidth="4" strokeLinecap="round" strokeLinejoin="round"/>{filled.map((value,index)=><circle key={index} cx={x(index)} cy={190-(value/max)*150} r="5" fill="#fff" stroke="#2f6fed" strokeWidth="3"/>)}</svg><div className="chart-labels">{Array.from({length:12},(_,index)=><span key={index}>{index+1}月</span>)}</div></div>; }
