+++ apps/web/src/shared-ui.tsx
@@ -1,0 +1,8 @@
+import { type ReactNode, useEffect, useId, useRef } from "react";
+import { createPortal } from "react-dom";
+import type { OrderLifecycle } from "./app-types";
+
+export function Field({ label, value, onChange, type="text" }: { label: string; value: string; onChange: (value: string) => void; type?: string }) { return <label className="field"><span>{label}</span><input required value={value} type={type} min={type === "number" ? "0" : undefined} step={type === "number" ? "0.01" : undefined} onChange={(event) => onChange(event.target.value)}/></label>; }
+export function Modal({ title, note, onClose, preventClose=false, children }: { title: string; note: string; onClose: () => void; preventClose?: boolean; children: ReactNode }) { const titleId=useId();const noteId=useId();const dialogRef=useRef<HTMLElement>(null);const closeRef=useRef(onClose);const requestClose=()=>{if(!preventClose)onClose();};closeRef.current=requestClose;useEffect(()=>{const prior=document.activeElement instanceof HTMLElement?document.activeElement:null;const root=document.getElementById("root");const priorOverflow=document.body.style.overflow;root?.setAttribute("inert","");root?.setAttribute("aria-hidden","true");document.body.style.overflow="hidden";const dialog=dialogRef.current;dialog?.focus();function keydown(event:KeyboardEvent){if(event.key==="Escape"){event.preventDefault();closeRef.current();return;}if(event.key!=="Tab"||!dialog)return;const controls=[...dialog.querySelectorAll<HTMLElement>('button:not([disabled]),input:not([disabled]),select:not([disabled]),textarea:not([disabled]),[href],[tabindex]:not([tabindex="-1"])')].filter((item)=>!item.hasAttribute("hidden"));if(!controls.length){event.preventDefault();dialog.focus();return;}const first=controls[0]!;const last=controls.at(-1)!;if(event.shiftKey&&document.activeElement===first){event.preventDefault();last.focus();}else if(!event.shiftKey&&document.activeElement===last){event.preventDefault();first.focus();}}document.addEventListener("keydown",keydown);return()=>{document.removeEventListener("keydown",keydown);root?.removeAttribute("inert");root?.removeAttribute("aria-hidden");document.body.style.overflow=priorOverflow;prior?.focus();};},[]);return createPortal(<div className="modal-backdrop" role="presentation" onMouseDown={(event) => { if (event.target === event.currentTarget) requestClose(); }}><section ref={dialogRef} tabIndex={-1} className="modal" role="dialog" aria-modal="true" aria-labelledby={titleId} aria-describedby={noteId}><header><div><h2 id={titleId}>{title}</h2><p id={noteId}>{note}</p></div><button type="button" onClick={requestClose} disabled={preventClose} aria-label="关闭">×</button></header>{children}</section></div>,document.body); }
+export function Status({ state }: { state: OrderLifecycle }) { const names:Record<OrderLifecycle,string> = { draft:"草稿", active:"正向计入", paused:"已暂停", zero:"零金额",historical_review_required:"待历史核对" }; return <span className={`status status-${state}`}>{names[state]}</span>; }
+export function Metric({label,value,note,warning,negative,onClick,actionLabel}:{label:string;value:string;note:string;warning?:boolean;negative?:boolean;onClick?:(()=>void)|undefined;actionLabel?:string|undefined}) { return <div className="metric"><span>{label}</span>{onClick?<button type="button" className={`metric-value-action${negative?" negative":""}`} aria-label={actionLabel} onClick={onClick}>{value}</button>:<strong className={negative?"negative":""}>{value}</strong>}<small className={warning?"warning":""}>{note}</small></div>; }
